% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgroup.R
\name{AddWGroup}
\alias{AddWGroup}
\title{Add a plotting object to a group}
\usage{
AddWGroup(group.obj, new.obj)
}
\arguments{
\item{group.obj}{WGroup object to be added to}

\item{new.obj}{plotting object to be added}
}
\value{
a WGroup object where new.obj is added.
}
\description{
The object to be added are in the same coordinate system as the group.
}
