% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{as.data.frame,sesameQC-method}
\alias{as.data.frame,sesameQC-method}
\title{Convert sesameQC to data.frame}
\usage{
\S4method{as.data.frame}{sesameQC}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{a sesameQC object}

\item{...}{additional argument to as.data.frame}
}
\value{
a data.frame
}
\description{
Convert sesameQC to data.frame
}
\examples{
as.data.frame(new("sesameQC"))
}
