% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{as.data.frame.sesameQC}
\alias{as.data.frame.sesameQC}
\title{Coerce a sesameQC into a dataframe}
\usage{
\method{as.data.frame}{sesameQC}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{a sesameQC object}

\item{row.names}{see as.data.frame}

\item{optional}{see as.data.frame}

\item{...}{see as.data.frame}
}
\value{
a data.frame
}
\description{
Coerce a sesameQC into a dataframe
}
\examples{
sset <- sesameDataGet('EPIC.1.LNCaP')$sset
qc <- sesameQC(sset)
df <- as.data.frame(qc)
}
