% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strain.R
\name{betaToAF}
\alias{betaToAF}
\title{convert betas to variant allele frequency for mouse array}
\usage{
betaToAF(betas, platform = "MM285")
}
\arguments{
\item{betas}{beta value}
}
\value{
SNP variant allele frequency
}
\description{
see formatVCF if you would like to convert color-channel
probes on the human array.
}
\examples{
sesameDataCache("MM285") # if not done yet
sdf = sesameDataGet('MM285.1.SigDF')
vafs = betaToAF(getBetas(dyeBiasNL(noob(sdf))))
}
