% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{BetaValueToMValue}
\alias{BetaValueToMValue}
\title{Convert beta-value to M-value}
\usage{
BetaValueToMValue(b)
}
\arguments{
\item{b}{vector of beta values}
}
\value{
a vector of M values
}
\description{
Convert beta-value to M-value (aka logit transform)
}

