% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.R
\name{binSignals}
\alias{binSignals}
\title{Bin signals from probe signals}
\usage{
binSignals(probe.signals, bin.coords, probe.coords)
}
\arguments{
\item{probe.signals}{probe signals}

\item{bin.coords}{bin coordinates}

\item{probe.coords}{probe coordinates}
}
\description{
require GenomicRanges
}
