% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bSubMostVariable}
\alias{bSubMostVariable}
\title{Get most variable probes}
\usage{
bSubMostVariable(betas, n = 2000)
}
\arguments{
\item{betas}{beta value matrix (row: cpg; column: sample)}

\item{n}{number of most variable probes}
}
\value{
beta value matrix for the most variable probes
}
\description{
Get most variable probes
}
\examples{
## get most variable autosome probes
betas <- sesameDataGet('HM450.10.TCGA.PAAD.normal')
betas.most.variable <- bSubMostVariable(
    betas[getAutosomeProbes('HM450'),],2000)

## clear cache
rm(list=ls(env=sesameData:::cacheEnv), envir=sesameData:::cacheEnv)
gc()
}
