% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{buildControlMatrix450k}
\alias{buildControlMatrix450k}
\title{Build control summary matrix}
\usage{
buildControlMatrix450k(sset)
}
\arguments{
\item{sset}{an object of class SigSet}
}
\value{
a vector with control summaries
}
\description{
Build control summary matrix
}
\examples{
sset <- makeExampleSeSAMeDataSet()
control.summary <- buildControlMatrix450k(sset)

}
