% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KYCG.R
\name{buildGeneDBs}
\alias{buildGeneDBs}
\title{build gene-probe association database}
\usage{
buildGeneDBs(query = NULL, platform = NULL, max_distance = 10000)
}
\arguments{
\item{query}{the query probe list. If NULL, use all the probes
on the platform}

\item{platform}{HM450, EPIC, MM285, Mammal40, will infer from
query if not given}

\item{max_distance}{probe-gene distance for association}
}
\description{
build gene-probe association database
}
