% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kyCG.R
\name{calcDatabaseSetStatisticsAll}
\alias{calcDatabaseSetStatisticsAll}
\title{calcDatabaseSetStatisticsAll builds dataset for a given betas matrix 
composed of engineered features from the given database sets}
\usage{
calcDatabaseSetStatisticsAll(betas, databaseSets)
}
\arguments{
\item{betas}{matrix of beta values where probes are on the rows and samples
are on the columns}

\item{databaseSets}{List of vectors corresponding to probe locations for
which the features will be extracted}
}
\value{
Vector for a given sample columns are features across different
databaseSets
}
\description{
calcDatabaseSetStatisticsAll builds dataset for a given betas matrix 
composed of engineered features from the given database sets
}
