% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kyCG.R
\name{calcFoldChange}
\alias{calcFoldChange}
\title{calcFoldChange calculates fold change given a 2x2 matrix of counts.}
\usage{
calcFoldChange(mtx)
}
\arguments{
\item{mtx}{2x2 matrix of values corresponding to overlapping counts between
two sets of a categorical variable.}
}
\value{
A numerical value corresponding to the fold change enrichment,
}
\description{
calcFoldChange calculates fold change given a 2x2 matrix of counts.
}
