% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods.R
\name{CalcTextBounding}
\alias{CalcTextBounding}
\title{Calculate Text Bounding}
\usage{
CalcTextBounding(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{extra options}
}
\description{
Calculate bounding box including texts.
}
\details{
W.R.T lower left corner of the view port in the unit of points
}
