% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wheatmap.R
\name{CalcTextBounding.WHeatmap}
\alias{CalcTextBounding.WHeatmap}
\title{Calculate Texting Bounding for WHeatmap}
\usage{
\method{CalcTextBounding}{WHeatmap}(hm, group)
}
\arguments{
\item{hm}{an object of class WHeatmap}

\item{group}{an object of class WGroup}
}
\value{
an object of class WDim in coordinate points
}
\description{
Calculate Texting Bounding for WHeatmap
}
