% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{checkLevels}
\alias{checkLevels}
\title{filter data matrix by factor completeness
only works for discrete factors}
\usage{
checkLevels(betas, fc)
}
\arguments{
\item{betas}{matrix data}

\item{fc}{factors, or characters}
}
\value{
a boolean vector whether there is non-NA value for each tested
group for each probe
}
\description{
filter data matrix by factor completeness
only works for discrete factors
}
\examples{
se0 = sesameDataGet("MM285.10.tissues")[1:1000,]
se_ok = checkLevels(SummarizedExperiment::assay(se0),
    SummarizedExperiment::colData(se0)$tissue)
sum(se_ok) # number of good probes
se1 = se0[se_ok,]

sesameDataClearCache()
}
