% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.R
\name{cn.segmentation}
\alias{cn.segmentation}
\title{perform copy number segmentation}
\usage{
cn.segmentation(sset, ssets.normal = NULL, refversion = "hg19")
}
\arguments{
\item{sset}{\code{SignalSet}}

\item{ssets.normal}{\code{SignalSet} for normalization}

\item{refversion}{hg19 or hg38}
}
\value{
segmentation
}
\description{
perform copy number segmentation
}

