% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colormap.R
\name{ColorMap}
\alias{ColorMap}
\title{Constructor for ColoMap object}
\usage{
ColorMap(continuous = TRUE, colors = NULL, dmin = NULL, dmax = NULL,
  scaler = NULL, mapper = NULL)
}
\arguments{
\item{continuous}{whether colormap is continuous}

\item{colors}{colors for each data point}

\item{dmin}{miminum in continuous color map}

\item{dmax}{maximum in continuous color map}

\item{scaler}{scaler function from data range to 0-1}

\item{mapper}{function that maps data to color}
}
\value{
an object of class ColorMap
}
\description{
Create color maps
}
