% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kyCG.R
\name{compareDatbaseSetOverlap}
\alias{compareDatbaseSetOverlap}
\title{calculates the pariwise overlap between given list
of database sets using a distance metric.}
\usage{
compareDatbaseSetOverlap(databases = NA, metric = "Jaccard", verbose = FALSE)
}
\arguments{
\item{databases}{List of vectors corresponding to the database sets of
interest with associated meta data as an attribute to each element. Optional.
(Default: NA)}

\item{metric}{String representing the similarity metric to use. Optional.
(Default: "Jaccard").}

\item{verbose}{Logical value indicating whether to display intermediate
text output about the type of test. Optional. (Default: FALSE)}
}
\value{
An upper triangular matrix containing a metric (Jaccard) comparing
the pairwise distances between database sets.
}
\description{
calculates the pariwise overlap between given list
of database sets using a distance metric.
}
\examples{

databaseNames = c('KYCG.MM285.seqContextN.20210630')
databases = do.call(c, lapply(databaseNames, sesameDataGet))
compareDatbaseSetOverlap(databases)
sesameDataClearCache()

}
