% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue.R
\name{compareMouseBloodReference}
\alias{compareMouseBloodReference}
\title{Compare beta value against mouse blood reference}
\usage{
compareMouseBloodReference(betas = NULL, color = "blueYellow")
}
\arguments{
\item{betas}{matrix of betas for the target sample}

\item{color}{either blueYellow or fullJet}
}
\value{
grid object that co-plots with a pre-built mouse blood reference
}
\description{
Compare beta value against mouse blood reference
}
\examples{
sesameDataCache("MM285") # if not done yet
b = sesameDataGet("MM285.10.tissue")$betas[,10]
compareMouseBloodReference(b)
}
