% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue.R
\name{compareMouseTissueReference}
\alias{compareMouseTissueReference}
\title{Compare mouse array data with mouse tissue references}
\usage{
compareMouseTissueReference(betas)
}
\arguments{
\item{betas}{matrix of betas for the target sample}
}
\value{
grid object that contrast the target sample with
pre-built mouse tissue reference
}
\description{
Compare mouse array data with mouse tissue references
}
\examples{
b = sesameDataGet("MM285.10.tissue")$betas[,1:2]
compareMouseTissueReference(b)
}
