% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigDFMethods.R
\name{controls}
\alias{controls}
\title{get the controls attributes}
\usage{
controls(sdf, verbose = FALSE)
}
\arguments{
\item{sdf}{a \code{SigDF}}

\item{verbose}{print more messages}
}
\value{
the controls data frame
}
\description{
get the controls attributes
}
\examples{
sesameDataCache() # if not done yet
sdf <- sesameDataGet('EPIC.1.SigDF')
head(controls(sdf))
}
