% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{convertProbeID}
\alias{convertProbeID}
\title{Convert Probe ID}
\usage{
convertProbeID(
  x,
  target_platform,
  source_platform = NULL,
  mapping = NULL,
  target_uniq = TRUE,
  include_new = FALSE,
  include_old = FALSE,
  return_mapping = FALSE
)
}
\arguments{
\item{x}{source probe IDs}

\item{target_platform}{the platform to take the data to}

\item{source_platform}{optional source platform}

\item{mapping}{a liftOver mapping file. Typically this file
contains empirical evidence whether a probe mapping is reliable.
If given, probe ID-based mapping will be skipped. This is to
perform more stringent probe ID mapping.}

\item{target_uniq}{whether the target Probe ID should be kept unique.}

\item{include_new}{if true, include mapping of added probes}

\item{include_old}{if true, include mapping of deleted probes}

\item{return_mapping}{return mapping table, instead of the target IDs.}
}
\value{
mapped probe IDs, or mapping table if return_mapping = T
}
\description{
Convert Probe ID
}
