% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{convertTo}
\alias{convertTo}
\title{Convert human arrays to previous platforms
Missing probes are replaced using NAs.}
\usage{
convertTo(sdf, target_platform = c("HM450", "EPIC"))
}
\arguments{
\item{sdf}{SigDF data frame}

\item{target_platform}{HM450 or EPIC}
}
\value{
a new SigDF for the older platform
}
\description{
Convert human arrays to previous platforms
Missing probes are replaced using NAs.
}
\examples{
sdf <- sesameDataGet("EPIC.5.SigDF.normal")[[1]]
sdf_out <- convertTo(sdf, "HM450")
}
