% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigSetMethods.R
\docType{methods}
\name{ctl}
\alias{ctl}
\alias{ctl,SigSet-method}
\title{ctl getter generic}
\usage{
ctl(x)

\S4method{ctl}{SigSet}(x)
}
\arguments{
\item{x}{object of \code{SigSet}}
}
\value{
The ctl slot of \code{SigSet}
}
\description{
ctl getter generic

Get ctl slot of SigSet class
}
\examples{
sset <- sesameDataGet('HM450.1.TCGA.PAAD')$sset
head(ctl(sset))
}
