% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigSetMethods.R
\docType{methods}
\name{ctl<-}
\alias{ctl<-}
\alias{ctl<-,SigSet-method}
\title{ctl replacement generic}
\usage{
ctl(x) <- value

\S4method{ctl}{SigSet}(x) <- value
}
\arguments{
\item{x}{object of \code{SigSet}}

\item{value}{new value}
}
\value{
a new \code{SigSet}
}
\description{
ctl replacement generic

Replace ctl slot of SigSet class
}
\examples{
sset <- sesameDataGet('HM450.1.TCGA.PAAD')$sset
df <- ctl(sset)
df[1,1] <- 10
ctl(sset) <- df 
}
