% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{darkjet.stops}
\alias{darkjet.stops}
\title{darker jet color stops}
\format{An object of class \code{character} of length 6.}
\usage{
darkjet.stops
}
\description{
darker jet color stops
}
\keyword{datasets}
