% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{dataFrame2sesameQC}
\alias{dataFrame2sesameQC}
\title{Convert data frame to sesameQC object}
\usage{
dataFrame2sesameQC(df)
}
\arguments{
\item{df}{a publicQC data frame}
}
\value{
a list sesameQC objects
}
\description{
The function convert a data frame back to a list of sesameQC objects
}
\examples{
df <- sesameDataGet("MM285.publicQC")
qcs <- dataFrame2sesameQC(df[1:2,])
}
