% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deidentify.R
\name{deidentify}
\alias{deidentify}
\title{De-identify IDATs by removing SNP probes}
\usage{
deidentify(path, out_path = NULL, snps = NULL, mft = NULL)
}
\arguments{
\item{path}{input IDAT file}

\item{out_path}{output IDAT file}

\item{snps}{SNP definition, if not given, default to SNP probes}

\item{mft}{sesame-compatible manifest if non-standard}
}
\description{
De-identify IDATs by removing SNP probes
}
\examples{

temp_out <- tempfile("test")
deidentify(system.file(
    "extdata", "4207113116_A_Grn.idat", package = "sesameData"),
     temp_out)
unlink(temp_out)
}
