% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{detectionIB}
\alias{detectionIB}
\title{Mask detection by intermediate beta values}
\usage{
detectionIB(
  sdf,
  return.pval = FALSE,
  pval.threshold = 0.05,
  capMU = 3000,
  delta.beta = 0.2,
  n.windows = 500
)
}
\arguments{
\item{sdf}{a \code{SigDF}}

\item{return.pval}{whether to return p-values, instead of a SigDF}

\item{pval.threshold}{minimum p-value to mask}

\item{capMU}{the maximum M+U to search for intermediate betas}

\item{delta.beta}{the delta beta from the intermediate beta to cut}

\item{n.windows}{number of windows for smoothing}
}
\value{
a \code{SigDF} with mask added
}
\description{
Mask detection by intermediate beta values
}
\examples{
sdf <- sesameDataGet("EPIC.1.SigDF")
sum(sdf$mask)
sum(detectionIB(sdf)$mask)
}
