% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{detectionMask}
\alias{detectionMask}
\title{Mask Sigset by detection p-value}
\usage{
detectionMask(sset, pval.method = NULL, pval.threshold = 0.05)
}
\arguments{
\item{sset}{a \code{SigSet}}

\item{pval.method}{which method to use in calculating p-values}

\item{pval.threshold}{the p-value threshold}
}
\value{
a filtered \code{SigSet}
}
\description{
Mask Sigset by detection p-value
}
\examples{
sset <- sesameDataGet('EPIC.1.LNCaP')$sset
sset.masked <- detectionMask(sset)
}
