% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{detectionPfixedNorm}
\alias{detectionPfixedNorm}
\title{Detection P-value based on normal fitting with gived parameters}
\usage{
detectionPfixedNorm(sset, muG = 500, sdG = 200, muR = 500, sdR = 200)
}
\arguments{
\item{sset}{a \code{SigSet}}

\item{muG}{mean of background in Grn channel}

\item{sdG}{SD of background in Grn channel}

\item{muR}{mean of background in Red channel}

\item{sdR}{SD of background in Red channel}
}
\value{
detection p-value
}
\description{
The function takes a \code{SigSet} as input, computes detection p-value
using negative control probes parametrized in a normal distribution and
returns a new \code{SigSet} with an updated pval slot.
}
\details{
Background of Grn and Red are estimated separately from a fixed normal 
distribution. p-value is taken from the minimum of the p-value of the 
two alleles (color depends on probe design).
}
\examples{
sset <- makeExampleSeSAMeDataSet()
sset <- detectionPfixedNorm(sset)

}
