% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{detectionPoobEcdf}
\alias{detectionPoobEcdf}
\alias{pOOBAH}
\title{Detection P-value based on ECDF of out-of-band signal}
\usage{
detectionPoobEcdf(sset, force = FALSE)

pOOBAH(sset, force = FALSE)
}
\arguments{
\item{sset}{a \code{SigSet}}

\item{force}{force rerun even if result already exists}
}
\value{
detection p-value
}
\description{
aka pOOBAH (p-vals by Out-Of-Band Array Hybridization)
}
\details{
The function takes a \code{SigSet} as input, computes detection p-value
using out-of-band probes empirical distribution and returns a new
\code{SigSet} with an updated pval slot.
}
\examples{
sset <- makeExampleSeSAMeDataSet()
sset <- detectionPoobEcdf(sset)

sset <- makeExampleSeSAMeDataSet()
sset <- pOOBAH(sset)
}
