% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{detectionPoobEcdf2}
\alias{detectionPoobEcdf2}
\alias{pOOBAH2}
\title{Detection P-value based on ECDF of out-of-band signal}
\usage{
detectionPoobEcdf2(sset, force = FALSE)

pOOBAH2(sset, force = FALSE)
}
\arguments{
\item{sset}{a \code{SigSet}}

\item{force}{force rerun even if result already exists}
}
\value{
detection p-value
}
\description{
aka pOOBAH2 (p-vals by Out-Of-Band Array Hybridization)
}
\details{
The function takes a \code{SigSet} as input, computes detection p-value
using out-of-band probes empirical distribution and returns a new
\code{SigSet} with an updated pval slot.

The difference between this function and the original pOOBAH
is that pOOBAH2 is based on background-subtracted and dyebias
corrected signal and do not distinguish the color channel difference.
}
\examples{
sset <- makeExampleSeSAMeDataSet()
sset <- detectionPoobEcdf(sset)

sset <- makeExampleSeSAMeDataSet()
sset <- pOOBAH2(sset)
}
