% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{diffMeth}
\alias{diffMeth}
\title{test differential methylation on each locus}
\usage{
diffMeth(betas, sample.data, formula, se.lb = 0.1, cf.test = NULL)
}
\arguments{
\item{betas}{beta values}

\item{sample.data}{data frame for sample information, column names
are predictor variables (e.g., sex, age, treatment, tumor/normal etc)
and are referenced in formula. Rows are samples.}

\item{formula}{formula}

\item{se.lb}{lower bound to standard error of slope}

\item{cf.test}{factors to test (default to all factors in formula except
intercept). Use "all" for all factors.}
}
\value{
cf coefficient table for each factor
}
\description{
test differential methylation on each locus
}

