% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellComposition.R
\name{diffRefSet}
\alias{diffRefSet}
\title{restrict refset to differentially methylated probes
use with care, might introduce bias}
\usage{
diffRefSet(g)
}
\arguments{
\item{g}{reference set}
}
\value{
g
}
\description{
restrict refset to differentially methylated probes
use with care, might introduce bias
}
\examples{
g <- diffRefSet(getRefSet(platform='HM450'))
}
