% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{dmContrasts}
\alias{dmContrasts}
\title{List all contrasts of a DMLSummary}
\usage{
dmContrasts(smry)
}
\arguments{
\item{smry}{a DMLSummary object}
}
\value{
a character vector of contrasts
}
\description{
List all contrasts of a DMLSummary
}
\examples{
data <- sesameDataGet('HM450.76.TCGA.matched')
smry <- DML(data$betas[1:1000,], ~type, meta=data$sampleInfo)
dmContrasts(smry)

sesameDataClearCache()
}
