% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{DML}
\alias{DML}
\title{test differential methylation on each locus}
\usage{
DML(betas, sample.data, formula, se.lb = 0.06, balanced = TRUE,
  cf.test = NULL)
}
\arguments{
\item{betas}{beta values}

\item{sample.data}{data frame for sample information, column names
are predictor variables (e.g., sex, age, treatment, tumor/normal etc)
and are referenced in formula. Rows are samples.}

\item{formula}{formula}

\item{se.lb}{lower bound to standard error of slope, lower this to get
more difference of small effect size.}

\item{balanced}{whether design is balanced or not. default to TRUE, when
unbalanced will use Welch's method to estimate standard error. This is slower
balanced=TRUE still give a good result in most cases.}

\item{cf.test}{factors to test (default to all factors in formula except
intercept). Use "all" for all factors.}
}
\value{
cf coefficient table for each factor
}
\description{
test differential methylation on each locus
}

