% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{DMR}
\alias{DMR}
\title{Find Differentially Methylated Region (DMR)}
\usage{
DMR(
  betas,
  cf,
  platform = NULL,
  refversion = NULL,
  dist.cutoff = NULL,
  seg.per.locus = 0.5
)
}
\arguments{
\item{betas}{beta values for distance calculation}

\item{cf}{coefficient table from DML or DMLShrinkage}

\item{platform}{EPIC, HM450, MM285, ...}

\item{refversion}{hg38, hg19, mm10, ...}

\item{dist.cutoff}{distance cutoff (default to use dist.cutoff.quantile)}

\item{seg.per.locus}{number of segments per locus
higher value leads to more segments}
}
\value{
coefficient table with segment ID and segment P-value
}
\description{
This subroutine uses Euclidean distance to group CpGs and
then combine p-values for each segment. The function performs DML test first
if cf is NULL. It groups the probe testing results into differential
methylated regions in a coefficient table with additional columns
designating the segment ID and statistical significance (P-value) testing
the segment.
}
\examples{

sesameDataCache("HM450") # in case not done yet

data <- sesameDataGet('HM450.76.TCGA.matched')
cf_list = summaryExtractCfList(DML(data$betas, ~type, meta=data$sampleInfo))
cf_list = DMR(data$betas, cf_list$typeTumour)
}
