% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_correction.R
\name{dyeBiasCorr}
\alias{dyeBiasCorr}
\title{Correct dye bias}
\usage{
dyeBiasCorr(sset, ref = 5000)
}
\arguments{
\item{sset}{a \code{SignalSet}}

\item{ref}{reference signal level}
}
\value{
a normalized \code{SignalSet}
}
\description{
Correct dye bias
}

