% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_correction.R
\name{dyeBiasCorrTypeINorm}
\alias{dyeBiasCorrTypeINorm}
\title{Correct dye bias using type I}
\usage{
dyeBiasCorrTypeINorm(sset)
}
\arguments{
\item{sset}{\code{SignalSet} to be modified in-place}
}
\value{
modified sset
}
\description{
Correct dye bias using type I G,R distribution
}

