% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EPIC.controls}
\alias{EPIC.controls}
\title{EPIC controls}
\format{An object of class \code{data.frame} with 635 rows and 4 columns.}
\usage{
EPIC.controls
}
\description{
EPIC controls
}
\keyword{datasets}

