% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EPIC.female.clean.chrY.probes}
\alias{EPIC.female.clean.chrY.probes}
\title{EPIC female clean chrY probes}
\format{An object of class \code{character} of length 314.}
\usage{
EPIC.female.clean.chrY.probes
}
\description{
EPIC female clean chrY probes
}
\keyword{datasets}

