% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EPIC.hg19.probe2chr}
\alias{EPIC.hg19.probe2chr}
\title{EPIC probe2chromosome}
\format{An object of class \code{factor} of length 866895.}
\usage{
EPIC.hg19.probe2chr
}
\description{
EPIC probe2chromosome
}
\keyword{datasets}

