% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EPIC.mapped.probes.hg19}
\alias{EPIC.mapped.probes.hg19}
\title{EPIC mapped probes hg19}
\format{An object of class \code{GRanges} of length 840750.}
\usage{
EPIC.mapped.probes.hg19
}
\description{
EPIC mapped probes hg19
}
\keyword{datasets}

