% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EPIC.mapped.probes.hg38}
\alias{EPIC.mapped.probes.hg38}
\title{EPIC mapped probes hg38}
\format{An object of class \code{GRanges} of length 838881.}
\usage{
EPIC.mapped.probes.hg38
}
\description{
EPIC mapped probes hg38
}
\keyword{datasets}

