% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EPIC.ordering}
\alias{EPIC.ordering}
\title{EPIC data ordering}
\format{An object of class \code{data.frame} with 866895 rows and 6 columns.}
\usage{
EPIC.ordering
}
\description{
EPIC data ordering
}
\keyword{datasets}

