% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellComposition.R
\name{estimateLeukocyte}
\alias{estimateLeukocyte}
\title{Estimate leukocyte fraction using a two-component model}
\usage{
estimateLeukocyte(betas.tissue, betas.leuko = NULL, betas.tumor = NULL,
  platform = "HM450")
}
\arguments{
\item{betas.tissue}{tissue beta value matrix (#probes X #samples)}

\item{betas.leuko}{leukocyte beta value matrix,
if missing, use the SeSAMe default by infinium platform}

\item{betas.tumor}{optional, tumor beta value matrix}

\item{platform}{"HM450", "HM27" or "EPIC"}
}
\value{
leukocyte estimate, a numeric vector
}
\description{
Estimate leukocyte fraction using a two-component model
}
\examples{

betas.tissue <- readRDS(system.file(
    'extdata','HM450.betas.TCGA-2L-AAQA-01A-21D-A38H-05.rds',
    package='sesameData'))
estimateLeukocyte(betas.tissue)

}
