% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ethnicity.model}
\alias{ethnicity.model}
\title{model for ethnicity inference}
\format{An object of class \code{randomForest} of length 18.}
\usage{
ethnicity.model
}
\description{
model for ethnicity inference
}
\keyword{datasets}

