% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ethnicity.rs.probes}
\alias{ethnicity.rs.probes}
\title{rs probes for ethnicity inference}
\format{An object of class \code{character} of length 59.}
\usage{
ethnicity.rs.probes
}
\description{
rs probes for ethnicity inference
}
\keyword{datasets}

