% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigSetMethods.R
\docType{methods}
\name{extra}
\alias{extra}
\alias{extra,SigSet-method}
\title{extra getter generic}
\usage{
extra(x)

\S4method{extra}{SigSet}(x)
}
\arguments{
\item{x}{object of \code{SigSet}}
}
\value{
The extra slot of \code{SigSet}
}
\description{
extra getter generic

Get extra slot of SigSet class
}
\examples{
sset <- sesameDataGet('HM450.1.TCGA.PAAD')$sset
head(extra(sset))
}
