% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigSetMethods.R
\docType{methods}
\name{extra<-}
\alias{extra<-}
\alias{extra<-,SigSet-method}
\title{extra replacement generic}
\usage{
extra(x) <- value

\S4method{extra}{SigSet}(x) <- value
}
\arguments{
\item{x}{object of \code{SigSet}}

\item{value}{new value}
}
\value{
a new \code{SigSet}
}
\description{
extra replacement generic

Replace extra slot of SigSet class
}
\examples{
sset <- sesameDataGet('HM450.1.TCGA.PAAD')$sset
df <- extra(sset)
df[1,1] <- 10
extra(sset) <- df 
}
