% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extractDesign}
\alias{extractDesign}
\title{Extract the first design category}
\usage{
extractDesign(design_str)
}
\arguments{
\item{design_str}{Design string in e.g., the mouse array}
}
\value{
a character vector for the design category
}
\description{
Extract the first design category
}
