% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgroup.R
\name{FromAffine}
\alias{FromAffine}
\title{Convert from affine coordinates to absolute coordinates}
\usage{
FromAffine(dm.affine, dm.sys)
}
\arguments{
\item{dm.affine}{dimension on affine coordinates (relative coordinates)}

\item{dm.sys}{dimension of the affine system}
}
\value{
dimension on the same coordinate system
}
\description{
Convert from affine coordinates to absolute coordinates
}
