% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{getAFs}
\alias{getAFs}
\title{Get allele frequency}
\usage{
getAFs(sdf, ...)
}
\arguments{
\item{sdf}{\code{SigDF}}

\item{...}{additional options to getBetas}
}
\value{
allele frequency
}
\description{
Get allele frequency
}
\examples{
sesameDataCache() # if not done yet
sdf <- sesameDataGet('EPIC.1.SigDF')
af <- getAFs(sdf)
}
