% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{getAFTypeIbySumAlleles}
\alias{getAFTypeIbySumAlleles}
\title{Get allele frequency treating type I by summing alleles}
\usage{
getAFTypeIbySumAlleles(sset, known.ccs.only = TRUE)
}
\arguments{
\item{sset}{\code{SigSet}}

\item{known.ccs.only}{consider only known CCS probes}
}
\value{
beta values
}
\description{
Takes a \code{SigSet} as input and returns a numeric vector containing
extra allele frequencies based on Color-Channel-Switching (CCS) probes.
If no CCS probes exist in the \code{SigSet}, then an numeric(0) is
returned.
}
\examples{
sset <- sesameDataGet('EPIC.1.LNCaP')$sset
betas <- getAFTypeIbySumAlleles(sset)
}
