% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{getBetasTypeIbySumAlleles}
\alias{getBetasTypeIbySumAlleles}
\title{get beta values treating type I by summing alleles}
\usage{
getBetasTypeIbySumAlleles(sset, quality.mask = TRUE,
  nondetection.mask = TRUE, pval.threshold = 0.05)
}
\arguments{
\item{sset}{\code{SignalSet}}

\item{quality.mask}{whether to mask low quality probes}

\item{nondetection.mask}{whether to mask nondetection}

\item{pval.threshold}{p-value threshold for nondetection mask}
}
\value{
beta values
}
\description{
used for extra allele frequency on Color-Channel-Switching CCS probes
}

