% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{getBetasTypeIbySumChannels}
\alias{getBetasTypeIbySumChannels}
\title{Get beta values treating type I by summing channels}
\usage{
getBetasTypeIbySumChannels(sset, quality.mask = TRUE,
  nondetection.mask = TRUE, pval.threshold = 0.05)
}
\arguments{
\item{sset}{\code{SigSet}}

\item{quality.mask}{whether to mask low quality probes}

\item{nondetection.mask}{whether to mask nondetection}

\item{pval.threshold}{p-value threshold for nondetection mask}
}
\value{
beta values
}
\description{
This function is used for rescuing beta values on
Color-Channel-Switching CCS probes. The function takes a \code{SigSet}
and returns beta value except that Type-I in-band signal and out-of-band
signal are combined. This prevents color-channel switching due to SNPs.
}
\examples{
sset <- makeExampleSeSAMeDataSet()
betas <- getBetasTypeIbySumChannels(sset)
}
