% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.R
\name{getBinCoordinates}
\alias{getBinCoordinates}
\title{Get bin coordinates}
\usage{
getBinCoordinates(chrominfo, probe.coords)
}
\arguments{
\item{chrominfo}{chromosome information object}

\item{probe.coords}{probe coordinates}
}
\value{
bin.coords
}
\description{
requires GenomicRanges, IRanges
}

