% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getManifest}
\alias{getManifest}
\title{Retrieve manifest file from the supporting website
at http://zwdzwd.github.io/InfiniumAnnotation}
\usage{
getManifest(
  platform = c("EPIC", "HM450", "HM27"),
  refversion = c("hg19", "hg38"),
  version = "current",
  probeType = c("all", "cg", "ch", "rs"),
  designType = c("all", "I", "II")
)
}
\arguments{
\item{platform}{Infinium platform}

\item{refversion}{human reference version, irrelevant for mouse array}

\item{version}{manifest version, default to the latest/current.}

\item{probeType}{cg, ch or rs, default to all probes}

\item{designType}{I (Infinium-I) or II (Infinium-II), default to both}
}
\value{
manifest file of requested probes
}
\description{
Retrieve manifest file from the supporting website
at http://zwdzwd.github.io/InfiniumAnnotation
}
\examples{

mft <- getManifest('HM27', 'hg38')

}
