% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMostVariableProbes}
\alias{getMostVariableProbes}
\title{Get most variable probes}
\usage{
getMostVariableProbes(betas, n = 2000)
}
\arguments{
\item{betas}{beta value matrix (row: cpg; column: sample)}

\item{n}{number of most variable probes}
}
\value{
beta value matrix for the most variable probes
}
\description{
Get most variable probes
}
\examples{
## get most variable autosome probes
betas <- sesameDataGet('HM450.1.TCGA.PAAD')$betas
betas.most.variable <- getMostVariableProbes(
    betas[getAutosomeProbes('HM450'),],2000)
}
