% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dye_bias.R
\name{getNormCtls}
\alias{getNormCtls}
\title{get normalization control signal}
\usage{
getNormCtls(sset, average = FALSE)
}
\arguments{
\item{sset}{a SigSet}

\item{average}{whether to average}
}
\value{
a data frame of normalization control signals
}
\description{
get normalization control signal from SigSet. 
The function optionally takes mean for each channel.
}
\examples{
sset <- readIDATpair(file.path(system.file(
    'extdata','',package='sesameData'), '4207113116_B'))

df.ctl <- getNormCtls(sset)

}
