% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getProbeByChromosome}
\alias{getProbeByChromosome}
\title{Get probe by chromosome}
\usage{
getProbeByChromosome(chrm, platform)
}
\arguments{
\item{chrm}{chromosome}

\item{platform}{platform}
}
\value{
probes in the specified chromosome
}
\description{
Get probe by chromosome
}
