% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getProbesByChromosome}
\alias{getProbesByChromosome}
\title{Get Probes by Chromosome}
\usage{
getProbesByChromosome(
  chrms,
  platform = c("EPIC", "HM450", "MM285"),
  refversion = NULL
)
}
\arguments{
\item{chrms}{chromosomes to subset}

\item{platform}{EPIC, HM450, Mouse}

\item{refversion}{hg19, hg38, or mm10, inference by default}
}
\value{
a vector of probes on the selected chromosomes
}
\description{
Get Probes by Chromosome
}
\examples{
sex.probes <- getProbesByChromosome(c('chrX','chrY'))
}
