% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{getProbesByGene}
\alias{getProbesByGene}
\title{get probes by gene}
\usage{
getProbesByGene(geneName, platform = "EPIC", refversion = "hg38")
}
\arguments{
\item{geneName}{gene name}

\item{platform}{EPIC or hm450}

\item{refversion}{hg38 or hg19}
}
\value{
probes that fall into the given gene
}
\description{
get probes by gene
}

