% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{getProbesByRegion}
\alias{getProbesByRegion}
\title{get probes by genomic region}
\usage{
getProbesByRegion(chrm, beg, end, platform = "EPIC", refversion = "hg38")
}
\arguments{
\item{chrm}{chromosome}

\item{beg}{begin}

\item{end}{end}

\item{platform}{EPIC or hm450}

\item{refversion}{hg38 or hg19}
}
\value{
probes that fall into the given region
}
\description{
get probes by genomic region
}
