% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{getProbesByTSS}
\alias{getProbesByTSS}
\title{Get Probes by Gene Transcription Start Site (TSS)}
\usage{
getProbesByTSS(geneName, upstream = 1500, dwstream = 1500,
  platform = c("EPIC", "HM450"), refversion = c("hg38", "hg19"))
}
\arguments{
\item{geneName}{gene name}

\item{upstream}{the number of base pairs to expand upstream the TSS}

\item{dwstream}{the number of base pairs to expand dwstream the TSS}

\item{platform}{EPIC or HM450}

\item{refversion}{hg38 or hg19}
}
\value{
probes that fall into the given gene
}
\description{
Get probes mapped to a TSS. All transcripts for the gene are considered.
The function takes a gene name as appears in UCSC RefGene database. The
platform and reference genome build can be changed with `platform` and
`refversion` options. The function returns a vector of probes that falls
into the TSS region of the gene.
}
\examples{
probes <- getProbesByTSS('CDKN2A')
}
