% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellComposition.R
\name{getRefSet}
\alias{getRefSet}
\title{retrieve reference set}
\usage{
getRefSet(cells = NULL, platform = "EPIC")
}
\arguments{
\item{cells}{reference cell types}

\item{platform}{EPIC or hm450}
}
\value{
g
}
\description{
retrieve reference set
}
