% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{getSegment}
\alias{getSegment}
\title{select segment from coefficient table}
\usage{
getSegment(cf1, seg.id)
}
\arguments{
\item{cf1}{coefficient table of one factor from segmentDMR}

\item{seg.id}{segment ID}
}
\value{
coefficient table from given segment
}
\description{
select segment from coefficient table
}

