% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getSesameExample}
\alias{getSesameExample}
\title{Retrieve SeSAMe examples}
\usage{
getSesameExample(nm, platform = "")
}
\arguments{
\item{nm}{name}

\item{platform}{optional, hm450, EPIC or hm27}
}
\value{
example object
}
\description{
Retrieve SeSAMe examples
}

