% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{getSexInfo}
\alias{getSexInfo}
\title{Get sex-related information}
\usage{
getSexInfo(sdf)
}
\arguments{
\item{sdf}{a \code{SigDF}}
}
\value{
medianY and medianX, fraction of XCI, methylated and unmethylated X
probes, median intensities of auto-chromosomes.
}
\description{
The function takes a \code{SigDF} and returns a vector of three
numerics: the median intensity of chrY probes; the median intensity of
chrX probes; and fraction of intermediate chrX probes. chrX and chrY
probes excludes pseudo-autosomal probes.
}
\examples{
sesameDataCache("EPIC") # if not done yet
sdf <- sesameDataGet('EPIC.1.LNCaP')$sdf
getSexInfo(sdf)
}
