% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getVariantAnno_InfiniumI}
\alias{getVariantAnno_InfiniumI}
\title{Retrieve variant annotation file for Infinium-I probes
from the supporting website
at http://zwdzwd.github.io/InfiniumAnnotation}
\usage{
getVariantAnno_InfiniumI(
  platform = c("EPIC"),
  refversion = c("hg19", "hg38"),
  version = "20200704"
)
}
\arguments{
\item{platform}{Infinium platform}

\item{refversion}{human reference version, irrelevant for mouse array}

\item{version}{manifest version, default to the latest/current.}
}
\value{
variant annotation file of infinium I probes
}
\description{
Retrieve variant annotation file for Infinium-I probes
from the supporting website
at http://zwdzwd.github.io/InfiniumAnnotation
}
\examples{

annoI <- getVariantAnno_InfiniumI('EPIC', 'hg38')

}
