% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgroup.R
\name{GroupCheckNameUnique}
\alias{GroupCheckNameUnique}
\title{Check whether group names are unique}
\usage{
GroupCheckNameUnique(group.obj)
}
\arguments{
\item{group.obj}{a WGroup}
}
\value{
TRUE or FALSE
}
\description{
Check whether group names are unique
}
