% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgroup.R
\name{GroupDeepGet}
\alias{GroupDeepGet}
\title{Get an plotting object from a group's descendants}
\usage{
GroupDeepGet(x, nm, force.unique = TRUE)
}
\arguments{
\item{x}{a WGroup object}

\item{nm}{name}

\item{force.unique}{assume the name is unique in the descendants and get one object instead of a list}
}
\value{
if `force.unique==FALSE` return a list. Otherwise, one plotting object.
}
\description{
Get an plotting object from a group's descendants
}
