% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg19.chrominfo}
\alias{hg19.chrominfo}
\title{hg19 chromosome info}
\format{An object of class \code{list} of length 2.}
\usage{
hg19.chrominfo
}
\description{
hg19 chromosome info
}
\keyword{datasets}

