% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg38.chrominfo}
\alias{hg38.chrominfo}
\title{hg38 chromosome info}
\format{An object of class \code{list} of length 2.}
\usage{
hg38.chrominfo
}
\description{
hg38 chromosome info
}
\keyword{datasets}

