% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hm27.controls}
\alias{hm27.controls}
\title{hm27 controls}
\format{An object of class \code{data.frame} with 144 rows and 4 columns.}
\usage{
hm27.controls
}
\description{
hm27 controls
}
\keyword{datasets}

