% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hm27.hg19.probe2chr}
\alias{hm27.hg19.probe2chr}
\title{hm27 probe2chromosome}
\format{An object of class \code{factor} of length 27588.}
\usage{
hm27.hg19.probe2chr
}
\description{
hm27 probe2chromosome
}
\keyword{datasets}

