% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hm27.ordering}
\alias{hm27.ordering}
\title{hm27 data ordering}
\format{An object of class \code{data.frame} with 27578 rows and 6 columns.}
\usage{
hm27.ordering
}
\description{
hm27 data ordering
}
\keyword{datasets}

