% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hm450.controls}
\alias{hm450.controls}
\title{hm450 controls}
\format{An object of class \code{data.frame} with 850 rows and 4 columns.}
\usage{
hm450.controls
}
\description{
hm450 controls
}
\keyword{datasets}

