% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hm450.female.clean.chrY.probes}
\alias{hm450.female.clean.chrY.probes}
\title{hm450 female clean chrY probes}
\format{An object of class \code{character} of length 337.}
\usage{
hm450.female.clean.chrY.probes
}
\description{
hm450 female clean chrY probes
}
\keyword{datasets}

