% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hm450.female.xlinked.chrX.probes}
\alias{hm450.female.xlinked.chrX.probes}
\title{hm450 female X linked probes}
\format{An object of class \code{character} of length 3797.}
\usage{
hm450.female.xlinked.chrX.probes
}
\description{
hm450 female X linked probes
}
\keyword{datasets}

