% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hm450.hg19.probe2chr}
\alias{hm450.hg19.probe2chr}
\title{hm450 probe2chromosome}
\format{An object of class \code{factor} of length 485577.}
\usage{
hm450.hg19.probe2chr
}
\description{
hm450 probe2chromosome
}
\keyword{datasets}

