% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hm450.mapped.probes.hg19}
\alias{hm450.mapped.probes.hg19}
\title{hm450 mapped probes hg19}
\format{An object of class \code{GRanges} of length 467097.}
\usage{
hm450.mapped.probes.hg19
}
\description{
hm450 mapped probes hg19
}
\keyword{datasets}

