% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hm450.mapped.probes.hg38}
\alias{hm450.mapped.probes.hg38}
\title{hm450 mapped probes hg38}
\format{An object of class \code{GRanges} of length 466007.}
\usage{
hm450.mapped.probes.hg38
}
\description{
hm450 mapped probes hg38
}
\keyword{datasets}

