% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hm450.ordering}
\alias{hm450.ordering}
\title{hm450 data ordering}
\format{An object of class \code{data.frame} with 485577 rows and 6 columns.}
\usage{
hm450.ordering
}
\description{
hm450 data ordering
}
\keyword{datasets}

