% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigSetMethods.R
\docType{methods}
\name{IG<-}
\alias{IG<-}
\alias{IG<-,SigSet-method}
\title{IG replacement generic}
\usage{
IG(x) <- value

\S4method{IG}{SigSet}(x) <- value
}
\arguments{
\item{x}{object of \code{SigSet}}

\item{value}{new value}
}
\value{
a new \code{SigSet}
}
\description{
IG replacement generic

Replace IG slot of SigSet class
}
\examples{

sesameDataCache("HM450") # if not done yet
sset <- sesameDataGet('HM450.1.TCGA.PAAD')$sset
df <- IG(sset)
df[1,1] <- 10
IG(sset) <- df 
}
