% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigSetMethods.R
\name{IGpass}
\alias{IGpass}
\title{Get IG slot of SigSet class that passes mask
This is the same as IG() if there is no mask set}
\usage{
IGpass(sset)
}
\arguments{
\item{sset}{SigSet object}
}
\value{
The IG slot that passes extra$mask filter
}
\description{
Get IG slot of SigSet class that passes mask
This is the same as IG() if there is no mask set
}
\examples{

sesameDataCache("EPIC") # if not done yet
sset <- sesameDataGet('EPIC.1.LNCaP')$sset
head(IGpass(sset))
}
