% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigSetMethods.R
\docType{methods}
\name{II}
\alias{II}
\alias{II,SigSet-method}
\title{II getter generic}
\usage{
II(x)

\S4method{II}{SigSet}(x)
}
\arguments{
\item{x}{object of \code{SigSet}}
}
\value{
The II slot of \code{SigSet}
}
\description{
II getter generic

Get II slot of SigSet class
}
\examples{
sesameDataCache("HM450") # if not done yet
sset <- sesameDataGet('HM450.1.TCGA.PAAD')$sset
head(II(sset))
}
