% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigSetMethods.R
\docType{methods}
\name{II<-}
\alias{II<-}
\alias{II<-,SigSet-method}
\title{II replacement generic}
\usage{
II(x) <- value

\S4method{II}{SigSet}(x) <- value
}
\arguments{
\item{x}{object of \code{SigSet}}

\item{value}{new value}
}
\value{
a new \code{SigSet}
}
\description{
II replacement generic

Replace II slot of SigSet class
}
\examples{
sesameDataCache("HM450") # if not done yet
sset <- sesameDataGet('HM450.1.TCGA.PAAD')$sset
df <- II(sset)
df[1,1] <- 10
II(sset) <- df 
}
