% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigSetMethods.R
\name{IIpass}
\alias{IIpass}
\title{Get II slot of SigSet class that passes mask
This is the same as II() if there is no mask set}
\usage{
IIpass(sset)
}
\arguments{
\item{sset}{SigSet object}
}
\value{
The II slot that passes extra$mask filter
}
\description{
Get II slot of SigSet class that passes mask
This is the same as II() if there is no mask set
}
\examples{
sesameDataCache("EPIC") # if not done yet
sset <- sesameDataGet('EPIC.1.LNCaP')$sset
head(IIpass(sset))
}
