% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{imputeTo}
\alias{imputeTo}
\title{Impute to platform}
\usage{
imputeTo(betas, platform = NULL, probes = NULL, tissue = "Blood")
}
\arguments{
\item{betas}{named vector or matrix of beta values}

\item{platform}{platform of target imputation}

\item{probes}{Probe ID if not set to vector names}

\item{tissue}{tissue context of imputation}
}
\value{
imputed data, vector or matrix
}
\description{
Impute to platform
}
\examples{

betas <- c("cg04707299"=0.2, "cg13380562"=0.9, "cg00000103"=0.1)
betas_imputed <- imputeTo(betas, "HM450")

betas <- setNames(seq(0,1,length.out=3),
    c("cg00004963_TC21", "cg00004963_TC22", "cg00004747_TC21"))
betas_imputed <- imputeTo(betas, "HM450")

}
