% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{inferOobG}
\alias{inferOobG}
\title{Infer out-of-band red instead of using what is specified in manifest}
\usage{
inferOobG(sset)
}
\arguments{
\item{sset}{a \code{SigSet}}
}
\value{
out-of-band data frame
}
\description{
Infer out-of-band red instead of using what is specified in manifest
}
