% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{inferOobR}
\alias{inferOobR}
\title{Infer out-of-band green instead of using what is specified in manifest}
\usage{
inferOobR(sset)
}
\arguments{
\item{sset}{a \code{SigSet}}
}
\value{
out-of-band data frame
}
\description{
Infer out-of-band green instead of using what is specified in manifest
}
