% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex.R
\name{inferSex}
\alias{inferSex}
\title{Infer Sex}
\usage{
inferSex(x, platform = NULL, verbose = FALSE)
}
\arguments{
\item{x}{either a raw \code{SigDF} or a beta value vector named by probe ID
SigDF is preferred over beta values.}

\item{platform}{Only MM285, EPIC and HM450 are supported.}

\item{verbose}{print more messages}
}
\value{
'F' or 'M'
We established our sex calling based on the CpGs hypermethylated in
inactive X (XiH), CpGs hypomethylated in inactive X (XiL) and signal
intensity ratio of Y-chromosome over autosomes. Currently human inference
uses a random forest and mouse inference uses a support vector machine.

The function checks the sample quality. If the sample is of poor quality
the inference return NA.

Note many factors such as Dnmt genotype, XXY male (Klinefelter's),
45,X female (Turner's) can confuse the model sometimes.
This function works on a single sample.
}
\description{
Infer Sex
}
\examples{
sesameDataCache() # if not done yet
sdf <- sesameDataGet('EPIC.1.SigDF')
inferSex(sdf)
}
