% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{inferTypeIChannel}
\alias{inferTypeIChannel}
\title{Infer and reset color channel for Type-I probes instead of
using what is specified in manifest}
\usage{
inferTypeIChannel(sset)
}
\arguments{
\item{sset}{a \code{SigSet}}
}
\value{
a \code{SigSet}
}
\description{
Infer and reset color channel for Type-I probes instead of
using what is specified in manifest
}
\examples{

sset <- sesameDataGet('EPIC.1.LNCaP')$sset
inferTypeIChannel(sset)

}
