% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigSetMethods.R
\docType{methods}
\name{IR}
\alias{IR}
\alias{IR,SigSet-method}
\title{IR getter generic}
\usage{
IR(x)

\S4method{IR}{SigSet}(x)
}
\arguments{
\item{x}{object of \code{SigSet}}
}
\value{
The IR slot of \code{SigSet}
}
\description{
IR getter generic

Get IR slot of SigSet class
}
\examples{
sesameDataCache("HM450") # if not done yet
sset <- sesameDataGet('HM450.1.TCGA.PAAD')$sset
head(IR(sset))
}
