% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigSetMethods.R
\name{IRpass}
\alias{IRpass}
\title{Get IR slot of SigSet class that passes mask
This is the same as IR() if there is no mask set}
\usage{
IRpass(sset)
}
\arguments{
\item{sset}{SigSet object}
}
\value{
The IR slot that passes extra$mask filter
}
\description{
Get IR slot of SigSet class that passes mask
This is the same as IR() if there is no mask set
}
\examples{
sesameDataCache("EPIC") # if not done yet
sset <- sesameDataGet('EPIC.1.LNCaP')$sset
head(IRpass(sset))
}
