% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isUniqProbeID}
\alias{isUniqProbeID}
\title{Whether the probe ID is the uniq probe ID like in
the mouse array, e.g., cg36609548}
\usage{
isUniqProbeID(Probe_ID)
}
\arguments{
\item{Probe_ID}{Probe ID}
}
\value{
a logical(1), whether the probe ID is based on the new ID system
}
\description{
Whether the probe ID is the uniq probe ID like in
the mouse array, e.g., cg36609548
}
