% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KYCG.R
\name{KYCG_buildGeneDBs}
\alias{KYCG_buildGeneDBs}
\title{build gene-probe association database}
\usage{
KYCG_buildGeneDBs(
  query = NULL,
  platform = NULL,
  max_distance = 10000,
  silent = FALSE
)
}
\arguments{
\item{query}{the query probe list. If NULL, use all the probes
on the platform}

\item{platform}{HM450, EPIC, MM285, Mammal40, will infer from
query if not given}

\item{max_distance}{probe-gene distance for association}

\item{silent}{suppress messages}
}
\value{
gene databases
}
\description{
build gene-probe association database
}
\examples{
query <- c("cg04707299", "cg13380562", "cg00480749")
dbs <- KYCG_buildGeneDBs(query, platform = "EPIC")
testEnrichment(query, dbs, platform = "EPIC")
}
