% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KYCG.R
\name{KYCG_downloadDBs}
\alias{KYCG_downloadDBs}
\title{A convenience function for downloading knowledgebase sets}
\usage{
KYCG_downloadDBs(platform, fdr)
}
\arguments{
\item{platform}{EPICv2, EPIC, HM450 etc.}

\item{fdr}{directory to which feature files will be downloaded}
}
\value{
untarred feature folders
}
\description{
A convenience function for downloading knowledgebase sets
}
