% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KYCG.R
\name{KYCG_getDBs}
\alias{KYCG_getDBs}
\title{Get databases by full or partial names of the database group(s)}
\usage{
KYCG_getDBs(group_nms)
}
\arguments{
\item{group_nms}{database group names}
}
\value{
a list of databases
}
\description{
Get databases by full or partial names of the database group(s)
}
\examples{
dbs <- KYCG_getDBs("MM285.chromHMM")
}
