% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KYCG.R
\name{KYCG_listDBGroups}
\alias{KYCG_listDBGroups}
\title{List database group names}
\usage{
KYCG_listDBGroups(filter = NULL, type = NULL)
}
\arguments{
\item{filter}{keywords for filtering}

\item{type}{categorical, numerical (default: all)}
}
\value{
a list of db group names
}
\description{
List database group names
}
\examples{
head(KYCG_listDBGroups("chromHMM"))
}
