% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KYCG.R
\name{KYCG_loadDBs}
\alias{KYCG_loadDBs}
\title{Load database groups}
\usage{
KYCG_loadDBs(in_paths, group_use_filename = FALSE)
}
\arguments{
\item{in_paths}{folder that contains all databases}

\item{group_use_filename}{whether to use file name for groups}
}
\value{
a list of db group names
}
\description{
Load database groups
}
\examples{

## download regulatory annotations from
## http://zwdzwd.github.io/InfiniumAnnotation
## unzip the file
if (FALSE) {
dbs <- KYCG_loadDBs(path_to_unzipped_folder)
}
}
