% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KYCG_plot.R
\name{KYCG_plotEnrichAll}
\alias{KYCG_plotEnrichAll}
\title{plot enrichment test result}
\usage{
KYCG_plotEnrichAll(df, fdr_max = 25)
}
\arguments{
\item{df}{test enrichment result data frame}
}
\value{
grid object
}
\description{
plot enrichment test result
}
