% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kyCG_plot.R
\name{KYCG_plotLollipop}
\alias{KYCG_plotLollipop}
\title{creates a lollipop plot of log(estimate) given data with
fields estimate.}
\usage{
KYCG_plotLollipop(df, n = 10)
}
\arguments{
\item{df}{DataFrame where each field is a database name with its
estimate.}

\item{n}{Integer representing the number of top enrichments to report.
Optional. (Default: 10)}
}
\value{
ggplot lollipop plot
}
\description{
creates a lollipop plot of log(estimate) given data with
fields estimate.
}
\examples{
KYCG_plotLollipop(data.frame(estimate=c(runif(10, 0, 10))))

}
