% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KYCG_plot.R
\name{KYCG_plotManhattan}
\alias{KYCG_plotManhattan}
\title{KYCG_plotManhattan makes a manhattan plot to summarize EWAS results}
\usage{
KYCG_plotManhattan(
  vals,
  platform = NULL,
  gr = NULL,
  genome = NULL,
  title = NULL,
  label_min = 100,
  col = c("wheat1", "sienna3"),
  ylabel = "Value"
)
}
\arguments{
\item{vals}{named vector of values (P,Q etc), vector name is Probe ID.}

\item{platform}{String corresponding to the type of platform to use for
retrieving GRanges coordinates 
of probes. Either MM285, EPIC, HM450, or HM27. If it is not provided, it
will be inferred from the query set probeIDs (Default: NA).}

\item{gr}{GRanges object containing genomic coordinates of probes. If not
provided, GRanges for provided or inferred platform will be retrieved}

\item{genome}{genome build}

\item{title}{title for plot}

\item{label_min}{Threshold above which data points will be labelled with
Probe ID}

\item{col}{color}

\item{ylabel}{y-axis label}
}
\value{
a ggplot object
}
\description{
KYCG_plotManhattan makes a manhattan plot to summarize EWAS results
}
\examples{

## see vignette for examples
sesameDataGet_resetEnv()

}
