% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KYCG_plot.R
\name{KYCG_plotWaterfall}
\alias{KYCG_plotWaterfall}
\title{create a waterfall plot of log(estimate) given test enrichment}
\usage{
KYCG_plotWaterfall(df, n_label = 10, label_column = "dbname")
}
\arguments{
\item{df}{data frame where each row is a database with test
enrichment result}

\item{n_label}{number of datapoints to label}

\item{label_column}{column in df to be used as the label (default: dbname)}
}
\value{
grid
}
\description{
create a waterfall plot of log(estimate) given test enrichment
}
\examples{

library(SummarizedExperiment)
df <- rowData(sesameDataGet('MM285.tissueSignature'))
query <- df$Probe_ID[df$branch == "fetal_brain" & df$type == "Hypo"]
results <- testEnrichment(query, "TFBS")
KYCG_plotWaterfall(results)

}
