% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dims.R
\name{LeftOf}
\alias{LeftOf}
\title{LeftOf}
\usage{
LeftOf(x = NULL, width = NULL, pad = 0.01, min.ratio = 0.02,
  v.aln = NULL, h.scale = NULL, h.scale.proportional = FALSE)
}
\arguments{
\item{x}{an object with dimension}

\item{width}{the width of the new object (when NULL, set proportional to data)}

\item{pad}{padding between the target and current}

\item{min.ratio}{minimum ratio of dimensions when auto-scale}

\item{v.aln}{object for vertical alignment (when NULL, set to x)}

\item{h.scale}{object for horizontal scaling (when NULL, set to x)}

\item{h.scale.proportional}{when h.scale is provided, whether to make proportional to data}
}
\value{
a dimension to the left of x
}
\description{
Generate dimension to the left of another object
}
