% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{liftOver}
\alias{liftOver}
\title{lift over beta values or SigDFs to another Infinium platform
This function wraps ID conversion and provide optional
imputation functionality.}
\usage{
liftOver(
  x,
  target_platform,
  mapping = NULL,
  impute = FALSE,
  celltype = "Blood"
)
}
\arguments{
\item{x}{either beta value (vector or matrix) or SigDF(s)}

\item{target_platform}{the platform to take the data to}

\item{mapping}{a liftOver mapping file. Typically this file
contains empirical evidence whether a probe mapping is reliable.
If given, probe ID-based mapping will be skipped. This is to
perform more stringent probe ID mapping.}

\item{impute}{whether to impute or not, default is FALSE}

\item{celltype}{the cell type / tissue context of imputation,
if not given, will use nearest neighbor to find out.}
}
\value{
imputed data, vector, matrix, SigDF(s)
}
\description{
lift over beta values or SigDFs to another Infinium platform
This function wraps ID conversion and provide optional
imputation functionality.
}
\examples{

sesameDataCache()
## sdf = sesameDataGet("EPICv2.8.SigDF")[["GM12878_206909630042_R08C01"]]
## betas = openSesame(sdf)
## betas_HM450 = liftOver(betas, "HM450", impute=TRUE)
}
