% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{listAvailableMasks}
\alias{listAvailableMasks}
\title{list existing quality masks for a SigDF}
\usage{
listAvailableMasks(sdf)
}
\arguments{
\item{sdf}{a SigDF object}
}
\value{
a tibble of masks
}
\description{
list existing quality masks for a SigDF
}
\examples{
sdf <- sesameDataGet('EPIC.1.SigDF')
listAvailableMasks(sdf)
}
