% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{makeExampleSeSAMeDataSet}
\alias{makeExampleSeSAMeDataSet}
\title{Make a simulated SeSAMe data set}
\usage{
makeExampleSeSAMeDataSet(platform = "HM450")
}
\arguments{
\item{platform}{optional, HM450, EPIC or HM27}
}
\value{
Object of class \code{SignalSet}
}
\description{
Constructs a simulated dataset.
}
\examples{
sset <- makeExampleSeSAMeDataSet()

}
