% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colormap.R
\name{MapToContinuousColors}
\alias{MapToContinuousColors}
\title{map data to continuous color}
\usage{
MapToContinuousColors(data, cmp = CMPar(), given.cm = NULL)
}
\arguments{
\item{data}{numeric vector}

\item{cmp}{an color map parameter object of class CMPar}

\item{given.cm}{given colormap}
}
\value{
an object of ColorMap
}
\description{
map data to continuous color
}
