% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colormap.R
\name{MapToDiscreteColors}
\alias{MapToDiscreteColors}
\title{map data to discrete color}
\usage{
MapToDiscreteColors(data, cmp = CMPar(), given.cm = NULL)
}
\arguments{
\item{data}{numeric vector}

\item{cmp}{an color map parameter object of class CMPar}

\item{given.cm}{given color map}
}
\value{
an object of ColorMap
}
\description{
map data to discrete color
}
