% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{mapToMammal40}
\alias{mapToMammal40}
\title{Map the SDF (from overlap array platforms)
Replicates are merged by picking the best detection}
\usage{
mapToMammal40(sdf)
}
\arguments{
\item{sdf}{a \code{SigDF} object}
}
\value{
a named numeric vector for beta values
}
\description{
Map the SDF (from overlap array platforms)
Replicates are merged by picking the best detection
}
\examples{
sdf <- sesameDataGet("Mammal40.1.SigDF")
betas <- mapToMammal40(sdf[1:10,])
}
