% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigSetMethods.R
\docType{methods}
\name{mask}
\alias{mask}
\alias{mask,SigSet-method}
\title{Get Probe Names of SigSet class}
\usage{
mask(x)

\S4method{mask}{SigSet}(x)
}
\arguments{
\item{x}{object of \code{Sigset}}
}
\value{
A char vector
}
\description{
Get Probe Names of SigSet class
}
\examples{
sesameDataCache("HM450") # if not done yet
sset <- sesameDataGet('HM450.1.TCGA.PAAD')$sset
head(mask(sset))
}
