% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigSetMethods.R
\name{mask}
\alias{mask}
\title{Get current mask of SigSet class}
\usage{
mask(x)
}
\arguments{
\item{x}{object of \code{Sigset}}
}
\value{
A named logical vector
}
\description{
Get current mask of SigSet class
}
\examples{
sesameDataCache("HM450") # if not done yet
sset <- sesameDataGet('HM450.1.TCGA.PAAD')$sset
head(mask(sset))
}
