% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{maskRepeatSNPs}
\alias{maskRepeatSNPs}
\title{Mask repeat and SNPs}
\usage{
maskRepeatSNPs(x, platform)
}
\arguments{
\item{x}{a \code{SignalSet}, a beta value vector or matrix}

\item{platform}{platfrom}
}
\value{
masked object of the corresponding class of x
}
\description{
Mask repeat and SNPs
}

