% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{medianIntensity}
\alias{medianIntensity}
\title{Whole-dataset-wide Median Intensity}
\usage{
medianIntensity(sdf, mask = TRUE)
}
\arguments{
\item{sdf}{a \code{SigDF}}

\item{mask}{whether to mask probes using mask column}
}
\value{
median of all intensities
}
\description{
The function takes one single \code{SigDF} and computes median
intensity of all the in-band measurements. This includes all Type-I
in-band measurements and all Type-II probe measurements. Both methylated
and unmethylated alleles are considered. This function outputs a single
numeric for the median.
}
\examples{
sesameDataCache("EPIC") # if not done yet
sdf <- sesameDataGet('EPIC.1.SigDF')
medianIntensity(sdf)
}
