% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{medianTotalIntensity}
\alias{medianTotalIntensity}
\title{Whole-dataset-wide Median Total Intensity (M+U)}
\usage{
medianTotalIntensity(sdf, mask = TRUE)
}
\arguments{
\item{sdf}{a \code{SigDF}}

\item{mask}{whether to mask probes using mask column}
}
\value{
median of all intensities
}
\description{
The function takes one single \code{SigDF} and computes median
intensity of M+U for each probe. This function outputs a single
numeric for the median.
}
\examples{
sesameDataCache("EPIC") # if not done yet
sdf <- sesameDataGet('EPIC.1.SigDF')
medianIntensity(sdf)
}
