% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MValueToBetaValue}
\alias{MValueToBetaValue}
\title{Convert M-value to beta-value}
\usage{
MValueToBetaValue(m)
}
\arguments{
\item{m}{a vector of M values}
}
\value{
a vector of beta values
}
\description{
Convert M-value to beta-value (aka inverse logit transform)
}
\examples{
MValueToBetaValue(c(-3, 0, 3))
}
