% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{negControls}
\alias{negControls}
\title{get negative control signal}
\usage{
negControls(sdf)
}
\arguments{
\item{sdf}{a SigDF}
}
\value{
a data frame of negative control signals
}
\description{
get negative control signal
}
