% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_correction.R
\name{neob}
\alias{neob}
\title{Negative control plus out-of-band background correction}
\usage{
neob(sdf, offset = 15)
}
\arguments{
\item{sdf}{a \code{SigDF}}

\item{offset}{offset}
}
\value{
a new \code{SigDF} with neob background correction
}
\description{
The function takes a \code{SigDF} and returns a modified \code{SigDF}
with background subtracted. Background was modelled in a normal distribution
and true signal in an exponential distribution. The Norm-Exp deconvolution
is parameterized using both negative control and Out-Of-Band (oob) probes
}
\examples{
sdf <- sesameDataGet('EPIC.1.SigDF')
sdf.nb <- neob(sdf)
}
