% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigDFMethods.R
\name{noMasked}
\alias{noMasked}
\title{remove masked probes from SigDF}
\usage{
noMasked(sdf)
}
\arguments{
\item{sdf}{input SigDF object}
}
\value{
a SigDF object without masked probes
}
\description{
remove masked probes from SigDF
}
\examples{
sesameDataCache()
sdf <- sesameDataGet("EPIC.1.SigDF")
sdf <- pOOBAH(sdf)

sdf_noMasked <- noMasked(sdf)

}
