% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dye_bias.R
\name{normControls}
\alias{normControls}
\title{get normalization control signal}
\usage{
normControls(sdf, average = FALSE, verbose = FALSE)
}
\arguments{
\item{sdf}{a SigDF}

\item{average}{whether to average}

\item{verbose}{print more messages}
}
\value{
a data frame of normalization control signals
}
\description{
get normalization control signal from SigDF. 
The function optionally takes mean for each channel.
}
