% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigSetMethods.R
\docType{methods}
\name{oobG}
\alias{oobG}
\alias{oobG,SigSet-method}
\title{oobG getter generic}
\usage{
oobG(x)

\S4method{oobG}{SigSet}(x)
}
\arguments{
\item{x}{object of \code{SigSet}}
}
\value{
The oobG slot of \code{SigSet}
}
\description{
oobG getter generic

Get oobG slot of SigSet class
}
\examples{
sesameDataCache("HM450") # if not done yet
sset <- sesameDataGet('HM450.1.TCGA.PAAD')$sset
head(oobG(sset))
}
