% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigSetMethods.R
\name{oobGpass}
\alias{oobGpass}
\title{Get oobG slot of SigSet class that passes mask
This is the same as oobG() if there is no mask set}
\usage{
oobGpass(sset)
}
\arguments{
\item{sset}{SigSet object}
}
\value{
The oobG slot that passes extra$mask filter
}
\description{
Get oobG slot of SigSet class that passes mask
This is the same as oobG() if there is no mask set
}
\examples{
sesameDataCache("EPIC") # if not done yet
sset <- sesameDataGet('EPIC.1.LNCaP')$sset
head(oobGpass(sset))
}
