% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigSetMethods.R
\docType{methods}
\name{oobR}
\alias{oobR}
\alias{oobR,SigSet-method}
\title{oobR getter generic}
\usage{
oobR(x)

\S4method{oobR}{SigSet}(x)
}
\arguments{
\item{x}{object of \code{SigSet}}
}
\value{
The oobR slot of \code{SigSet}
}
\description{
oobR getter generic

Get oobR slot of SigSet class
}
\examples{
sesameDataCache("HM450") # if not done yet
sset <- sesameDataGet('HM450.1.TCGA.PAAD')$sset
head(oobR(sset))
}
