% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigSetMethods.R
\name{oobRpass}
\alias{oobRpass}
\title{Get oobR slot of SigSet class that passes mask
This is the same as oobR() if there is no mask set}
\usage{
oobRpass(sset)
}
\arguments{
\item{sset}{SigSet object}
}
\value{
The oobR slot that passes extra$mask filter
}
\description{
Get oobR slot of SigSet class that passes mask
This is the same as oobR() if there is no mask set
}
\examples{
sesameDataCache("EPIC") # if not done yet
sset <- sesameDataGet('EPIC.1.LNCaP')$sset
head(oobRpass(sset))
}
