% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palgen.R
\name{palgen}
\alias{palgen}
\title{Generate some additional color palettes}
\usage{
palgen(pal, n = 150, space = "Lab")
}
\arguments{
\item{pal}{a string for adhoc pals}

\item{n}{the number of colors for interpolation}

\item{space}{rgb or Lab}
}
\value{
a palette-generating function
}
\description{
Generate some additional color palettes
}
\examples{
library(pals)
pal.bands(palgen("whiteturbo"))
}
