% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO.R
\name{parseGEOSignalABFile}
\alias{parseGEOSignalABFile}
\title{Parse GEO signal-A/B File into a SigSet List}
\usage{
parseGEOSignalABFile(path, platform = "HM450", drop = TRUE,
  parallel = TRUE)
}
\arguments{
\item{path}{path to Signal-A/B file downlaoded from GEO. 
The file can remain gzipped.}

\item{platform}{HM450, EPIC or HM27}

\item{drop}{whether to reduce to SigSet when there is only
one sample.}

\item{parallel}{whether to use multiple cores.}
}
\value{
a SigSetList or a SigSet
}
\description{
This function is meant to be a convenience function for 
parsing data from Signal_A and Signal_B file provided by
GEO. In many cases, this function generates a "partial"
SigSet due to lack of out-of-band signal and control 
probe measurement in those Signal_A/B files.
The detection p-value is based on a fixed normal distribution
rather than from negative control or OOB probes.
}
\examples{
path = system.file(
    'extdata',
    'GSE36369_NonEBV_SignalA_SignalB_3samples_1k.txt.gz',
    package='sesame')
ssets <- parseGEOSignalABFile(path)
}
