% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigDFMethods.R
\name{platform}
\alias{platform}
\title{Convenience function to output platform attribute of SigDF}
\usage{
platform(sdf)
}
\arguments{
\item{sdf}{a SigDF object}
}
\description{
Convenience function to output platform attribute of SigDF
}
\examples{
sesameDataCache("EPIC")
sdf = = sesameDataGet('EPIC.1.SigDF')
platform(sdf)
}
