% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{+.WObject}
\alias{+.WObject}
\title{merge plotting objects}
\usage{
\method{+}{WObject}(group, p)
}
\arguments{
\item{group}{a WGroup or a plotting object}

\item{p}{a new plotting object}
}
\value{
a WGroup
}
\description{
merge plotting objects
}
