% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{predictAgePheno}
\alias{predictAgePheno}
\title{Phenotypic age predictor}
\usage{
predictAgePheno(betas)
}
\arguments{
\item{betas}{a probeID-named vector of beta values}
}
\value{
age in years
}
\description{
The function takes a named numeric vector of beta values. The name attribute
contains the probe ID (cg, ch or rs IDs). The function looks for overlapping
probes and estimate age using Horvath aging model (Levine et al. 2018
Aging, 513 probes). The function outputs a single numeric of age in years.
}
\examples{
betas <- sesameDataGet('HM450.1.TCGA.PAAD')$betas
predictAgePheno(betas)
}
