% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{predictMouseAgeInMonth}
\alias{predictMouseAgeInMonth}
\title{Mouse age predictor}
\usage{
predictMouseAgeInMonth(betas, na_fallback = TRUE)
}
\arguments{
\item{betas}{a probeID-named vector of beta values}

\item{na_fallback}{use the fallback default for NAs.}
}
\value{
age in month
}
\description{
The function takes a named numeric vector of beta values. The name attribute
contains the probe ID. The function looks for overlapping
probes and estimate age using an aging model built from 321 MM285 probes.
The function outputs a single numeric of age in months. The clock is most
accurate with the sesame preprocessing.
}
\examples{

betas = sesameDataGet('MM285.10.tissue')$betas
predictMouseAgeInMonth(betas[,1])
sesameDataClearCache()

}
