% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{prefixMaskButCG}
\alias{prefixMaskButCG}
\title{Mask all but CG probes in SigDF}
\usage{
prefixMaskButCG(sdf)
}
\arguments{
\item{sdf}{SigDF}
}
\value{
SigDF
}
\description{
Mask all but CG probes in SigDF
}
\examples{
sdf <- prefixMaskButCG(sesameDataGet("MM285.1.SigDF"))
}
