% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{prefixMaskCtlUK}
\alias{prefixMaskCtlUK}
\title{Mask control and uk probes in SigDF}
\usage{
prefixMaskCtlUK(sdf)
}
\arguments{
\item{sdf}{SigDF}
}
\value{
SigDF
}
\description{
Mask control and uk probes in SigDF
}
\examples{
sdf <- prefixMaskCtlUK(sesameDataGet("MM285.1.SigDF"))
}
