% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{prepSesame}
\alias{prepSesame}
\title{Apply a chain of sesame preprocessing functions in an arbitrary order}
\usage{
prepSesame(sdf, code = "UCDPB")
}
\arguments{
\item{sdf}{SigDF}

\item{code}{code that indicates preprocessing functions and their
execution order (functions on the left is executed first).}
}
\value{
SigDF
}
\description{
Apply a chain of sesame preprocessing functions in an arbitrary order
}
\examples{
sdf <- sesameDataGet("MM285.1.SigDF")
sdf1 <- prepSesame(sdf, "UCDPB")
}
