% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{prepSesameList}
\alias{prepSesameList}
\title{List supported prepSesame functions}
\usage{
prepSesameList()
}
\value{
a data frame with code, func, description
}
\description{
List supported prepSesame functions
}
\examples{
prepSesameList()
}
