% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{print.sesameQC}
\alias{print.sesameQC}
\title{Print sesameQC object}
\usage{
\method{print}{sesameQC}(x, ...)
}
\arguments{
\item{x}{a sesameQC object}

\item{...}{extra parameter for print}
}
\value{
print sesameQC result on screen
}
\description{
Print sesameQC object
}
\examples{
sesameDataCache("EPIC") # if not done yet
sdf <- sesameDataGet('EPIC.1.SigDF')
sesameQC(sdf)

# release memory for Windows package builder
rm(list=ls(env=sesameData:::cacheEnv), envir=sesameData:::cacheEnv)
gc()

}
