% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogram.R
\name{print.WDendrogram}
\alias{print.WDendrogram}
\title{print a dendrogram}
\usage{
\method{print}{WDendrogram}(x, stand.alone = TRUE, layout.only = FALSE,
  cex = 1, ...)
}
\arguments{
\item{x}{a dendrogram}

\item{stand.alone}{plot is stand alone}

\item{layout.only}{plot layout only}

\item{cex}{factor to scaling texts}

\item{...}{additional options (ignored)}
}
\description{
print a dendrogram
}
