% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{print.WGenerator}
\alias{print.WGenerator}
\title{print a WGenerator}
\usage{
\method{print}{WGenerator}(x, ...)
}
\arguments{
\item{x}{a WGenerator object}

\item{...}{additional options}
}
\value{
the WGroup containing the plotting object
}
\description{
This calls WGenerator and creates a WGroup to enclose the produced object.
}
