% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{print.WGG}
\alias{print.WGG}
\title{plot WGG object}
\usage{
\method{print}{WGG}(x, cex = 1, layout.only = FALSE, stand.alone = TRUE,
  ...)
}
\arguments{
\item{x}{WGG}

\item{cex}{scaling factor for text}

\item{layout.only}{plot layout}

\item{stand.alone}{produce a stand.alone plot}

\item{...}{extra options}
}
\value{
printed ggobj object
}
\description{
plot WGG object
}
