% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob.R
\name{print.WGrob}
\alias{print.WGrob}
\title{plot WGrob object}
\usage{
\method{print}{WGrob}(x, cex = 1, layout.only = FALSE, stand.alone = TRUE,
  ...)
}
\arguments{
\item{x}{WGrob}

\item{cex}{scaling factor for text}

\item{layout.only}{plot layout}

\item{stand.alone}{produce a stand.alone plot}

\item{...}{extra options}
}
\description{
plot WGrob object
}
