% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgroup.R
\name{print.WGroup}
\alias{print.WGroup}
\title{Draw WGroup}
\usage{
\method{print}{WGroup}(x, stand.alone = TRUE, cex = 1,
  layout.only = FALSE, ...)
}
\arguments{
\item{x}{a WGroup}

\item{stand.alone}{to plot stand alone}

\item{cex}{factor for scaling fonts}

\item{layout.only}{to plot layout only}

\item{...}{additional options}
}
\description{
Draw WGroup
}
