% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wheatmap.R
\name{print.WHeatmap}
\alias{print.WHeatmap}
\title{plot WHeatmap}
\usage{
\method{print}{WHeatmap}(x, cex = 1, layout.only = FALSE,
  stand.alone = TRUE, ...)
}
\arguments{
\item{x}{a WHeatmap}

\item{cex}{factor to scaling texts}

\item{layout.only}{plot layout only}

\item{stand.alone}{plot is stand alone}

\item{...}{additional options}
}
\value{
\code{NULL}
}
\description{
plot WHeatmap
}
