% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{print.WLabel}
\alias{print.WLabel}
\title{print WLabel}
\usage{
\method{print}{WLabel}(x, cex = 1, layout.only = FALSE,
  stand.alone = TRUE, ...)
}
\arguments{
\item{x}{a WLabel object}

\item{cex}{factor to scale text}

\item{layout.only}{plot layout only}

\item{stand.alone}{plot label stand alone}

\item{...}{additional options}
}
\description{
print WLabel
}
