% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rect.R
\name{print.WRect}
\alias{print.WRect}
\title{print WRect}
\usage{
\method{print}{WRect}(x, cex = 1, layout.only = FALSE, stand.alone = TRUE,
  ...)
}
\arguments{
\item{x}{a WRect object}

\item{cex}{factor for scaling text}

\item{layout.only}{print layout only}

\item{stand.alone}{plot WRect standalone}

\item{...}{additional options}
}
\description{
print WRect
}
