% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigSetMethods.R
\docType{methods}
\name{probeNames}
\alias{probeNames}
\alias{probeNames,SigSet-method}
\title{Get Probe Names of SigSet class}
\usage{
probeNames(x)

\S4method{probeNames}{SigSet}(x)
}
\arguments{
\item{x}{object of \code{Sigset}}
}
\value{
A char vector
}
\description{
Get Probe Names of SigSet class
}
\examples{
sset <- sesameDataGet('HM450.1.TCGA.PAAD')$sset
head(probeNames(sset))
}
