% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.R
\name{probeSignals}
\alias{probeSignals}
\title{Compute probe signals}
\usage{
probeSignals(sset.target, ssets.normal)
}
\arguments{
\item{sset.target}{\code{SignalSet} for target}

\item{ssets.normal}{\code{SignalSet}s from normal samples}
}
\value{
probe.signals
}
\description{
Compute probe signals
}
