% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigSetMethods.R
\docType{methods}
\name{pval}
\alias{pval}
\alias{pval,SigSet-method}
\title{pval getter generic}
\usage{
pval(x)

\S4method{pval}{SigSet}(x)
}
\arguments{
\item{x}{object of \code{SigSet}}
}
\value{
The pval slot of \code{SigSet}
}
\description{
pval getter generic

Get pval slot of SigSet class
}
\examples{

sesameDataCache("HM450") # if not done yet
sset <- sesameDataGet('HM450.1.TCGA.PAAD')$sset
head(pval(sset))
}
