% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigSetMethods.R
\docType{methods}
\name{pval<-}
\alias{pval<-}
\alias{pval<-,SigSet-method}
\title{pval replacement generic}
\usage{
pval(x) <- value

\S4method{pval}{SigSet}(x) <- value
}
\arguments{
\item{x}{object of \code{SigSet}}

\item{value}{new value}
}
\value{
a new \code{SigSet}
}
\description{
pval replacement generic

Replace pval slot of SigSet class
}
\examples{
sset <- sesameDataGet('HM450.1.TCGA.PAAD')$sset
df <- pval(sset)
df[1] <- 0.01
pval(sset) <- df
}
