% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{readIDATs}
\alias{readIDATs}
\title{Import IDATs from a list of samples}
\usage{
readIDATs(sample.names, base.dir = NULL, mc = FALSE, mc.cores = 8)
}
\arguments{
\item{sample.names}{a sample list}

\item{base.dir}{base directory}

\item{mc}{use multiple cores}

\item{mc.cores}{number of cores to use}
}
\value{
a list of \code{SignalSet}s
}
\description{
Each element of the returned list contains a matrix
having signal intensity addressed by chip address
}

