% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{readIDATsFromDir}
\alias{readIDATsFromDir}
\title{Import IDATs from a directory}
\usage{
readIDATsFromDir(dir.name, ...)
}
\arguments{
\item{dir.name}{directory name.}

\item{...}{multiple core parameters: mc and mc.cores see \code{readIDATs}}
}
\value{
a list of \code{SignalSet}s
}
\description{
Each element of the returned list contains a matrix
having signal intensity addressed by chip address
}
\details{
dir.name is the folder containing the IDAT files.
}
\examples{
ssets <- readIDATsFromDir(
  system.file("extdata", "", package = "sesameData"))
}
