% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{readIDATsFromSampleSheet}
\alias{readIDATsFromSampleSheet}
\title{Import IDATs from a sample sheet}
\usage{
readIDATsFromSampleSheet(sample.sheet, base.dir = NULL, ...)
}
\arguments{
\item{sample.sheet}{path to sample sheet}

\item{base.dir}{directory on which the \code{sample.sheet.path} is based}

\item{...}{multiple core parameters: mc and mc.cores see \code{readIDATs}}
}
\value{
a list of \code{SignalSet}s
}
\description{
Each element of the returned list contains a matrix
having signal intensity addressed by chip address
}

