% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{recommendedMaskNames}
\alias{recommendedMaskNames}
\title{Recommended mask names for each Infinium platform}
\usage{
recommendedMaskNames(platform)
}
\arguments{
\item{platform}{string for platform, e.g., EPIC, EPICv2, MSA}
}
\value{
a named list of mask names
}
\description{
The returned name is the db name used in KYCG.mask
}
\examples{
recommendedMaskNames("EPICv2")
recommendedMaskNames("EPIC")

}
