% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{resetMask}
\alias{resetMask}
\title{Reset Masking}
\usage{
resetMask(sdf)
}
\arguments{
\item{sdf}{a \code{SigDF}}
}
\value{
a new \code{SigDF} with mask reset to all FALSE
}
\description{
Reset Masking
}
\examples{
sesameDataCache("EPIC") # if not done yet
sdf <- sesameDataGet('EPIC.1.LNCaP')$sdf
sum(mask(sdf))
sdf <- addMask(sdf, c("cg14057072", "cg22344912"))
sum(mask(sdf))
sum(mask(resetMask(sdf)))
}
