% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{restoreMask}
\alias{restoreMask}
\title{Save current mask}
\usage{
restoreMask(sset, from = "mask2")
}
\arguments{
\item{sset}{a \code{SigSet} object}

\item{from}{name of a previously saved mask}
}
\value{
a new \code{SigSet} object
}
\description{
Save current mask
}
\examples{
sset <- sesameDataGet('EPIC.1.LNCaP')$sset
sset <- resetMask(sset)
sset <- saveMask(sset)
sset <- restoreMask(sset)
}
