% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesamize.R
\name{RGChannelSetToSigDFs}
\alias{RGChannelSetToSigDFs}
\title{Convert RGChannelSet (minfi) to a list of SigDF (SeSAMe)}
\usage{
RGChannelSetToSigDFs(rgSet, manifest = NULL, BPPARAM = SerialParam())
}
\arguments{
\item{rgSet}{a minfi::RGChannelSet}

\item{manifest}{manifest file}

\item{BPPARAM}{get parallel with MulticoreParam(n)}
}
\value{
a list of sesame::SigDF
}
\description{
Notice the colData() and rowData() is lost. Most cases, rowData is empty
anyway.
}
\examples{

if (require(FlowSorted.Blood.450k)) {
    rgSet <- FlowSorted.Blood.450k[,1:2]
    sdfs <- RGChannelSetToSigDFs(rgSet)
}
}
