% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesamize.R
\name{RGChannelSetToSigSet}
\alias{RGChannelSetToSigSet}
\title{Convert RGChannelSet (minfi) to SigSet (SeSAMe)}
\usage{
RGChannelSetToSigSet(rgSet, BPPARAM = SerialParam())
}
\arguments{
\item{rgSet}{a minfi::RGChannelSet}

\item{BPPARAM}{get parallel with MulticoreParam(2)}
}
\value{
a list of sesame::SigSet
}
\description{
Notice the colData() and rowData() is lost. Most cases, rowData is empty
anyway.
}
\examples{

if (require(FlowSorted.Blood.450k)) {
    rgSet <- FlowSorted.Blood.450k[,1:2]
    ssets <- RGChannelSetToSigSet(rgSet)
}
}
