% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{saveMask}
\alias{saveMask}
\title{Save current mask}
\usage{
saveMask(sset, to = "mask2")
}
\arguments{
\item{sset}{a \code{SigSet} object}

\item{to}{new mask name}
}
\value{
a new \code{SigSet} object
}
\description{
Save current mask
}
\examples{
sesameDataCache("EPIC") # if not done yet
sset <- sesameDataGet('EPIC.1.LNCaP')$sset
sset <- resetMask(sset)
sset <- saveMask(sset)
}
