% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgroup.R
\name{ScaleGroup}
\alias{ScaleGroup}
\title{Scale group}
\usage{
ScaleGroup(group.obj)
}
\arguments{
\item{group.obj}{group object that needs to be scaled}
}
\value{
scaled group obj
}
\description{
Scale group to incorporate text on margins
}
