% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{SDFcollapseToPfx}
\alias{SDFcollapseToPfx}
\title{collapse to probe prefix}
\usage{
SDFcollapseToPfx(sdf)
}
\arguments{
\item{sdf}{a SigDF object}
}
\value{
a data frame with updated Probe_ID
}
\description{
collapse to probe prefix
}
