% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigDFMethods.R
\name{sdfPlatform}
\alias{sdfPlatform}
\title{Convenience function to output platform attribute of SigDF}
\usage{
sdfPlatform(sdf, silent = FALSE)
}
\arguments{
\item{sdf}{a SigDF object}

\item{silent}{whether to print message}
}
\value{
the platform string for the SigDF object
}
\description{
Convenience function to output platform attribute of SigDF
}
\examples{
sesameDataCache()
sdf <- sesameDataGet('EPIC.1.SigDF')
sdfPlatform(sdf)

}
