% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sesame_checkVersion}
\alias{sesame_checkVersion}
\title{Check SeSAMe versions}
\usage{
sesame_checkVersion()
}
\value{
print the version of sesame, sesameData, biocondcutor and R
}
\description{
print package verison of sesame and depended packages to help troubleshoot
installation issues.
}
\examples{
sesame_checkVersion()
}
