% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesameAnno.R
\name{sesameAnno_buildAddressFile}
\alias{sesameAnno_buildAddressFile}
\title{Build sesame ordering address file from tsv}
\usage{
sesameAnno_buildAddressFile(tsv)
}
\arguments{
\item{tsv}{a platform name, a file path or a tibble/data.frame manifest file}
}
\value{
a list of ordering and controls

## addr <- sesameAnno_buildAddressFile("HM450")
## or
## download tsv directly from
## http://zwdzwd.github.io/InfiniumAnnotation or
## addr <- sesameAnno_buildAddressFile("downloaded_file")
}
\description{
Build sesame ordering address file from tsv
}
