% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesameAnno.R
\name{sesameAnno_buildManifestGRanges}
\alias{sesameAnno_buildManifestGRanges}
\title{Build manifest GRanges from tsv}
\usage{
sesameAnno_buildManifestGRanges(tsv, genome, decoy = FALSE, columns = NULL)
}
\arguments{
\item{tsv}{a file path, a platform (e.g., EPIC), or
a tibble/data.frame object}

\item{genome}{a genome string, e.g., hg38, mm10}

\item{decoy}{consider decoy sequence in chromosome order}

\item{columns}{the columns to include in the GRanges}
}
\value{
GRanges

## genome <- sesameAnno_buildManifestGRanges("HM450", "hg38")
## or
## download tsv directly from
## http://zwdzwd.github.io/InfiniumAnnotation or
## genome <- sesameAnno_buildAddressFile("downloaded_file")
}
\description{
Build manifest GRanges from tsv
}
