% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesameAnno.R
\name{sesameAnno_download}
\alias{sesameAnno_download}
\title{Download additional annotation files}
\usage{
sesameAnno_download(title, dest_dir, version = 1)
}
\arguments{
\item{title}{title of the annotation file}

\item{dest_dir}{download to this directory}

\item{version}{version number}
}
\value{
annotation file
}
\description{
see also
http://zwdzwd.github.io/InfiniumAnnotation
}
\details{
This function acts similarly as sesameAnno_get except that it directly
download files without invoking BiocFileCache. This is needed in some
situation because BiocFileCache may change the file name and downstream
program may depend on the correct file names. It also lets you download
files in a cleaner way without routing through BiocFileCache
}
\examples{

## avoid testing as this function uses external host
if (FALSE) {
sesameAnno_download("Test/3999492009_R01C01_Grn.idat", tempdir())
}

}
