% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesameAnno.R
\name{sesameAnno_getManifestDF}
\alias{sesameAnno_getManifestDF}
\title{download Infinium manifest from Github repositories}
\usage{
sesameAnno_getManifestDF(
  platform,
  genome = NULL,
  version = anno_base_default_version
)
}
\arguments{
\item{platform}{Mammal40, MM285, EPIC, and HM450}

\item{genome}{hg38, mm10 etc.}

\item{version}{release version, default is the latest}
}
\value{
tibble
}
\description{
download Infinium manifest from Github repositories
}
\examples{

## avoid testing since it depends on external host
if (FALSE) {
mft <- sesameAnno_getManifestDF("Mammal40")
}
}
