% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesameAnno.R
\name{sesameAnno_readManifestTSV}
\alias{sesameAnno_readManifestTSV}
\title{Read manifest file to a tsv format}
\usage{
sesameAnno_readManifestTSV(tsv_fn)
}
\arguments{
\item{tsv_fn}{tsv file path}
}
\value{
a manifest as a tibble
}
\description{
Read manifest file to a tsv format
}
\examples{
\dontrun{
## download manifest from
## http://zwdzwd.github.io/InfiniumAnnotation
mft <- sesameAnno_readManifestTSV("downloaded_file")
}
}
