% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesameAnno.R
\name{sesameData_getAnno}
\alias{sesameData_getAnno}
\title{retrieve additional annotation files}
\usage{
sesameData_getAnno(title, version = 1, dest_dir = NULL)
}
\arguments{
\item{title}{title of the annotation file}

\item{version}{version number}

\item{dest_dir}{if not NULL, download to this directory}
}
\value{
annotation file
}
\description{
retrieve additional annotation files
}
\examples{
cat("Deprecated!")

}
