% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{sesamePlotRedGrnQQ}
\alias{sesamePlotRedGrnQQ}
\title{Plot red-green QQ-Plot using Infinium-I Probes}
\usage{
sesamePlotRedGrnQQ(sset)
}
\arguments{
\item{sset}{a \code{SigSet}}
}
\value{
create a qqplot
}
\description{
Plot red-green QQ-Plot using Infinium-I Probes
}
\examples{
sesameDataCache("EPIC")
sset <- # if not done yet
sset <- sesameDataGet('EPIC.1.LNCaP')$sset
sesamePlotRedGrnQQ(sset)
}
