% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\docType{class}
\name{sesameQC-class}
\alias{sesameQC-class}
\title{An S4 class to hold QC statistics}
\value{
sesameQC object
}
\description{
An S4 class to hold QC statistics
}
\section{Slots}{

\describe{
\item{\code{stat}}{a list to store qc stats}
}}

