% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{sesameQC}
\alias{sesameQC}
\title{Generate summary numbers that indicative of experiment quality
Please provide a raw SigDF(before any preprocessing). Usually
directly from readIDATpair}
\usage{
sesameQC(sdf)
}
\arguments{
\item{sdf}{a \code{SigDF} object}
}
\value{
a sesameQC class object
}
\description{
Generate summary numbers that indicative of experiment quality
Please provide a raw SigDF(before any preprocessing). Usually
directly from readIDATpair
}
\examples{
sesameDataCache("EPIC") # if not done yet
sdf <- sesameDataGet('EPIC.1.SigDF')
sesameQC(sdf)

# release memory for Windows package builder
rm(list=ls(env=sesameData:::cacheEnv), envir=sesameData:::cacheEnv)
gc()

}
