% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{sesameQC_asDataFrame}
\alias{sesameQC_asDataFrame}
\title{Convert data frame to sesameQC object}
\usage{
sesameQC_asDataFrame(df)
}
\arguments{
\item{df}{a publicQC data frame}
}
\value{
a list sesameQC objects
}
\description{
The function convert a data frame back to a list of sesameQC objects
}
\examples{
df <- sesameDataGet("MM285.publicQC")
qcs <- sesameQC_asDataFrame(df[1:2,])
}
