% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{sesameQC_calcStats}
\alias{sesameQC_calcStats}
\title{Calculate QC stats on multiple SigDFs}
\usage{
sesameQC_calcStats(sdfs, funs = NULL)
}
\arguments{
\item{sdfs}{a list of SigDFs}

\item{funs}{a sesameQC_calcStats_* function or a list of them
default to sesameQC_calcStats_detection}
}
\value{
a data frame
}
\description{
Calculate QC stats on multiple SigDFs
}
\examples{
sdfs <- sesameDataGet("EPIC.5.SigDF.normal")
sesameQC_calcStats(sdfs)
sesameQC_calcStats(sdfs, sesameQC_calcStats_numProbes)
sesameQC_calcStats(sdfs,
    c(sesameQC_calcStats_numProbes, sesameQC_calcStats_channel))
}
