% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{sesameQC_calcStats_dyeBias}
\alias{sesameQC_calcStats_dyeBias}
\title{Quantify deviation of dye bias in the high signal range from the
global median}
\usage{
sesameQC_calcStats_dyeBias(sdf, qc = NULL)
}
\arguments{
\item{sdf}{a \code{SigDF}}

\item{qc}{existing sesameQC object to add to (optional)}
}
\value{
a sesameQC
}
\description{
Positive value indicates augmentation of high-end dye bias over
low-end. negative value represents high-end dye bias contradicts
that at low-end (a distorted dye bias). Negative distortion score
(< -1) suggests low experiment quality. 0 suggests a consistent
dye bias at high and low-end.
}
\examples{
sdf <- sesameDataGet('EPIC.1.SigDF')
sesameQC_calcStats_dyeBias(sdf)
}
