% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{sesameQC_calcStats_intens}
\alias{sesameQC_calcStats_intens}
\title{Generate summary numbers that indicative of experiment quality
based on intensity.}
\usage{
sesameQC_calcStats_intens(sdf)
}
\arguments{
\item{sdf}{a \code{SigDF} object}
}
\value{
a sesameQC
}
\description{
Please provide a raw SigDF(before any preprocessing). Usually
directly from readIDATpair.
}
\examples{
sesameDataCache("EPIC") # if not done yet
sdf <- sesameDataGet('EPIC.1.SigDF')
sesameQC_calcStats_intens(sdf)
}
