% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{sesameQC_plotBar}
\alias{sesameQC_plotBar}
\title{Bar plot of color channel switch}
\usage{
sesameQC_plotBar(qcs)
}
\arguments{
\item{sdfs}{a list of SigDFs}
}
\value{
a bar plot comparing color channel switches
}
\description{
Bar plot of color channel switch
}
\examples{
sesameDataCache("EPIC") # if not done yet
sdfs <- sesameDataGet("EPIC.5.SigDF.normal")
sesameQC_plotBar(lapply(sdfs, sesameQC_calcStats_detection))
}
