% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{sesameQC_plotBetaByDesign}
\alias{sesameQC_plotBetaByDesign}
\title{Plot betas distinguishing different Infinium chemistries}
\usage{
sesameQC_plotBetaByDesign(sdf, legend_pos = "top", main = "", ...)
}
\arguments{
\item{sdf}{SigDF}

\item{...}{additional options to plot}
}
\value{
create a density plot
}
\description{
Plot betas distinguishing different Infinium chemistries
}
\examples{
sdf <- sesameDataGet("EPIC.1.SigDF")
sesameQC_plotBetaByDesign(sdf)
}
