% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{sesameQC_plotHeatSNPs}
\alias{sesameQC_plotHeatSNPs}
\title{Plot SNP heatmap}
\usage{
sesameQC_plotHeatSNPs(sdfs, cluster = TRUE, filter.nonvariant = TRUE)
}
\arguments{
\item{cluster}{show clustered heatmap}

\item{filter.nonvariant}{whether to filter nonvariant (range < 0.3)}

\item{betas}{beta value matrix, row: probes; column: samples}
}
\description{
Plot SNP heatmap
}
\examples{

sdfs <- sesameDataGet("EPIC.5.SigDF.normal")
sesameQC_plotHeatSNPs(betas)
}
