% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{sesameQC_plotRedGrnQQ}
\alias{sesameQC_plotRedGrnQQ}
\title{Plot red-green QQ-Plot using Infinium-I Probes}
\usage{
sesameQC_plotRedGrnQQ(sdf)
}
\arguments{
\item{sdf}{a \code{SigDF}}
}
\value{
create a qqplot
}
\description{
Plot red-green QQ-Plot using Infinium-I Probes
}
\examples{
sesameDataCache("EPIC") # if not done yet
sdf <- sesameDataGet('EPIC.1.SigDF')
sesameQC_plotRedGrnQQ(sdf)
}
