% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{sesameQC_publicQC}
\alias{sesameQC_publicQC}
\title{Retrieve stats of public data of similar nature
e.g., tissue, FFPE vs non-FFPE, etc.}
\usage{
sesameQC_publicQC(platform = NULL, tissue = NULL, samplePrep = NULL)
}
\arguments{
\item{platform}{HM450, EPIC, MM285 etc.}

\item{tissue}{optional, blood, buccal, saliva, etc.}

\item{samplePrep}{optional, fresh, FF, etc.}
}
\value{
a data frame of public data QC stats
}
\description{
Retrieve stats of public data of similar nature
e.g., tissue, FFPE vs non-FFPE, etc.
}
\examples{
publicQC <- sesameQC_publicQC()

}
