% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{sesameQC_rankDetectionSuccess}
\alias{sesameQC_rankDetectionSuccess}
\title{This function compares the input sample with public data
in terms of detection success rate.}
\usage{
sesameQC_rankDetectionSuccess(sdf, publicQC = NULL)
}
\arguments{
\item{sdf}{a raw (unprocessed) \code{SigDF}}

\item{publicQC}{output of sesameQC_publicQC, optional}
}
\value{
a fraction to represent the rank of the test stat
}
\description{
This function compares the input sample with public data
in terms of detection success rate.
}
\examples{

sesameDataCache("EPIC") # if not done yet
sdf <- sesameDataGet('EPIC.1.SigDF')
ranks <- sesameQC_rankDetectionSuccess(sdf)

}
