% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{sesameQC_rankStats}
\alias{sesameQC_rankStats}
\title{This function compares the input sample with public data.
Only overlapping metrics will be compared.}
\usage{
sesameQC_rankStats(qc, publicQC = NULL)
}
\arguments{
\item{qc}{a sesameQC object}

\item{publicQC}{output of sesameQC_publicQC, optional}
}
\value{
a sesameQC
}
\description{
This function compares the input sample with public data.
Only overlapping metrics will be compared.
}
\examples{

sesameDataCache("EPIC") # if not done yet
sdf <- sesameDataGet('EPIC.1.SigDF')
sesameQC_rankStats(sesameQC_calcStats(sdf, "intensity"))

}
