% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{sesameQCtoDF}
\alias{sesameQCtoDF}
\title{Convert a list of sesameQC to data frame}
\usage{
sesameQCtoDF(qcs, cols = c("frac_dt_cg", "RGdistort", "RGratio"))
}
\arguments{
\item{qcs}{sesameQCs}

\item{cols}{QC columns, use NULL to report all}
}
\value{
a data frame
}
\description{
Convert a list of sesameQC to data frame
}
\examples{
sdf <- sesameDataGet("EPIC.1.SigDF")
qcs <- sesameQC_calcStats(sdf, "detection")
sesameQCtoDF(qcs)
}
