% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{setMask}
\alias{setMask}
\title{Set mask to only the probes specified}
\usage{
setMask(sset, probes)
}
\arguments{
\item{sset}{a \code{SigSet}}

\item{probes}{a vector of probe IDs or a logical vector with TRUE
representing masked probes}
}
\description{
Set mask to only the probes specified
}
\examples{
sset <- sesameDataGet('EPIC.1.LNCaP')$sset
sum(mask(sset))
sum(mask(setMask(sset, "cg14959801")))
sum(mask(setMask(sset, c("cg14057072", "cg22344912"))))
}
