% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sex.inference.model}
\alias{sex.inference.model}
\title{sex inference model}
\format{An object of class \code{randomForest} of length 18.}
\usage{
sex.inference.model
}
\description{
sex inference model
}
\keyword{datasets}

