% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{show,sesameQC-method}
\alias{show,sesameQC-method}
\title{The display method for sesameQC}
\usage{
\S4method{show}{sesameQC}(object)
}
\arguments{
\item{object}{object to be displayed}
}
\value{
None
}
\description{
The function outputs the number of probes in each category and the first
few signal measurements.
}
\examples{
new("sesameQC")
}
