% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{SigDF}
\alias{SigDF}
\title{SigDF constructor}
\usage{
SigDF(df, platform = "EPIC", ctl = NULL)
}
\arguments{
\item{df}{a \code{data.frame}}

\item{platform}{a string to specify the array platform}
}
\description{
SigDF constructor
}
\examples{
sesameDataCache("EPIC") # if not done yet
df <- as.data.frame(sesameDataGet('EPIC.1.SigDF'))
}
