% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesamize.R
\name{SigDFsToRGChannelSet}
\alias{SigDFsToRGChannelSet}
\title{Convert sesame::SigDF to minfi::RGChannelSet}
\usage{
SigDFsToRGChannelSet(sdfs, BPPARAM = SerialParam(), annotation = NA)
}
\arguments{
\item{sdfs}{a list of sesame::SigDF}

\item{BPPARAM}{get parallel with MulticoreParam(n)}

\item{annotation}{the minfi annotation string, guessed if not given}
}
\value{
a minfi::RGChannelSet
}
\description{
This function does not support the mouse array.
}
\examples{

sesameDataCache("EPIC") # if not done yet
sdf <- sesameDataGet('EPIC.1.SigDF')
rgSet <- SigDFsToRGChannelSet(sdf)

}
