% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesamize.R
\name{SigDFToRatioSet}
\alias{SigDFToRatioSet}
\title{Convert one sesame::SigDF to minfi::RatioSet}
\usage{
SigDFToRatioSet(sdf, annotation = NA)
}
\arguments{
\item{sdf}{a sesame::SigDF}

\item{annotation}{minfi annotation string}
}
\value{
a minfi::RatioSet
}
\description{
Convert one sesame::SigDF to minfi::RatioSet
}
\examples{

sesameDataCache("EPIC") # if not done yet
sdf <- sesameDataGet('EPIC.1.LNCaP')$sdf
ratioSet <- SigDFToRatioSet(sdf)

}
