% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{signalMU}
\alias{signalMU}
\title{report M and U for regular probes}
\usage{
signalMU(sset)
}
\arguments{
\item{sset}{a \code{SigSet}}
}
\value{
a data frame of M and U columns
}
\description{
report M and U for regular probes
}
\examples{
sesameDataCache("EPIC") # if not done yet
sset <- sesameDataGet('EPIC.1.LNCaP')$sset
signalMU(sset)
}
