% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{signalR6toS4}
\alias{signalR6toS4}
\title{sesame R6 to S4}
\usage{
signalR6toS4(sset)
}
\arguments{
\item{sset}{signalset in R6}
}
\value{
signalset in S4
}
\description{
SignalSet-SigSet conversion
}
\examples{

sset <- SignalSet$new('EPIC')
sset$IG <- matrix(1:4, nrow=2)
sset$IR <- matrix(1:4, nrow=2)
sset$II <- matrix(1:4, nrow=2)
sset$oobG <- matrix(1:4, nrow=2)
sset$oobR <- matrix(1:4, nrow=2)
sset$ctl <- data.frame(G=1:2,R=3:4)
sset$pval <- rep(0,2)

signalR6toS4(sset)
}
