% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\docType{class}
\name{SignalSet}
\alias{SignalSet}
\title{SignalSet class}
\format{An \code{\link{R6Class}} object.}
\usage{
SignalSet
}
\value{
Object of class \code{SignalSet}
}
\description{
SignalSet class
}
\section{Fields}{

\describe{
\item{\code{platform}}{platform name, currently supports "EPIC", "hm450" and "hm27"}

\item{\code{IG}}{intensity table for type I probes in green channel}

\item{\code{IR}}{intensity table for type I probes in red channel}

\item{\code{II}}{intensity table for type II probes}

\item{\code{oobG}}{out-of-band probes in green channel}

\item{\code{oobR}}{out-of-band probes in red channel}

\item{\code{ctl}}{all the control probe intensities}

\item{\code{pval}}{named numeric vector of p-values}

\item{\code{mask}}{probe mask
\describe{
  \item{Documentation}{For full documentation of each method go to }
  \item{\code{new(platform)}}{Create a SignalSet in the specified platform}
  \item{\code{detectPValue()}}{Detect P-value for each probe}
  \item{\code{toM()}}{Convert to M values}
  \item{\code{totalIntensities()}}{Total intensity on each probe}
}}
}}

\examples{
SignalSet$new("EPIC")
}
\keyword{datasets}
