% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{signalToBeta}
\alias{signalToBeta}
\title{Convert multiple signals to beta}
\usage{
signalToBeta(ssets, na.mask = TRUE, mc = FALSE, mc.cores = NULL)
}
\arguments{
\item{ssets}{a list of \code{SignalSet}s}

\item{na.mask}{NA-mask repetitive and SNP probes}

\item{mc}{parallel}

\item{mc.cores}{number of cores}
}
\value{
beta value matrix
}
\description{
Convert multiple signals to beta
}

