% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigSetList.R
\docType{methods}
\name{SigSetList-methods}
\alias{SigSetList-methods}
\alias{show,SigSetList-method}
\title{SigSetList methods (centralized).
 
`show`         print a summary of the SigSetList.}
\usage{
\S4method{show}{SigSetList}(object)
}
\arguments{
\item{object}{a SigSetList}

\item{x}{a SigSetList}

\item{sset}{a SigSetList}
}
\description{
SigSetList methods (centralized).
 
`show`         print a summary of the SigSetList.
}
