% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigSetList.R
\name{SigSetList}
\alias{SigSetList}
\title{constructor}
\usage{
SigSetList(...)
}
\arguments{
\item{...}{the SigSet objects that will be the List elements}
}
\value{
a SigSetList
}
\description{
constructor
}
