% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigSetList.R
\name{SigSetListFromIDATs}
\alias{SigSetListFromIDATs}
\title{read IDATs into a SigSetList}
\usage{
SigSetListFromIDATs(stubs, parallel = FALSE)
}
\arguments{
\item{stubs}{the IDAT filename stubs}

\item{parallel}{run in parallel? (default FALSE)}
}
\value{
a SigSetList
}
\description{
FIXME: switch from `parallel` to BiocParallel
}
