% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesamize.R
\name{SigSetsToRGChannelSet}
\alias{SigSetsToRGChannelSet}
\title{Convert sesame::SigSet to minfi::RGChannelSet}
\usage{
SigSetsToRGChannelSet(ssets, BPPARAM = SerialParam(), annotation = NA)
}
\arguments{
\item{ssets}{a list of sesame::SigSet}

\item{BPPARAM}{get parallel with MulticoreParam(n)}

\item{annotation}{the minfi annotation string, guessed if not given}
}
\value{
a minfi::RGChannelSet
}
\description{
Convert sesame::SigSet to minfi::RGChannelSet
}
\examples{

sesameDataCache("EPIC") # if not done yet
sset <- sesameDataGet('EPIC.1.LNCaP')$sset
rgSet <- SigSetsToRGChannelSet(sset)

}
