% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesamize.R
\name{SigSetToRatioSet}
\alias{SigSetToRatioSet}
\title{Convert one sesame::SigSet to minfi::RatioSet}
\usage{
SigSetToRatioSet(sset, annotation = NA)
}
\arguments{
\item{sset}{a sesame::SigSet}

\item{annotation}{minfi annotation string}
}
\value{
a minfi::RatioSet
}
\description{
Convert one sesame::SigSet to minfi::RatioSet
}
\examples{

sset <- sesameDataGet('EPIC.1.LNCaP')$sset
ratioSet <- SigSetToRatioSet(sset)

}
