% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kyCG.R
\name{skew}
\alias{skew}
\title{skew determines the skew of a distribution x, taken from the Moments package}
\usage{
skew(x, na.rm = FALSE)
}
\arguments{
\item{x}{Vector of numeric values}

\item{na.rm}{Logical value corresponding to whether NA will be ignored}
}
\value{
Numeric value quantifying the skew of the distribution x
}
\description{
skew determines the skew of a distribution x, taken from the Moments package
}
