% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{[.WGroup}
\alias{[.WGroup}
\title{subset a SignalSet}
\usage{
\method{[}{WGroup}(sset, probes)
}
\arguments{
\item{sset}{a \code{SignalSet}}

\item{probes}{probe names}
}
\value{
a \code{SignalSet} with only probes
}
\description{
subset a SignalSet
}

