% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{summaryExtractCfList}
\alias{summaryExtractCfList}
\title{Extract slope information from DMLSummary}
\usage{
summaryExtractCfList(smry)
}
\arguments{
\item{smry}{DMLSummary from DML command}
}
\value{
a list of coefficients for each tested factor
}
\description{
Extract slope information from DMLSummary
}
\examples{
sesameDataCache("HM450") # in case not done yet
data <- sesameDataGet('HM450.76.TCGA.matched')
smry <- DML(data$betas[1:1000,], ~type, meta=data$sampleInfo)
cf_list <- summaryExtractCfList(smry)
}
