% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{summaryExtractTest}
\alias{summaryExtractTest}
\title{Extract slope information from DMLSummary}
\usage{
summaryExtractTest(smry)
}
\arguments{
\item{smry}{DMLSummary from DML command}
}
\value{
a table of slope and p-value
}
\description{
Extract slope information from DMLSummary
}
\examples{
sesameDataCache("HM450") # in case not done yet
data = sesameDataGet('HM450.76.TCGA.matched')
smry = DML(data$betas[1:1000,], ~type, meta=data$sampleInfo)
slopes = summaryExtractTest(smry)

sesameDataClearCache()
}
