% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kyCG.R
\name{testEnrichment1}
\alias{testEnrichment1}
\title{testEnrichment1 tests for the enrichment of set of probes (query set) in a
single given feature (database set)}
\usage{
testEnrichment1(querySet, databaseSet, universeSet, estimate.type = "ES")
}
\arguments{
\item{querySet}{Vector of probes of interest (e.g., significant probes)}

\item{databaseSet}{Vector corresponding to the database sets of
interest with associated meta data as an attribute to each element.}

\item{universeSet}{Vector of probes in the universe set containing all of
the probes to be considered in the test.}

\item{estimate.type}{String indicating the estimate to report. (Default:
"ES")}
}
\value{
One list containing features corresponding the test estimate,
p-value, and type of test.
}
\description{
testEnrichment1 tests for the enrichment of set of probes (query set) in a
single given feature (database set)
}
