% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kyCG.R
\name{testEnrichmentFisher}
\alias{testEnrichmentFisher}
\title{testEnrichmentFisher uses Fisher's exact test to estimate the association
between two categorical variables.}
\usage{
testEnrichmentFisher(query, database, universe)
}
\arguments{
\item{query}{Vector of probes of interest (e.g., significant probes)}

\item{database}{Vectors corresponding to the database set of
interest with associated meta data as an attribute to each element.}

\item{universe}{Vector of probes in the universe set containing all of
the probes to be considered in the test. (Default: NULL)}
}
\value{
A DataFrame with the estimate/statistic, p-value, and name of test
for the given results.
}
\description{
testEnrichmentFisher uses Fisher's exact test to estimate the association
between two categorical variables.
}
