% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KYCG.R
\name{testEnrichmentGSEA}
\alias{testEnrichmentGSEA}
\title{testEnrichmentGSEA uses the GSEA test to estimate the association of a
categorical variable against a continuous variable.}
\usage{
testEnrichmentGSEA(query, database, precise = FALSE)
}
\arguments{
\item{query}{Vector of probes of interest (e.g., significant probes)}

\item{database}{Vector of probes corresponding to a single database set}

\item{precise}{whether to compute precise p-value (up to numerical limit)
of interest.}
}
\value{
A DataFrame with the estimate/statistic, p-value, and name of test
for the given results.
}
\description{
estimate represent enrichment score and negative estimate indicate a
test for depletion
}
