% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KYCG.R
\name{testEnrichmentSpearman}
\alias{testEnrichmentSpearman}
\title{testEnrichmentSpearman uses the Spearman statistical test to estimate
the association between two continuous variables.}
\usage{
testEnrichmentSpearman(query, database)
}
\arguments{
\item{query}{Vector of probes of interest (e.g., significant probes)}

\item{database}{List of vectors corresponding to the database set of
interest with associated meta data as an attribute to each element.}
}
\value{
A DataFrame with the estimate/statistic, p-value, and name
of test for the given results.
}
\description{
testEnrichmentSpearman uses the Spearman statistical test to estimate
the association between two continuous variables.
}
