% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgroup.R
\name{ToAffine}
\alias{ToAffine}
\title{Convert from absolute coordinates to affine coordinates}
\usage{
ToAffine(dm, dm.sys)
}
\arguments{
\item{dm}{dimension on the same coordinate system
as the affine system (absolute coordinates)}

\item{dm.sys}{dimension of the affine system}
}
\value{
dimension on affine coordinates (relative coordinates)
}
\description{
Convert from absolute coordinates to affine coordinates
}
