% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{topSegments}
\alias{topSegments}
\title{top segments in differential methylation}
\usage{
topSegments(cf1)
}
\arguments{
\item{cf1}{coefficient table of one factor from segmentDMR}
}
\value{
coefficient table ordered by adjusted p-value of segments
}
\description{
This is a convenience function to show top differential methylated segments.
}
