% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{totalIntensityZscore}
\alias{totalIntensityZscore}
\title{Calculate intensity Z-score}
\usage{
totalIntensityZscore(sset)
}
\arguments{
\item{sset}{a \code{SigSet}}
}
\value{
a vector of Z-score for each probe
}
\description{
This function compute intensity Z-score with respect to the mean.
Log10 transformation is done first. Probes of each design type are
grouped before Z-scores are computed.
}
\examples{
sset <- sesameDataGet('HM450.1.TCGA.PAAD')$sset
head(totalIntensityZscore(sset))
}
