% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{updateSigBySpecies}
\alias{updateSigBySpecies}
\title{Set color and mask using species-specific manifest}
\usage{
updateSigBySpecies(sdf, species, addr = NULL)
}
\arguments{
\item{sdf}{a \code{SigDF}}

\item{species}{the species the sample is considered to be}

\item{addr}{species-specific address species, optional}
}
\value{
a \code{SigDF} with updated color channel and mask
}
\description{
Set color and mask using species-specific manifest
}
\examples{
sdf <- sesameDataGet('Mammal40.1.SigDF')
sdf_mouse <- updateSigBySpecies(sdf, "mus_musculus")

}
