% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.R
\name{visualizeSegments}
\alias{visualizeSegments}
\title{Visualize segments}
\usage{
visualizeSegments(seg, to.plot = NULL)
}
\arguments{
\item{seg}{segments}

\item{to.plot}{chromosome to plot (by default plot all chromosomes)}
}
\description{
require ggplot2, scales
}

