% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{WColumnBind}
\alias{WColumnBind}
\title{column bind non-overlapping objects}
\usage{
WColumnBind(..., nr = NULL, nc = NULL)
}
\arguments{
\item{...}{plotting objects}

\item{nr}{number of rows}

\item{nc}{number of columns}
}
\value{
an object of class WDim
}
\description{
column bind non-overlapping objects
}
