% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customize.R
\name{WCustomize}
\alias{WCustomize}
\title{Customize an existing plot}
\usage{
WCustomize(mar.left = NULL, mar.right = NULL, mar.top = NULL,
  mar.bottom = NULL, mar = NULL)
}
\arguments{
\item{mar.left}{left margin [0.03]}

\item{mar.right}{right margin [0.03]}

\item{mar.top}{top margin [0.03]}

\item{mar.bottom}{bottom margin [0.03]}

\item{mar}{margin in all directions [0.03]}
}
\value{
an object of class WCustomize
}
\description{
Customize an existing plot
}
\examples{
WHeatmap(matrix(c('fred','frank','brad',
                'frank','fred','frank'), ncol=2)) +
  WLegendV(NULL, RightOf(), label.fontsize = 20) +
  WCustomize(mar.right=0.1)
}
