% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogram.R
\name{WDendrogram}
\alias{WDendrogram}
\title{WDendrogram class}
\usage{
WDendrogram(clust = NULL, dm = WDim(0, 0, 1, 1), name = "",
  facing = c("bottom", "top", "left", "right"))
}
\arguments{
\item{clust}{hclust object}

\item{dm}{plotting dimension}

\item{name}{name of the dendrogram plot}

\item{facing}{direction of the dendrogram plot}
}
\value{
an object of class WDendrogram
}
\description{
WDendrogram class
}
