% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dims.R
\name{WDim}
\alias{WDim}
\title{class WDim}
\usage{
WDim(left = 0, bottom = 0, width = 1, height = 1, nr = 1, nc = 1,
  text.x = 0, text.y = 0, text.just = c("center", "center"),
  column.split = NULL, row.split = NULL)
}
\arguments{
\item{left}{left coordinate}

\item{bottom}{bottom coordinate}

\item{width}{width}

\item{height}{height}

\item{nr}{number of row}

\item{nc}{number of column}

\item{text.x}{x anchor for text}

\item{text.y}{y anchor for text}

\item{text.just}{just for text}

\item{column.split}{a list of WDim objects for column split}

\item{row.split}{a list of WDim objects for row split}
}
\value{
a WDim object
}
\description{
class WDim
}
