% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{WGG}
\alias{WGG}
\title{WGG object
form ggplot with coordinates}
\usage{
WGG(ggobj, dm = NULL, name = "")
}
\arguments{
\item{ggobj}{ggplot plotting object}

\item{dm}{dimension}

\item{name}{name}
}
\value{
WGG object
}
\description{
WGG object
form ggplot with coordinates
}
