% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob.R
\name{WGrob}
\alias{WGrob}
\title{WGrob object
plot from a gList of grob objects}
\usage{
WGrob(glist, dm = NULL, name = "")
}
\arguments{
\item{glist}{gList object}

\item{dm}{dimension}

\item{name}{name}
}
\value{
WGrob object
}
\description{
WGrob object
plot from a gList of grob objects
}
