% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgroup.R
\name{WGroup}
\alias{WGroup}
\title{Construct a WGroup}
\usage{
WGroup(..., name = "", group.dm = NULL, group.from.member = FALSE,
  mar = WMar(), affine = FALSE, nr = NULL, nc = NULL)
}
\arguments{
\item{...}{plotting objects to be grouped}

\item{name}{name of the group}

\item{group.dm}{group dimension, by default use the dm of the merge of members}

\item{group.from.member}{group merged from member coordinates (require affine == FALSE),
the supplied group.dm is ignored}

\item{mar}{a WMar object}

\item{affine}{whether the group members are on affine coordinates already}

\item{nr}{number of rows}

\item{nc}{number of columns}
}
\value{
a WGroup object
}
\description{
Construct a WGroup
}
