% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{WLegendV}
\alias{WLegendV}
\title{WLegendV}
\usage{
WLegendV(x = NULL, dm = NULL, name = "", n.stops = 20, n.text = 5,
  label.fontsize = 12, width = 0.05, height = 0.02, ...)
}
\arguments{
\item{x}{a name or a plotting object, if NULL use the last plotting object}

\item{dm}{position}

\item{name}{name of the plotted legend}

\item{n.stops}{number of stops in computing continuous legend}

\item{n.text}{number of text labels in continuous legend}

\item{label.fontsize}{label font size}

\item{width}{width of each unit in plotted legend}

\item{height}{height of each unit in plotted legend}

\item{...}{additional options to WHeatmap}
}
\value{
an object of class WLegendV
}
\description{
a vertical legend
}
\examples{
WHeatmap(matrix(1:4,nrow=2))+WLegendV(NULL, RightOf())
}
