% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dims.R
\name{WPosition}
\alias{WPosition}
\title{place an arbitrary position w.r.t a subplot}
\usage{
WPosition(anchor.x, anchor.y, x = NULL, just = c("left", "bottom"),
  data.coord = FALSE)
}
\arguments{
\item{anchor.x}{x coordinates}

\item{anchor.y}{y coordinates}

\item{x}{plotting object to anchor}

\item{just}{adjustment of new plot}

\item{data.coord}{whether the coordinates is in term of data}
}
\value{
a WDimGenerator object
}
\description{
place an arbitrary position w.r.t a subplot
}
