% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rect.R
\name{WRect}
\alias{WRect}
\title{construct a WRect}
\usage{
WRect(obj = NULL, x.span = NULL, y.span = NULL, color = "black",
  lwd = 3, fill = NA, name = "")
}
\arguments{
\item{obj}{a plotting object or its name}

\item{x.span}{x-axis/horizontal span (e.g., c(2,4))}

\item{y.span}{y-axis/vertical span (e.g., c(5,9))}

\item{color}{edge color}

\item{lwd}{edge width}

\item{fill}{fill color}

\item{name}{name}
}
\value{
a WRect object
}
\description{
construct a WRect
}
