% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{WRowBind}
\alias{WRowBind}
\title{row bind non-overlapping objects}
\usage{
WRowBind(..., nr = NULL, nc = NULL)
}
\arguments{
\item{...}{plotting objects}

\item{nr}{number of rows}

\item{nc}{number of columns}
}
\value{
an object of class WDim
}
\description{
row bind non-overlapping objects
}
