% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{addMask}
\alias{addMask}
\title{Add probes to mask}
\usage{
addMask(sset, probes)
}
\arguments{
\item{sset}{a \code{SigSet}}

\item{probes}{a vector of probe IDs or a logical vector with TRUE
representing masked probes}
}
\value{
a \code{SigSet} with added mask
}
\description{
This function essentially merge existing probe masking
with new prboes to mask
}
\examples{
sset <- sesameDataGet('EPIC.1.LNCaP')$sset
sum(mask(sset))
sum(mask(addMask(sset, c("cg14057072", "cg22344912"))))
}
