% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{addMask}
\alias{addMask}
\title{Add probes to mask}
\usage{
addMask(sdf, probes)
}
\arguments{
\item{sdf}{a \code{SigDF}}

\item{probes}{a vector of probe IDs or a logical vector with TRUE
representing masked probes}
}
\value{
a \code{SigDF} with added mask
}
\description{
This function essentially merge existing probe masking
with new probes to mask
}
\examples{
sdf <- sesameDataGet('EPIC.1.SigDF')
sum(sdf$mask)
sum(addMask(sdf, c("cg14057072", "cg22344912"))$mask)
}
