% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strain.R
\name{betaToAF}
\alias{betaToAF}
\title{convert betas to variant allele frequency}
\usage{
betaToAF(betas)
}
\arguments{
\item{betas}{beta value}
}
\value{
SNP variant allele frequency
}
\description{
convert betas to variant allele frequency
}
\examples{
sesameDataCache("MM285") # if not done yet
sdf <- sesameDataGet('MM285.1.SigDF')
vafs <- betaToAF(getBetas(dyeBiasNL(noob(sdf))))
}
