% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{buildControlMatrix450k}
\alias{buildControlMatrix450k}
\title{Build control summary matrix}
\usage{
buildControlMatrix450k(sset)
}
\arguments{
\item{sset}{an object of class SigSet}
}
\value{
a vector with control summaries
}
\description{
The function takes a \code{SigSet} as input and outputs the control matrix
summary vector. This vector summarizes one single QC metric for the array
control. This includes bisulfite control, stain signal extension efficiency
and more.
}
\examples{
sset <- makeExampleSeSAMeDataSet()
control.summary <- buildControlMatrix450k(sset)

}
