% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kyCG.R
\name{calcDatabaseSetStatisticsAll}
\alias{calcDatabaseSetStatisticsAll}
\title{calcDatabaseSetStatisticsAll builds dataset for a given betas matrix 
composed of engineered features from the given database sets}
\usage{
calcDatabaseSetStatisticsAll(betas, databaseSets)
}
\arguments{
\item{betas}{matrix of beta values where probes are on the rows and samples
are on the columns}

\item{databaseSets}{List of vectors corresponding to probe locations for
which the features will be extracted}
}
\value{
Vector for a given sample columns are features across different
databaseSets
}
\description{
calcDatabaseSetStatisticsAll builds dataset for a given betas matrix 
composed of engineered features from the given database sets
}
\examples{

library(SummarizedExperiment)
se = sesameDataGet('MM285.20Kx467.SE')
samplesheet = colData(se)[, c("Mouse_Age_Months", "Mouse_Age_Days", "Sex", 
"Strain_Corrected","Tissue_Corrected", 'Genotype')]
betas = assay(se)
databaseSetNames = c('KYCG.MM285.seqContextN.20210630', 
'KYCG.MM285.probeType.20210630')
databaseSets = do.call(c, lapply(databaseSetNames, sesameDataGet))
calcDatabaseSetStatisticsAll(betas, databaseSets=databaseSets)
sesameDataClearCache()

}
