% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.R
\name{cnSegmentation}
\alias{cnSegmentation}
\title{perform copy number segmentation}
\usage{
cnSegmentation(sset, ssets.normal, refversion = "hg19")
}
\arguments{
\item{sset}{\code{SigSet}}

\item{ssets.normal}{\code{SigSet} for normalization}

\item{refversion}{hg19 or hg38}
}
\value{
an object of \code{CNSegment}
}
\description{
perform copy number segmentation
}
\examples{
sset <- readRDS(system.file(
    "extdata", "EPIC.sset.LNCaP.Rep1.chr4.rds", package = "sesameData"))
ssets.normal <- readRDS(system.file(
    "extdata", "EPIC.ssets.5normals.chr4.rds", package = "sesameData"))
seg <- cnSegmentation(sset, ssets.normal)

}
