% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.R
\name{cnSegmentation}
\alias{cnSegmentation}
\title{Perform copy number segmentation}
\usage{
cnSegmentation(sset, ssets.normal, refversion = "hg19")
}
\arguments{
\item{sset}{\code{SigSet}}

\item{ssets.normal}{\code{SigSet} for normalization}

\item{refversion}{hg19 or hg38}
}
\value{
an object of \code{CNSegment}
}
\description{
Perform copy number segmentation using the signals in the signal set.
The function takes a \code{SigSet} for the target sample and a set of
normal \code{SigSet} for the normal samples. An optional arguments specifies
the version of genome build that the inference will operate on. The function
outputs an object of class \code{CNSegment} with signals for the segments (
seg.signals), chromosome information (chrominfo), the bin coordinates (
bin.coords) and bin signals (bin.signals).
}
\examples{
sset <- readRDS(system.file(
    "extdata", "EPIC.sset.LNCaP.Rep1.chr4.rds", package = "sesameData"))
ssets.normal <- readRDS(system.file(
    "extdata", "EPIC.ssets.5normals.chr4.rds", package = "sesameData"))
seg <- cnSegmentation(sset, ssets.normal)

}
