% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.R
\name{cnSegmentation}
\alias{cnSegmentation}
\title{Perform copy number segmentation}
\usage{
cnSegmentation(sdf, sdfs.normal = NULL, refversion = c("hg19", "hg38"))
}
\arguments{
\item{sdf}{\code{SigDF}}

\item{sdfs.normal}{a list of \code{SigDF}s for normalization, if not given,
use the stored normal data from sesameData. However, we do recommend using
a matched copy number normal dataset for normalization.}

\item{refversion}{hg19 or hg38}
}
\value{
an object of \code{CNSegment}
}
\description{
Perform copy number segmentation using the signals in the signal set.
The function takes a \code{SigDF} for the target sample and a set of
normal \code{SigDF} for the normal samples. An optional arguments specifies
the version of genome build that the inference will operate on. The function
outputs an object of class \code{CNSegment} with signals for the segments (
seg.signals), the bin coordinates (
bin.coords) and bin signals (bin.signals).
}
\examples{

sesameDataCache("EPIC") # in case not done yet

sdf <- sesameDataGet('EPIC.1.SigDF')
sdfs.normal <- sesameDataGet('EPIC.5.SigDF.normal')
seg <- cnSegmentation(sdf, sdfs.normal)

}
