% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KYCG.R
\name{compareDatbaseSetOverlap}
\alias{compareDatbaseSetOverlap}
\title{calculates the pariwise overlap between given list
of database sets using a distance metric.}
\usage{
compareDatbaseSetOverlap(databases = NA, metric = "Jaccard")
}
\arguments{
\item{databases}{List of vectors corresponding to the database sets of
interest with associated meta data as an attribute to each element. Optional.
(Default: NA)}

\item{metric}{String representing the similarity metric to use. Optional.
(Default: "Jaccard").}
}
\value{
An upper triangular matrix containing a metric (Jaccard) comparing
the pairwise distances between database sets.
}
\description{
calculates the pariwise overlap between given list
of database sets using a distance metric.
}
