% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kyCG.R
\name{createDatabaseSetNetwork}
\alias{createDatabaseSetNetwork}
\title{createGeneNetwork creates databaseSet network using the Jaccard index.}
\usage{
createDatabaseSetNetwork(databaseSets)
}
\arguments{
\item{databaseSets}{Vector of probes corresponding to a single database set
of interest.}
}
\value{
ggplot lollipop plot
}
\description{
createGeneNetwork creates databaseSet network using the Jaccard index.
}
\examples{
databaseSets = list(a=c("a", "b"), b=c("a", "e", "f"), c=c("q", "a"))
createDatabaseSetNetwork(databaseSets)

}
