% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kyCG.R
\name{createDatabaseSetNetwork}
\alias{createDatabaseSetNetwork}
\title{createGeneNetwork creates databaseSet network using the Jaccard index.}
\usage{
createDatabaseSetNetwork(databaseSets)
}
\arguments{
\item{databaseSets}{Vector of probes corresponding to a single database set
of interest.}
}
\value{
ggplot lollipop plot
}
\description{
createGeneNetwork creates databaseSet network using the Jaccard index.
}
\examples{

databaseSetNames = c('KYCG.MM285.seqContextN.20210630')
databaseSets = do.call(c, lapply(databaseSetNames, sesameDataGet))
createDatabaseSetNetwork(databaseSets)
sesameDataClearCache()

}
