% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KYCG.R
\name{createDBNetwork}
\alias{createDBNetwork}
\title{createGeneNetwork creates database network using the Jaccard index.}
\usage{
createDBNetwork(databases)
}
\arguments{
\item{databases}{Vector of probes corresponding to a single database set
of interest.}
}
\value{
ggplot lollipop plot
}
\description{
createGeneNetwork creates database network using the Jaccard index.
}
