% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kyCG.R
\name{dbStats}
\alias{dbStats}
\title{dbStats builds dataset for a given betas matrix 
composed of engineered features from the given database sets}
\usage{
dbStats(betas, dbs, fun = mean, na.rm = TRUE)
}
\arguments{
\item{betas}{matrix of beta values where probes are on the rows and
samples are on the columns}

\item{dbs}{List of vectors corresponding to probe locations for
which the features will be extracted}

\item{fun}{aggregation function, default to mean}

\item{na.rm}{whether to remove NA}
}
\value{
matrix with samples on the rows and database set on the columns
}
\description{
dbStats builds dataset for a given betas matrix 
composed of engineered features from the given database sets
}
\examples{

library(SummarizedExperiment)
se <- sesameDataGet('MM285.20Kx467.SE')
stats <- dbStats(assay(se), c('KYCG.MM285.probeType.20210630'))
head(stats)
sesameDataClearCache()

}
