% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KYCG.R
\name{dbStats}
\alias{dbStats}
\title{dbStats builds dataset for a given betas matrix 
composed of engineered features from the given database sets}
\usage{
dbStats(
  betas,
  databases,
  fun = mean,
  na.rm = TRUE,
  n_min = NULL,
  f_min = 0.1,
  long = FALSE
)
}
\arguments{
\item{betas}{matrix of beta values where probes are on the rows and
samples are on the columns}

\item{databases}{List of vectors corresponding to probe locations for
which the features will be extracted}

\item{fun}{aggregation function, default to mean}

\item{na.rm}{whether to remove NA}

\item{n_min}{min number of non-NA for aggregation function to apply,
overrides f_min}

\item{f_min}{min fraction of non-NA for aggregation function to apply}

\item{long}{produce long-form result}
}
\value{
matrix with samples on the rows and database set on the columns
}
\description{
dbStats builds dataset for a given betas matrix 
composed of engineered features from the given database sets
}
\examples{

library(SummarizedExperiment)
se <- sesameDataGet('MM285.467.SE.tissue20Kprobes')
head(dbStats(assay(se), "MM285.chromHMM")[,1:3])
sesameDataGet_resetEnv()

}
