% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deidentify.R
\name{deIdentify}
\alias{deIdentify}
\title{De-identify IDATs by removing SNP probes}
\usage{
deIdentify(path, out_path = NULL, snps = NULL, mft = NULL, secret = NULL)
}
\arguments{
\item{path}{input IDAT file}

\item{out_path}{output IDAT file}

\item{snps}{SNP definition, if not given, default to SNP probes}

\item{mft}{sesame-compatible manifest if non-standard}

\item{secret}{a numeric integer, max 2^31-1
when not NULL randomize the signal intensities
when NULL (default), this sets all SNP intensities to zero.}
}
\description{
Mask SNP probe intensity mean by zero.
}
\examples{

temp_out <- tempfile("test")
deIdentify(system.file(
    "extdata", "4207113116_A_Grn.idat", package = "sesameData"),
     temp_out)
unlink(temp_out)
}
