% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{detectionPnegEcdf}
\alias{detectionPnegEcdf}
\title{Detection P-value based on ECDF of negative control}
\usage{
detectionPnegEcdf(sset, force = FALSE)
}
\arguments{
\item{sset}{a \code{SigSet}}

\item{force}{force rerun even if result already exists}
}
\value{
detection p-value
}
\description{
The function takes a \code{SigSet} as input, computes detection p-value
using negative control probes' empirical distribution and returns a new
\code{SigSet} with an updated pval slot.
}
\examples{
sset <- makeExampleSeSAMeDataSet()
sset <- detectionPnegEcdf(sset)
}
