% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{detectionPnegEcdf}
\alias{detectionPnegEcdf}
\title{detection P-value based on ECDF of negative control}
\usage{
detectionPnegEcdf(sset)
}
\arguments{
\item{sset}{a \code{SignalSet}}
}
\value{
detection p-value
}
\description{
detection P-value based on ECDF of negative control
}
\examples{
sset <- makeExampleSeSAMeDataSet()
sset$pval <- detectionPnegEcdf(sset)

}
