% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{detectionPnegNormGS}
\alias{detectionPnegNormGS}
\title{Detection P-value emulating Genome Studio}
\usage{
detectionPnegNormGS(sset)
}
\arguments{
\item{sset}{a \code{SigSet}}
}
\value{
detection p-value
}
\description{
The function takes a \code{SigSet} as input, computes detection p-value
using negative control probes parametrized in a normal distribution a la
Genome Studio and returns a new \code{SigSet} with an updated pval slot.
}
\examples{
sset <- makeExampleSeSAMeDataSet()
sset <- detectionPnegNormGS(sset)

}
