% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{detectionPnegNormGS}
\alias{detectionPnegNormGS}
\title{detection P-value emulating Genome Studio}
\usage{
detectionPnegNormGS(sset)
}
\arguments{
\item{sset}{a \code{SignalSet}}
}
\value{
detection p-value
}
\description{
detection P-value emulating Genome Studio
}
\examples{
sset <- makeExampleSeSAMeDataSet()
sset$pval <- detectionPnegNormGS(sset)

}
