% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{detectionPnegNormGS}
\alias{detectionPnegNormGS}
\title{Detection P-value emulating Genome Studio}
\usage{
detectionPnegNormGS(sset)
}
\arguments{
\item{sset}{a \code{SigSet}}
}
\value{
detection p-value
}
\description{
Returns a new \code{SigSet} with updated pval slot
}
\examples{
sset <- makeExampleSeSAMeDataSet()
sset <- detectionPnegNormGS(sset)

}
