% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{detectionPnegNormTotal}
\alias{detectionPnegNormTotal}
\title{detection P-value based on normal fitting the negative controls, channels are first summed}
\usage{
detectionPnegNormTotal(sset)
}
\arguments{
\item{sset}{a \code{SignalSet}}
}
\value{
detection p-value
}
\description{
detection P-value based on normal fitting the negative controls, channels are first summed
}
