% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{detectionPnegNormTotal}
\alias{detectionPnegNormTotal}
\title{Detection P-value based on normal fitting the negative controls,
channels are first summed}
\usage{
detectionPnegNormTotal(sset)
}
\arguments{
\item{sset}{a \code{SigSet}}
}
\value{
detection p-value
}
\description{
The function takes a \code{SigSet} as input, computes detection p-value
using negative control probes parametrized in a normal distribution with
the two channels summed first and returns a new \code{SigSet} with an
updated pval slot.
}
\examples{
sset <- makeExampleSeSAMeDataSet()
sset <- detectionPnegNormTotal(sset)

}
