% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{detectionPnegNormTotal}
\alias{detectionPnegNormTotal}
\title{Detection P-value based on normal fitting the negative controls,
channels are first summed}
\usage{
detectionPnegNormTotal(sset)
}
\arguments{
\item{sset}{a \code{SigSet}}
}
\value{
detection p-value
}
\description{
Returns a new \code{SigSet} with updated pval slot
}
\examples{
sset <- makeExampleSeSAMeDataSet()
sset <- detectionPnegNormTotal(sset)

}
