% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{detectionPoobEcdf}
\alias{detectionPoobEcdf}
\alias{pOOBAH}
\title{Detection P-value based on ECDF of out-of-band signal}
\usage{
detectionPoobEcdf(sdf, return.pval = FALSE, pval.threshold = 0.05)

pOOBAH(sdf, return.pval = FALSE, pval.threshold = 0.05)
}
\arguments{
\item{sdf}{a \code{SigDF}}

\item{return.pval}{whether to return p-values, instead of a
masked \code{SigDF}}

\item{pval.threshold}{minimum p-value to mask}
}
\value{
a \code{SigDF}, or a p-value vector if return.pval is TRUE
}
\description{
aka pOOBAH (p-vals by Out-Of-Band Array Hybridization)
}
\details{
The function takes a \code{SigDF} as input, computes detection p-value
using out-of-band probes empirical distribution and returns a new
\code{SigDF} with an updated mask slot.
}
\examples{
sdf <- sesameDataGet("HM450.1.TCGA.PAAD")$sdf
sum(mask(sdf))
sdf_with_mask <- detectionPoobEcdf(sdf)
sum(mask(sdf_with_mask))
sdf <- sesameDataGet("HM450.1.TCGA.PAAD")$sdf
sum(mask(sdf))
sdf_with_mask <- pOOBAH(sdf)
sum(mask(sdf_with_mask))
}
