% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{detectionPoobEcdf}
\alias{detectionPoobEcdf}
\title{Detection P-value based on ECDF of out-of-band signal}
\usage{
detectionPoobEcdf(sset)
}
\arguments{
\item{sset}{a \code{SigSet}}
}
\value{
detection p-value
}
\description{
The function takes a \code{SigSet} as input, computes detection p-value
using out-of-band probes empirical distribution and returns a new
\code{SigSet} with an updated pval slot.
}
\examples{
sset <- makeExampleSeSAMeDataSet()
sset <- detectionPoobEcdf(sset)

}
