% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{detectionPoobEcdf}
\alias{detectionPoobEcdf}
\title{detection P-value based on ECDF of out-of-band signal}
\usage{
detectionPoobEcdf(sset)
}
\arguments{
\item{sset}{a \code{SignalSet}}
}
\value{
detection p-value
}
\description{
detection P-value based on ECDF of out-of-band signal
}
\examples{
sset <- makeExampleSeSAMeDataSet()
sset$pval <- detectionPoobEcdf(sset)

}
