% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{detectionPoobEcdf2}
\alias{detectionPoobEcdf2}
\title{Detection P-value based on ECDF of out-of-band signal}
\usage{
detectionPoobEcdf2(sset)
}
\arguments{
\item{sset}{a \code{SigSet}}
}
\value{
detection p-value
}
\description{
The only difference between this and detectionPoobEcdf is that
the out-of-band color channel is inferred from data instead of
using the channel specified in the manifest.
}
\details{
aka pOOBAH (p-vals by Out-Of-Band Array Hybridization)

The function takes a \code{SigSet} as input, computes detection p-value
using out-of-band probes empirical distribution and returns a new
\code{SigSet} with an updated pval slot.
}
\examples{
sset <- makeExampleSeSAMeDataSet()
sset <- detectionPoobEcdf2(sset)

}
