% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{detectionPoobEcdf2}
\alias{detectionPoobEcdf2}
\alias{pOOBAH2}
\title{Detection P-value based on ECDF of out-of-band signal}
\usage{
detectionPoobEcdf2(sdf, return.pval = FALSE, pval.threshold = 0.05)

pOOBAH2(sdf, return.pval = FALSE, pval.threshold = 0.05)
}
\arguments{
\item{sdf}{a \code{SigDF}}

\item{return.pval}{whether to return p-values, instead of a
masked \code{SigDF}}

\item{pval.threshold}{minimum p-value to mask}
}
\value{
a \code{SigDF}, or a p-value vector if return.pval is TRUE
}
\description{
aka pOOBAH2 (p-vals by Out-Of-Band Array Hybridization)
}
\details{
The function takes a \code{SigDF} as input, computes detection p-value
using out-of-band probes empirical distribution and returns a new
\code{SigDF} with an updated mask slot.

The difference between this function and the original pOOBAH
is that pOOBAH2 is based on background-subtracted and dyebias
corrected signal and do not distinguish the color channel difference.
}
\examples{
sdf <- sesameDataGet("EPIC.1.SigDF")
sum(sdf$mask)
sdf <- detectionPoobEcdf(sdf)
sum(sdf$mask)
sdf <- sesameDataGet("EPIC.1.SigDF")
sum(sdf$mask)
sdf <- pOOBAH2(sdf)
sum(sdf$mask)
}
