% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{detectionZero}
\alias{detectionZero}
\title{Detection P-value set to all zero}
\usage{
detectionZero(sset)
}
\arguments{
\item{sset}{a \code{SigSet}}
}
\value{
detection p-value set to all zero
}
\description{
Detection P-value set to all zero
}
\examples{
sset <- makeExampleSeSAMeDataSet()
sset <- detectionZero(sset)

}
