% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_methylation.R
\name{DMR}
\alias{DMR}
\title{Find Differentially Methylated Region (DMR)}
\usage{
DMR(
  betas,
  formula = NULL,
  cf = NULL,
  dist.cutoff = NULL,
  meta = NULL,
  seg.per.locus = 0.5,
  platform = NULL,
  refversion = NULL,
  ...
)
}
\arguments{
\item{betas}{beta values for distance calculation}

\item{formula}{formula}

\item{cf}{coefficient table from diffMeth, when NULL will be computed from
beta. If cf is given, meta and formula are ignored.}

\item{dist.cutoff}{distance cutoff (default to use dist.cutoff.quantile)}

\item{meta}{data frame for sample information, column names
are predictor variables (e.g., sex, age, treatment, tumor/normal etc)
and are referenced in formula. Rows are samples.}

\item{seg.per.locus}{number of segments per locus
higher value leads to more segments}

\item{platform}{EPIC, HM450, MM285, ...}

\item{refversion}{hg38, hg19, mm10, ...}

\item{...}{additional parameters to DML}
}
\value{
coefficient table with segment ID and segment P-value
}
\description{
This subroutine uses Euclidean distance to group CpGs and
then combine p-values for each segment. The function performs DML test first
if cf is NULL. It groups the probe testing results into differential
methylated regions in a coefficient table with additional columns
designating the segment ID and statistical significance (P-value) testing
the segment.
}
\examples{

sesameDataCache("HM450") # in case not done yet
data <- sesameDataGet('HM450.76.TCGA.matched')
cf <- DMR(data$betas, ~type, meta=data$sampleInfo)

}
