% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_methylation.R
\name{DMR}
\alias{DMR}
\title{Find Differentially Methylated Region (DMR)}
\usage{
DMR(betas, sample.data = NULL, formula = NULL, cf = NULL,
  dist.cutoff = NULL, seg.per.locus = 0.5, platform = "EPIC",
  refversion = "hg38", ...)
}
\arguments{
\item{betas}{beta values for distance calculation}

\item{sample.data}{data frame for sample information, column names
are predictor variables (e.g., sex, age, treatment, tumor/normal etc)
and are referenced in formula. Rows are samples.}

\item{formula}{formula}

\item{cf}{coefficient table from diffMeth, when NULL will be computed from
beta. If cf is given, sample.data and formula are ignored.}

\item{dist.cutoff}{distance cutoff (default to use dist.cutoff.quantile)}

\item{seg.per.locus}{number of segments per locus
higher value leads to more segments}

\item{platform}{EPIC or HM450}

\item{refversion}{hg38 or hg19}

\item{...}{additional parameters to DML}
}
\value{
coefficient table with segment ID and segment P-value
}
\description{
This subroutine uses Euclidean distance to group CpGs and
then combine p-values for each segment. The function performs DML test first
if cf is NULL. It groups the probe testing results into differential
methylated regions in a coefficient table with additional columns
designating the segment ID and statistical significance (P-value) testing
the segment.
}
\examples{
betas <- readRDS(system.file(
    'extdata','HM450.betas.76matchedTCGAchr20.rds',
    package='sesameData'))
sample.info <- readRDS(system.file(
    'extdata','HM450.sampleinfo.76matchedTCGAchr20.rds',
    package='sesameData'))
cf <- DMR(betas, sample.info, ~type)
}
