% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dye_bias.R
\name{dyeBiasCorr}
\alias{dyeBiasCorr}
\title{Correct dye bias}
\usage{
dyeBiasCorr(sset, ref = NULL, in.place = FALSE)
}
\arguments{
\item{sset}{a \code{SignalSet}}

\item{ref}{reference signal level}

\item{in.place}{modify \code{SignalSet} in place, faster}
}
\value{
a normalized \code{SignalSet}
}
\description{
Correct dye bias
}
\examples{
sset <- makeExampleSeSAMeDataSet()
sset.db <- dyeBiasCorr(sset)
}
