% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dye_bias.R
\name{dyeBiasCorr}
\alias{dyeBiasCorr}
\title{Correct dye bias in by linear scaling.}
\usage{
dyeBiasCorr(sset, ref = NULL)
}
\arguments{
\item{sset}{a \code{SigSet}}

\item{ref}{reference signal level}
}
\value{
a normalized \code{SigSet}
}
\description{
The function takes a \code{SigSet} as input and scale both the Grn and Red
signal to a reference (ref) level. If the reference level is not given, it
is set to the mean intensity of all the in-band signals. The function
returns a \code{SigSet} with dye bias corrected.
}
\examples{
sset <- readRDS(system.file(
    "extdata", "EPIC.sset.LNCaP.Rep1.chr4.rds", package = "sesameData"))
sset.db <- dyeBiasCorr(sset)
}
