% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dye_bias.R
\name{dyeBiasCorrMostBalanced}
\alias{dyeBiasCorrMostBalanced}
\title{Correct dye bias using most balanced sample}
\usage{
dyeBiasCorrMostBalanced(ssets)
}
\arguments{
\item{ssets}{a list of normalized \code{SignalSet}s}
}
\value{
a list of normalized \code{SignalSet}s
}
\description{
In practice, it doesn't matter as long as the
reference level does not deviate much.
}
\examples{
ssets <- list(s1=makeExampleSeSAMeDataSet(), s2=makeExampleSeSAMeDataSet())
ssets.db <- dyeBiasCorrMostBalanced(ssets)
}
