% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dye_bias.R
\name{dyeBiasCorrTypeINorm}
\alias{dyeBiasCorrTypeINorm}
\title{Dye bias correction by matching green and red to mid point}
\usage{
dyeBiasCorrTypeINorm(sset)
}
\arguments{
\item{sset}{a \code{SigSet}}
}
\value{
a \code{SigSet} after dye bias correction.
}
\description{
This function compares the Type-I Red probes and Type-I Grn probes and
generates and mapping to correct signal of the two channels to the middle.
The function takes one single \code{SigSet} and returns a \code{SigSet}
with dye bias corrected.
}
\examples{
sset <- sesameDataGet('EPIC.1.LNCaP')$sset
sset.db <- dyeBiasCorrTypeINorm(sset)
}
