% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dye_bias.R
\name{dyeBiasCorrTypeINorm}
\alias{dyeBiasCorrTypeINorm}
\alias{dyeBiasNL}
\title{Dye bias correction by matching green and red to mid point}
\usage{
dyeBiasCorrTypeINorm(sdf)

dyeBiasNL(sdf)
}
\arguments{
\item{sdf}{a \code{SigDF}}
}
\value{
a \code{SigDF} after dye bias correction.
}
\description{
This function compares the Type-I Red probes and Type-I Grn probes and
generates and mapping to correct signal of the two channels to the middle.
The function takes one single \code{SigDF} and returns a \code{SigDF}
with dye bias corrected.
}
\examples{
sesameDataCache("EPIC") # if not done yet
sdf <- sesameDataGet('EPIC.1.LNCaP')$sdf
sdf.db <- dyeBiasCorrTypeINorm(sdf)
sdf <- sesameDataGet("HM450.1.TCGA.PAAD")$sdf
sdf <- dyeBiasNL(sdf)
}
