% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dye_bias.R
\name{dyeBiasCorrTypeINorm}
\alias{dyeBiasCorrTypeINorm}
\alias{dyeBiasNL}
\title{Dye bias correction by matching green and red to mid point}
\usage{
dyeBiasCorrTypeINorm(sdf, verbose = FALSE)

dyeBiasNL(sdf, verbose = FALSE)
}
\arguments{
\item{sdf}{a \code{SigDF}}

\item{verbose}{print more messages}
}
\value{
a \code{SigDF} after dye bias correction.
}
\description{
This function compares the Type-I Red probes and Type-I Grn probes and
generates and mapping to correct signal of the two channels to the middle.
The function takes one single \code{SigDF} and returns a \code{SigDF}
with dye bias corrected.
}
\examples{
sesameDataCache() # if not done yet
sdf <- sesameDataGet('EPIC.1.SigDF')
sdf.db <- dyeBiasCorrTypeINorm(sdf)
sdf <- sesameDataGet('EPIC.1.SigDF')
sdf <- dyeBiasNL(sdf)
}
