% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dye_bias.R
\name{dyeBiasCorrTypeINorm}
\alias{dyeBiasCorrTypeINorm}
\title{Dye bias correction by matching green and red to mid point}
\usage{
dyeBiasCorrTypeINorm(sset, in.place = FALSE)
}
\arguments{
\item{sset}{a \code{SignalSet}}

\item{in.place}{to modify sset in place}
}
\value{
a modified \code{SignalSet}
}
\description{
Dye bias correction by matching green and red to mid point
}
\examples{
sset <- makeExampleSeSAMeDataSet()
sset.db <- dyeBiasCorrTypeINorm(sset)
}
