% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellComposition.R
\name{estimateCellComposition}
\alias{estimateCellComposition}
\title{Estimate cell composition using reference}
\usage{
estimateCellComposition(g, q, refine = TRUE, dichotomize = FALSE, ...)
}
\arguments{
\item{g}{reference methylation}

\item{q}{target measurement: length(q) == nrow(g)}

\item{refine}{to refine estimate, takes longer}

\item{dichotomize}{to dichotomize query beta value before estimate,
this relieves unclean background subtraction}

\item{...}{extra parameters for optimization, this includes
temp - annealing temperature (0.5)
maxIter - maximum iteration to stop after converge (1000)
delta - delta score to reset counter (0.0001)
verbose - output debug info (FALSE)}
}
\value{
a list of fraction, min error and unknown component methylation state
}
\description{
Estimate cell composition using reference
}
