% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_composition.R
\name{estimateCellComposition}
\alias{estimateCellComposition}
\title{Estimate cell composition using reference}
\usage{
estimateCellComposition(g, q, refine = TRUE, dichotomize = FALSE, ...)
}
\arguments{
\item{g}{reference methylation}

\item{q}{target measurement: length(q) == nrow(g)}

\item{refine}{to refine estimate, takes longer}

\item{dichotomize}{to dichotomize query beta value before estimate,
this relieves unclean background subtraction}

\item{...}{extra parameters for optimization, this includes
temp - annealing temperature (0.5)
maxIter - maximum iteration to stop after converge (1000)
delta - delta score to reset counter (0.0001)
verbose - output debug info (FALSE)}
}
\value{
a list of fraction, min error and unknown component methylation state
}
\description{
This is a reference-based cell composition estimation. The function takes a
reference methylation status matrix (rows for probes and columns for cell
types, can be obtained by getRefSet function) and a query beta value
measurement. The length of the target beta values should be the same as
the number of rows of the reference matrix. The method assumes one unknown
component. It outputs a list containing the estimated cell fraction, the
error of optimization and methylation status of the unknown component.
}
\examples{
g <- diffRefSet(getRefSet(platform='HM450'))
M <- ncol(g)
trueFrac <- runif(M+1)
trueFrac <- trueFrac / sum(trueFrac)
g0 <- sample(c(0,1), nrow(g), replace=TRUE)
q <- cbind(g0, g) \%*\% trueFrac + rnorm(length(g0), mean=0, sd = 0.0)
q[q<0] <- 0
q[q>1] <- 1
est <- estimateCellComposition(g, q)
}
