% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{formatVCF}
\alias{formatVCF}
\title{Convert SNP from Infinium array to VCF file}
\usage{
formatVCF(sdf, annoS, annoI, vcf = NULL, genome = "hg19", verbose = FALSE)
}
\arguments{
\item{sdf}{SigDF}

\item{annoS}{SNP variant annotation, available at
https://github.com/zhou-lab/InfiniumAnnotationV1/tree/main/Anno/EPIC
EPIC.hg19.snp_overlap_b151.rds
EPIC.hg38.snp_overlap_b151.rds}

\item{annoI}{Infinium-I variant annotation, available at
https://github.com/zhou-lab/InfiniumAnnotationV1/tree/main/Anno/EPIC
EPIC.hg19.typeI_overlap_b151.rds
EPIC.hg38.typeI_overlap_b151.rds}

\item{vcf}{output VCF file path, if NULL output to console}

\item{genome}{genome}

\item{verbose}{print more messages}
}
\value{
VCF file. If vcf is NULL, a data.frame is output to
console. The data.frame does not contain VCF headers.

Note the vcf is not sorted. You can sort with
awk '$1 ~ /^#/ {print $0;next} {print $0 | "sort -k1,1 -k2,2n"}'
}
\description{
Convert SNP from Infinium array to VCF file
}
\examples{
sesameDataCacheAll() # if not done yet
sdf <- sesameDataGet('EPIC.1.SigDF')

\dontrun{
## download annoS and annoI from
## https://github.com/zhou-lab/InfiniumAnnotationV1/tree/main/Anno/EPIC
## output to console
head(formatVCF(sdf, annoS, annoI))
}

}
