% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{formatVCF}
\alias{formatVCF}
\title{Convert SNP from Infinium array to VCF file}
\usage{
formatVCF(sdf, anno, vcf = NULL, genome = "hg38", verbose = FALSE)
}
\arguments{
\item{sdf}{SigDF}

\item{anno}{SNP variant annotation, available at
https://github.com/zhou-lab/InfiniumAnnotationV1/tree/main/Anno/EPIC
EPIC.hg38.snp.tsv.gz}

\item{vcf}{output VCF file path, if NULL output to console}

\item{genome}{genome}

\item{verbose}{print more messages}
}
\value{
VCF file. If vcf is NULL, a data.frame is output to
console. The data.frame does not contain VCF headers.

Note the vcf is not sorted. You can sort with
awk '$1 ~ /^#/ {print $0;next} {print $0 | "sort -k1,1 -k2,2n"}'
}
\description{
Convert SNP from Infinium array to VCF file
}
\examples{
sesameDataCacheAll() # if not done yet
sdf <- sesameDataGet('EPIC.1.SigDF')

\dontrun{
## download anno from
## https://github.com/zhou-lab/InfiniumAnnotationV1/tree/main/Anno/EPIC
## output to console
anno = read_tsv(gzcon(url(sprintf(
"\%s/EPICv2/EPICv2.hg38.snp.tsv.gz",
"https://github.com/zhou-lab/InfiniumAnnotationV1/raw/main/Anno"))))
head(formatVCF(sdf, anno))
}

}
