% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{formatVCF}
\alias{formatVCF}
\title{Convert SNP from Infinium array to VCF file}
\usage{
formatVCF(sset, vcf = NULL, refversion = "hg19", annoS = NULL, annoI = NULL)
}
\arguments{
\item{sset}{SigSet}

\item{vcf}{output VCF file path, if NULL output to console}

\item{refversion}{reference version, currently only support}

\item{annoS}{SNP variant annotation, download if not given}

\item{annoI}{Infinium-I variant annotation, download if not given
hg19 and hg38 in human}
}
\value{
VCF file. If vcf is NULL, a data.frame is output to
console. The data.frame does not contain VCF headers.

Note the vcf is not sorted. You can sort with
awk '$1 ~ /^#/ {print $0;next} {print $0 | "sort -k1,1 -k2,2n"}'
}
\description{
Convert SNP from Infinium array to VCF file
}
\examples{
sset <- sesameDataGet('EPIC.1.LNCaP')$sset

annoS <- getVariantAnno_SNP('EPIC','hg19')
annoI <- getVariantAnno_InfiniumI('EPIC','hg19')
## output to console
head(formatVCF(sset, annoS=annoS, annoI=annoI))

}
