% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{getAFTypeIbySumAlleles}
\alias{getAFTypeIbySumAlleles}
\title{get allele frequency treating type I by summing alleles}
\usage{
getAFTypeIbySumAlleles(sset, quality.mask = TRUE, nondetection.mask = TRUE,
  pval.threshold = 0.05)
}
\arguments{
\item{sset}{\code{SignalSet}}

\item{quality.mask}{whether to mask low quality probes}

\item{nondetection.mask}{whether to mask nondetection}

\item{pval.threshold}{p-value threshold for nondetection mask}
}
\value{
beta values
}
\description{
used for extra allele frequency on Color-Channel-Switching CCS probes
}
\examples{
sset <- readRDS(system.file(
    'extdata','EPIC.sset.LNCaP.Rep1.rds',package='sesameData'))
betas <- getAFTypeIbySumAlleles(sset)
}
