% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{getBetas}
\alias{getBetas}
\title{get beta values}
\usage{
getBetas(sset, quality.mask = TRUE, nondetection.mask = TRUE,
  pval.threshold = 0.1)
}
\arguments{
\item{sset}{\code{SignalSet}}

\item{quality.mask}{whether to mask low quality probes}

\item{nondetection.mask}{whether to mask nondetection}

\item{pval.threshold}{p-value threshold for nondetection mask}
}
\value{
beta values
}
\description{
get beta values
}

