% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{getBetas}
\alias{getBetas}
\title{Get beta Values}
\usage{
getBetas(sset, quality.mask = TRUE, nondetection.mask = TRUE,
  mask.use.tcga = FALSE, pval.threshold = 0.05, sum.typeI = FALSE)
}
\arguments{
\item{sset}{\code{SigSet}}

\item{quality.mask}{whether to mask low quality probes}

\item{nondetection.mask}{whether to mask nondetection}

\item{mask.use.tcga}{whether to use TCGA masking, only applies to HM450}

\item{pval.threshold}{p-value threshold for nondetection mask}

\item{sum.type.I}{whether to sum type I channels}
}
\value{
a numeric vector, beta values
}
\description{
sum.typeI is used for rescuing beta values on
Color-Channel-Switching CCS probes. The function takes a \code{SigSet}
and returns beta value except that Type-I in-band signal and out-of-band
signal are combined. This prevents color-channel switching due to SNPs.
}
\examples{
sset <- sesameDataGet('EPIC.1.LNCaP')$sset
betas <- getBetas(sset)
}
