% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{getBetas}
\alias{getBetas}
\title{get beta values}
\usage{
getBetas(sset, quality.mask = TRUE, nondetection.mask = TRUE,
  mask.use.tcga = FALSE, pval.threshold = 0.05)
}
\arguments{
\item{sset}{\code{SigSet}}

\item{quality.mask}{whether to mask low quality probes}

\item{nondetection.mask}{whether to mask nondetection}

\item{mask.use.tcga}{whether to use TCGA masking, only applies to HM450}

\item{pval.threshold}{p-value threshold for nondetection mask}
}
\value{
a numeric vector, beta values
}
\description{
get beta values
}
\examples{
sset <- makeExampleSeSAMeDataSet('HM450')
betas <- getBetas(sset)
}
