% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.R
\name{getBinCoordinates}
\alias{getBinCoordinates}
\title{Get bin coordinates}
\usage{
getBinCoordinates(seqInfo, gapInfo, probe.coords)
}
\arguments{
\item{seqInfo}{chromosome information object}

\item{gapInfo}{chromosome gap information}

\item{probe.coords}{probe coordinates}
}
\value{
bin.coords
}
\description{
requires GenomicRanges, IRanges
}
