% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kyCG.R
\name{getDatabaseSetOverlap}
\alias{getDatabaseSetOverlap}
\title{getDatabaseSetOverlap tests for the overlap of set of probes (querySet) in a
single given feature (database set)}
\usage{
getDatabaseSetOverlap(querySet, databaseSets, platform = NA, verbose = TRUE)
}
\arguments{
\item{querySet}{Vector of probes corresponding to a single database set
of interest.}

\item{databaseSets}{List of vectors corresponding to the database sets of
interest with associated meta data as an attribute to each element.}

\item{platform}{String corresponding to the type of platform to use. Either
MM285, EPIC, HM450, or HM27. If it is not provided, it will be inferred
from the query set probeIDs (Default: NA).}

\item{verbose}{Logical value indicating whether to display intermediate
text output about the type of test. Optional. (Default: FALSE)}
}
\value{
A sparse data.frame containing all of the meta data from all database
sets.
}
\description{
getDatabaseSetOverlap tests for the overlap of set of probes (querySet) in a
single given feature (database set)
}
\examples{

library(SummarizedExperiment)
MM285.tissueSignature = sesameDataGet('MM285.tissueSignature')
df = rowData(MM285.tissueSignature)
querySet = df$Probe_ID[df$branch == "E-Brain"]
databaseSetNames = c('KYCG.MM285.seqContextN.20210630', 
'KYCG.MM285.designGroup.20210210')
databaseSets = do.call(c, lapply(databaseSetNames, sesameDataGet))
getDatabaseSetOverlap(querySet, databaseSets)
sesameDataClearCache()

}
