% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kyCG.R
\name{getDatabaseSets}
\alias{getDatabaseSets}
\title{getDatabaseSets retrieves database sets from a meta data sheet by querying 
the group, platform, reference columns. The data is returned as a list where 
names correspond to chosen database sets.}
\usage{
getDatabaseSets(
  accessions = NA,
  group = NA,
  platform = NA,
  reference = NA,
  verbose = TRUE
)
}
\arguments{
\item{accessions}{vector containing the characters associated with the
selected database sets; only non-NA locations will be returned. Optional.
(Default: c("20210810_MM285_TFBS_ENCODE").}

\item{group}{string representing the group for which the database sets will
be returned. Optional. (Default: NA).}

\item{platform}{string representing the platform (EPIC, HM450, HM27, MM285)
for which database sets will be returned. Optional. (Default: NA).}

\item{reference}{string representing the reference (hg19, hg38, mm9, mm10) 
for which the database sets will be returned. Optional. (Default NA).}

\item{verbose}{Logical value indicating whether intermediate outputs will be
displayed to console. Optional. (Default: TRUE).}
}
\value{
One list of vectors corresponding to aggregated database sets.
}
\description{
getDatabaseSets retrieves database sets from a meta data sheet by querying 
the group, platform, reference columns. The data is returned as a list where 
names correspond to chosen database sets.
}
\examples{
databaseSetNames = c('KYCG.MM285.seqContextN.20210630', 
'KYCG.MM285.designGroup.20210210')
databaseSets = getDatabaseSets(databaseSetNames, verbose=FALSE)

}
