% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kyCG.R
\name{getDatabaseSets}
\alias{getDatabaseSets}
\title{listDatabaseSets prints which database sets are available for a given release}
\usage{
getDatabaseSets(
  titles = NA,
  group = NA,
  platform = NA,
  reference = NA,
  verbose = TRUE
)
}
\arguments{
\item{titles}{vector containing the characters associated with the
selected database sets; only non-NA locations will be returned. Optional.
(Default: c("20210810_MM285_TFBS_ENCODE").}

\item{group}{string representing the group for which the database sets will
be returned. Optional. (Default: NA).}

\item{platform}{string representing the platform (EPIC, HM450, HM27, MM285)
for which database sets will be returned. Optional. (Default: NA).}

\item{reference}{string representing the reference (hg19, hg38, mm9, mm10) 
for which the database sets will be returned. Optional. (Default NA).}

\item{verbose}{Logical value indicating whether intermediate outputs will be
displayed to console. Optional. (Default: TRUE).}
}
\value{
One list of vectors corresponding to aggregated database sets.

One list of vectors corresponding to aggregated database sets.
}
\description{
listDatabaseSets prints which database sets are available for a given release
}
\examples{
listDatabaseSets()

getDatabaseSets()

}
