% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{getMask}
\alias{getMask}
\title{get probe masking by mask names}
\usage{
getMask(platform = "EPICv2", mask_names = "recommended")
}
\arguments{
\item{platform}{EPICv2, EPIC, HM450, HM27, ...}

\item{mask_names}{mask names (see listAvailableMasks)
by default: "recommended"
see recommendedMaskNames() for detail.}
}
\value{
a vector of probe ID
}
\description{
get probe masking by mask names
}
\examples{

length(getMask("MSA", "recommended"))
length(getMask("EPICv2", "recommended"))
length(getMask("EPICv2", c("recommended", "M_SNPcommon_1pt")))
length(getMask("EPICv2", "M_mapping"))
length(getMask("EPIC"))
length(getMask("HM450"))
length(getMask("MM285"))

}
