% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{getMask}
\alias{getMask}
\title{get probe masking by mask names}
\usage{
getMask(platform = "EPICv2", mask_names = "recommended")
}
\arguments{
\item{platform}{EPICv2, EPIC, HM450, HM27, ...}

\item{mask_names}{mask names (see listAvailableMasks), by default:
"recommended" masking will be applied. see recommendedMaskNames() for detail.}
}
\value{
a vector of probe ID
}
\description{
get probe masking by mask names
}
\examples{

recommendedMaskNames()[["EPIC"]]
length(getMask("EPIC"))
length(getMask("HM450"))
length(getMask("MM285"))

}
