% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{getProbesByGene}
\alias{getProbesByGene}
\title{Get probes by gene}
\usage{
getProbesByGene(geneName, platform = "EPIC", refversion = "hg38")
}
\arguments{
\item{geneName}{gene name}

\item{platform}{EPIC or HM450}

\item{refversion}{hg38 or hg19}
}
\value{
probes that fall into the given gene
}
\description{
Get probes by gene
}
\examples{
probes <- getProbesByGene('CDKN2A')
}
