% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{getProbesByGene}
\alias{getProbesByGene}
\title{Get Probes by Gene}
\usage{
getProbesByGene(geneName, platform = c("EPIC", "HM450"),
  refversion = c("hg38", "hg19"))
}
\arguments{
\item{geneName}{gene name}

\item{platform}{EPIC or HM450}

\item{refversion}{hg38 or hg19}
}
\value{
probes that fall into the given gene
}
\description{
Get probes mapped to a gene. All transcripts for the gene are considered.
The function takes a gene name as appears in UCSC RefGene database. The
platform and reference genome build can be changed with `platform` and
`refversion` options. The function returns a vector of probes that falls
into the given gene.
}
\examples{
probes <- getProbesByGene('CDKN2A')
}
