% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getProbesByRegion}
\alias{getProbesByRegion}
\title{get probes by genomic region}
\usage{
getProbesByRegion(chrm, beg = 1, end = -1, platform = "EPIC",
  refversion = "hg38")
}
\arguments{
\item{chrm}{chromosome}

\item{beg}{begin, 1 if omitted}

\item{end}{end, chromosome end if omitted}

\item{platform}{EPIC or HM450}

\item{refversion}{hg38 or hg19}
}
\value{
probes that fall into the given region
}
\description{
get probes by genomic region
}
\examples{
getProbesByRegion('chr5', 135413937, 135419936,
    refversion='hg19', platform='HM450')
}
