% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_composition.R
\name{getRefSet}
\alias{getRefSet}
\title{retrieve reference set}
\usage{
getRefSet(cells = NULL, platform = "EPIC")
}
\arguments{
\item{cells}{reference cell types}

\item{platform}{EPIC or HM450}
}
\value{
g
}
\description{
retrieve reference set
}
\examples{
betas <- getRefSet('CD4T', platform='HM450')
}
