% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_methylation.R
\name{getSegment}
\alias{getSegment}
\title{Select segment from coefficient table}
\usage{
getSegment(cf1, seg.id)
}
\arguments{
\item{cf1}{coefficient table of one factor from DMR}

\item{seg.id}{segment ID}
}
\value{
coefficient table from given segment
}
\description{
This function takes a coefficient table and returns a subset of the table
targeting only the specified segment using segment ID.
}
\examples{
data <- sesameDataGet('HM450.76.TCGA.matched')
cf <- DMR(data$betas, ~type, meta=data$sampleInfo)
getSegment(cf[[1]], cf[[1]][['Seg.ID']][1])
}
