% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{getSegment}
\alias{getSegment}
\title{select segment from coefficient table}
\usage{
getSegment(cf1, seg.id)
}
\arguments{
\item{cf1}{coefficient table of one factor from DMR}

\item{seg.id}{segment ID}
}
\value{
coefficient table from given segment
}
\description{
select segment from coefficient table
}
\examples{
betas <- readRDS(system.file(
    'extdata','HM450.betas.76matchedTCGAchr20.rds', package='sesameData'))
sample.info <- readRDS(system.file(
    'extdata','HM450.sampleinfo.76matchedTCGAchr20.rds', package='sesameData'))
cf <- DMR(betas, sample.info, ~type)
getSegment(cf[[1]], cf[[1]][['Seg.ID']][1])
}
