% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{getSexInfo}
\alias{getSexInfo}
\title{Get sex-related information}
\usage{
getSexInfo(sset)
}
\arguments{
\item{sset}{a \code{SignalSet}}
}
\value{
medianY and fracXlinked
}
\description{
The function takes a \code{SignalSet} and returns a vector of three
numerics: the median intensity of chrY probes; the median intensity of
chrX probes; and fraction of intermediate chrX probes. chrX and chrY
probes excludes pseudo-autosomal probes.
}
\examples{
sset <- makeExampleSeSAMeDataSet()
getSexInfo(sset)
}
