% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\docType{methods}
\name{IG}
\alias{IG}
\alias{IG,SigSet-method}
\title{IG getter generic}
\usage{
IG(x)

\S4method{IG}{SigSet}(x)
}
\arguments{
\item{x}{object of \code{SigSet}}
}
\value{
The IG slot of \code{SigSet}
}
\description{
IG getter generic

Get IG slot of SigSet class
}
