% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{imputeTo}
\alias{imputeTo}
\title{Impute to platform}
\usage{
imputeTo(betas, platform = NULL, probes = NULL, tissue = "Blood")
}
\arguments{
\item{betas}{named vector of beta values}

\item{platform}{platform of target imputation}

\item{probes}{Probe ID if not set to vector names}

\item{tissue}{tissue context of imputation}
}
\description{
Impute to platform
}
\examples{

betas <- c("cg04707299"=0.2, "cg13380562"=0.9, "cg00000103"=0.1)
betas2 <- imputeTo(betas, "EPIC")
betas2[names(betas)]
head(betas2)
}
