% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex.R
\name{inferSex}
\alias{inferSex}
\title{Infer Sex}
\usage{
inferSex(x, pfm = NULL)
}
\arguments{
\item{x}{either a \code{SigDF} or a beta value vector named by probe ID}
}
\value{
'F' or 'M'
We established our sex calling based on the CpGs hypermethylated in
inactive X (XiH), CpGs hypomethylated in inactive X (XiL) and signal
intensity ratio of Y-chromosome over autosomes. Currently human inference
uses a random forest and mouse inference uses a support vector machine.

XXY male (Klinefelter's), 45,X female (Turner's) can confuse the
model sometimes.
Our function works on a single sample.
}
\description{
Infer Sex
}
\examples{
sesameDataCache("EPIC") # if not done yet
sdf <- sesameDataGet('EPIC.1.SigDF')
inferSex(sdf)
}
