% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex.R
\name{inferSexKaryotypes}
\alias{inferSexKaryotypes}
\title{Infer Sex Karyotype}
\usage{
inferSexKaryotypes(sdf)
}
\arguments{
\item{sdf}{a \code{SigDF}}
}
\value{
Karyotype string, with XCI
}
\description{
The function takes a \code{SigDF} and infers the sex chromosome Karyotype
and presence/absence of X-chromosome inactivation (XCI). chrX, chrY and XCI
are inferred relatively independently. This function gives a more detailed
look of potential sex chromosome aberrations.
}
\examples{
sesameDataCache() # if not done yet
sdf <- sesameDataGet('EPIC.1.SigDF')
inferSexKaryotypes(sdf)
}
