% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{inferSexKaryotypes}
\alias{inferSexKaryotypes}
\title{Infer Sex Karyotype}
\usage{
inferSexKaryotypes(sset)
}
\arguments{
\item{sset}{a \code{SigSet}}
}
\value{
Karyotype string, with XCI
}
\description{
The function takes a \code{SigSet} and infers the sex chromosome Karyotype
and presence/absence of X-chromosome inactivation (XCI). chrX, chrY and XCI
are inferred relatively independently. This function gives a more detailed
look of potential sex chromosome aberrations.
}
\examples{
sset <- readRDS(system.file(
    'extdata','EPIC.sset.LNCaP.Rep1.rds',package='sesameData'))
inferSexKaryotypes(sset)
}
