% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{inferSpecies}
\alias{inferSpecies}
\title{Infer Species}
\usage{
inferSpecies(
  sdf,
  topN = 1000,
  threshold.pos = 0.01,
  threshold.neg = 0.1,
  return.auc = FALSE,
  return.species = FALSE,
  balance = TRUE,
  threshold.success.rate = 0.8
)
}
\arguments{
\item{sdf}{a \code{SigDF}}

\item{topN}{Top n positive and negative probes used to infer species.}

\item{threshold.pos}{pvalue < threshold.pos are considered positive
(default: 0.01).}

\item{threshold.neg}{pvalue > threshold.neg are considered negative
(default: 0.2).}

\item{return.auc}{return AUC calculated, override return.species}

\item{return.species}{return a string to represent species}

\item{balance}{whether to balance the postive and negative probe
number (default: TRUE).}

\item{threshold.success.rate}{threshold of success rate to determine
mouse species.}
}
\value{
a new SigDF with updated color channel specification and
masking unless return.auc or return.species is set to TRUE
}
\description{
We infer species based on probes pvalues and alignment score.
AUC was calculated for each specie, y_true is 1 or 0 
for pval < threshold.pos or pval > threshold.neg, respeceively,
}
\examples{
if (FALSE) { ## remove this, testing doesn't allow large file caching
  sdf <- sesameDataGet("MM285.1.SigDF")
  sdf <- inferSpecies(sdf)
}
}
