% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{inferTypeIChannel}
\alias{inferTypeIChannel}
\title{Infer and reset color channel for Type-I probes instead of
using what is specified in manifest}
\usage{
inferTypeIChannel(
  sset,
  switch_failed = FALSE,
  verbose = FALSE,
  summary = FALSE
)
}
\arguments{
\item{sset}{a \code{SigSet}}

\item{switch_failed}{whether to switch failed probes (default to FALSE)}

\item{verbose}{whether to print correction summary}

\item{summary}{return summarized numbers only.}
}
\value{
a \code{SigSet}, or numerics if summary == TRUE
}
\description{
Infer and reset color channel for Type-I probes instead of
using what is specified in manifest
}
\examples{

sset <- sesameDataGet('EPIC.1.LNCaP')$sset
inferTypeIChannel(sset)

}
