% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileSet.R
\name{initFileSet}
\alias{initFileSet}
\title{initialize a fileSet class by allocating appropriate storage}
\usage{
initFileSet(map_path, platform, samples, inc = 8)
}
\arguments{
\item{map_path}{path of file to map}

\item{platform}{EPIC, HM450 or HM27, consistent with sset@platform}

\item{samples}{sample names}

\item{inc}{bytes per unit data storage}
}
\value{
a sesame::fileSet object
}
\description{
initialize a fileSet class by allocating appropriate storage
}
\examples{

fset <- initFileSet('/tmp/mybetas2', 'HM27', c('s1','s2'))
}
