% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KYCG_plot.R
\name{KYCG_plotBar}
\alias{KYCG_plotBar}
\title{Bar plot to show most enriched CG groups from testEnrichment}
\usage{
KYCG_plotBar(df, n_min = 10, n_max = 30, max_fdr = 0.05, min_cap = -5)
}
\arguments{
\item{df}{KYCG result data frame}

\item{n_min}{minimum number of databases to report}

\item{n_max}{maximum number of databases to report}

\item{max_fdr}{maximum FDR}

\item{min_cap}{the minimum log2(OR), value below this are capped}
}
\value{
grid plot object
}
\description{
The input data frame should have an "estimate" and
a "FDR" columns.
}
\details{
Top CG groups are determined by estimate (descending order).
}
\examples{
KYCG_plotBar(data.frame(
  estimate=runif(10,0,10), FDR=runif(10,0,1), nD=10,
  overlap=as.integer(runif(10,0,30)), group="g", dbname=seq_len(10)))
}
