% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KYCG_plot.R
\name{KYCG_plotEnrichAll}
\alias{KYCG_plotEnrichAll}
\title{plot enrichment test result}
\usage{
KYCG_plotEnrichAll(df, fdr_max = 25, n_label = 15)
}
\arguments{
\item{df}{test enrichment result data frame}

\item{fdr_max}{maximum fdr for capping}

\item{n_label}{number of database to label}
}
\value{
grid object
}
\description{
plot enrichment test result
}
\examples{
query <- KYCG_getDBs("MM285.designGroup")[["PGCMeth"]]
res <- testEnrichment(query)
KYCG_plotEnrichAll(res)

}
