% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KYCG_plot.R
\name{KYCG_plotMeta}
\alias{KYCG_plotMeta}
\title{Plot meta gene or other meta genomic features}
\usage{
KYCG_plotMeta(betas, platform = NULL)
}
\arguments{
\item{betas}{a named numeric vector or a matrix
(row: probes; column: samples)}

\item{platform}{if not given and x is a SigDF, will be inferred
the meta features}
}
\value{
a grid plot object
}
\description{
Plot meta gene or other meta genomic features
}
\examples{
sdf <- sesameDataGet("EPIC.1.SigDF")
KYCG_plotMeta(getBetas(sdf))
}
