% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KYCG_plot.R
\name{KYCG_plotVolcano}
\alias{KYCG_plotVolcano}
\title{creates a volcano plot of -log2(p.value) and log(estimate)
given data with fields estimate and p.value.}
\usage{
KYCG_plotVolcano(data, alpha = 0.05)
}
\arguments{
\item{data}{DataFrame where each field is a database name with two fields
for the estimate and p.value.}

\item{alpha}{Float representing the cut-off alpha value for the plot. 
Optional. (Default: 0.05)}
}
\value{
ggplot volcano plot
}
\description{
creates a volcano plot of -log2(p.value) and log(estimate)
given data with fields estimate and p.value.
}
\examples{

data=data.frame(estimate=c(runif(10)), FDR=c(runif(10)))
KYCG_plotVolcano(data)

}
