% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mLiftOver.R
\name{liftOver}
\alias{liftOver}
\title{liftOver, see mLiftOver (renamed)}
\usage{
liftOver(...)
}
\arguments{
\item{...}{see mLiftOver}
}
\value{
imputed data, vector, matrix, SigDF(s)
}
\description{
liftOver, see mLiftOver (renamed)
}
