% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{liftOver}
\alias{liftOver}
\title{lift over beta values or SigDFs to another Infinium platform
This function wraps ID conversion and provide optional
imputation functionality.}
\usage{
liftOver(
  x,
  target_platform,
  source_platform = NULL,
  mapping = NULL,
  impute = FALSE,
  celltype = "Blood"
)
}
\arguments{
\item{x}{either named beta value (vector or matrix), probe IDs
or SigDF(s)
if input is a matrix, probe IDs should be in the row names
if input is a numeric vector, probe IDs should be in the vector
names.
If input is a character vector, the input will be
considered probe IDs.}

\item{target_platform}{the platform to take the data to}

\item{source_platform}{optional information of the source data
platform (when there might be ambiguity).}

\item{mapping}{a liftOver mapping file. Typically this file
contains empirical evidence whether a probe mapping is reliable.
If given, probe ID-based mapping will be skipped. This is to
perform more stringent probe ID mapping.}

\item{impute}{whether to impute or not, default is FALSE}

\item{celltype}{the cell type / tissue context of imputation,
if not given, will use nearest neighbor to find out.}
}
\value{
imputed data, vector, matrix, SigDF(s)
}
\description{
lift over beta values or SigDFs to another Infinium platform
This function wraps ID conversion and provide optional
imputation functionality.
}
\examples{

\dontrun{
sesameDataCache()
sdf = sesameDataGet("EPICv2.8.SigDF")[["GM12878_206909630042_R08C01"]]
betas = openSesame(sdf)
betas_HM450 = liftOver(betas, "HM450", impute=TRUE)

## directly map probes
cg_epic2 = grep("cg", names(sesameData_getManifestGRanges("EPICv2")), value=T)
head(liftOver(cg_epic2, "HM450"))

cg_hm450 = grep("cg", names(sesameData_getManifestGRanges("HM450")), value=T)
head(liftOver(cg_hm450, "EPICv2"))

rs_epic2 = grep("rs", names(sesameData_getManifestGRanges("EPICv2")), value=T)
head(liftOver(rs_epic2, "HM450", source_platform="EPICv2"))

}
}
