% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{listAvailableMasks}
\alias{listAvailableMasks}
\title{list existing quality masks for a SigDF}
\usage{
listAvailableMasks(platform)
}
\arguments{
\item{platform}{EPIC, MM285, HM450 etc}
}
\value{
a tibble of masks
}
\description{
list existing quality masks for a SigDF
}
\examples{
listAvailableMasks("EPIC")
}
