% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{makeExampleTinyEPICDataSet}
\alias{makeExampleTinyEPICDataSet}
\title{Make a tiny toy simulated EPIC data set}
\usage{
makeExampleTinyEPICDataSet()
}
\value{
Object of class \code{SignalSet}
}
\description{
Make a tiny toy simulated EPIC data set
}
\examples{
sset <- makeExampleTinyEPICDataSet()

}
