% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{makeExampleTinyEPICDataSet}
\alias{makeExampleTinyEPICDataSet}
\title{Make a tiny toy simulated EPIC data set}
\usage{
makeExampleTinyEPICDataSet()
}
\value{
Object of class \code{SignalSet}
}
\description{
Construct a tiny EPIC \code{SignalSet} of only 6 probes. In-band signals
were simulated using a N(4000, 200) normal distribution. Out-of-band signals
were simulated using a N(400, 200) normal distribution. Control signals were
simulated using a N(400, 300) normal distribution.
}
\examples{
sset <- makeExampleTinyEPICDataSet()

}
