% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{meanIntensity}
\alias{meanIntensity}
\title{Mean Intensity}
\usage{
meanIntensity(sdf, mask = TRUE)
}
\arguments{
\item{sdf}{a \code{SigDF}}

\item{mask}{whether to mask probes using mask column}
}
\value{
mean of all intensities
}
\description{
The function takes one single \code{SigDF} and computes mean
intensity of all the in-band measurements. This includes all Type-I
in-band measurements and all Type-II probe measurements. Both methylated
and unmethylated alleles are considered. This function outputs a single
numeric for the mean.
}
\examples{
sesameDataCache("EPIC") # if not done yet
sdf <- sesameDataGet('EPIC.1.LNCaP')$sdf
meanIntensity(sdf)
}
