% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{meanIntensity}
\alias{meanIntensity}
\title{Mean Intensity}
\usage{
meanIntensity(sset)
}
\arguments{
\item{sset}{a \code{SigSet}}
}
\value{
mean of all intensities
}
\description{
The function takes one single \code{SigSet} and computes mean
intensity of all the in-band measurements. This includes all Type-I
in-band measurements and all Type-II probe measurements. Both methylated
and unmethylated alleles are considered. This function outputs a single
numeric for the mean.
}
\examples{
sset <- makeExampleSeSAMeDataSet()
meanIntensity(sset)
}
