% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{meanIntensity}
\alias{meanIntensity}
\title{mean intensity}
\usage{
meanIntensity(sset)
}
\arguments{
\item{sset}{a \code{SignalSet}}
}
\value{
mean of all intensities
}
\description{
mean intensity of all the in-band measurements
}
\examples{
sset <- makeExampleSeSAMeDataSet()
intensities <- meanIntensity(sset)
}
