% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_correction.R
\name{noob}
\alias{noob}
\title{Noob background correction}
\usage{
noob(sdf, bgR = NULL, bgG = NULL, offset = 15)
}
\arguments{
\item{sdf}{a \code{SigDF}}

\item{bgR}{background red probes, if not given use all oobR}

\item{bgG}{background grn probes, if not given use all oobG}

\item{offset}{offset}
}
\value{
a new \code{SigDF} with noob background correction
}
\description{
The function takes a \code{SigDF} and returns a modified \code{SigDF}
with background subtracted. Background was modelled in a normal distribution
and true signal in an exponential distribution. The Norm-Exp deconvolution
is parameterized using Out-Of-Band (oob) probes
}
\examples{
sdf <- sesameDataGet('EPIC.1.SigDF')
sdf.nb <- noob(sdf)
}
