% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_correction.R
\name{noob}
\alias{noob}
\title{Noob background correction}
\usage{
noob(sset, offset = 15)
}
\arguments{
\item{sset}{a \code{SigSet}}

\item{offset}{offset}
}
\value{
a new \code{SigSet} with noob background correction
}
\description{
The function takes a \code{SigSet} and returns a modified \code{SigSet}
with background subtracted. Background was modelled in a normal distribution
and true signal in an exponential distribution. The Norm-Exp deconvolution
is parameterized using Out-Of-Band (oob) probes
}
\examples{
sset <- makeExampleTinyEPICDataSet()
sset.nb <- noob(sset)
}
