% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_correction.R
\name{noobsb}
\alias{noobsb}
\title{background subtraction with bleeding-through subtraction}
\usage{
noobsb(sset, offset = 15, detailed = FALSE)
}
\arguments{
\item{sset}{a \code{SignalSet}}

\item{offset}{offset}

\item{detailed}{if TRUE, return a list of \code{SignalSet}
and regression function}
}
\value{
a modified \code{SignalSet} with background correction
}
\description{
Norm-Exp deconvolution using Out-Of-Band (oob) probes.
Mean and standard deviations are estimated from cross-channel regression
}
\examples{
sset <- makeExampleSeSAMeDataSet('HM450')
sset.nb <- noobsb(sset)
}
