% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_correction.R
\name{openSeSAMe}
\alias{openSeSAMe}
\title{The openSeSAMe pipeline}
\usage{
openSeSAMe(sset)
}
\arguments{
\item{sset}{a \code{SigSet}}
}
\value{
a numeric vector for processed beta values
sset <- make
}
\description{
This function takes a \code{SigSet} and output beta values.It is
a simple wrapper of noob + nonlinear dye bias correction + pOOBAH masking.
}
\examples{
sset <- sesameDataGet('HM450.1.TCGA.PAAD')$sset
betas <- openSeSAMe(sset)
}
