% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{openSesame}
\alias{openSesame}
\title{The openSesame pipeline}
\usage{
openSesame(x)
}
\arguments{
\item{x}{SigSet(s), IDAT prefix(es), minfi GenomicRatioSet(s), 
or RGChannelSet(s)}
}
\value{
a numeric vector for processed beta values
}
\description{
This function is a simple wrapper of noob + nonlinear dye bias 
correction + pOOBAH masking.
}
\details{
If the input is an IDAT prefix or a \code{SigSet}, the output is
the beta value numerics. If the input is a minfi GenomicRatioSet
or RGChannelSet, the output is the sesamized GenomicRatioSet.
}
\examples{
sset <- sesameDataGet('HM450.1.TCGA.PAAD')$sset
IDATprefixes <- searchIDATprefixes(
    system.file("extdata", "", package = "sesameData"))
betas <- openSesame(IDATprefixes)
}
