% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{openSesame}
\alias{openSesame}
\title{The openSesame pipeline}
\usage{
openSesame(
  x,
  platform = "",
  manifest = NULL,
  what = "beta",
  BPPARAM = SerialParam(),
  ...
)
}
\arguments{
\item{x}{SigSet(s), IDAT prefix(es), minfi GenomicRatioSet(s), 
or RGChannelSet(s)}

\item{platform}{optional platform string}

\item{manifest}{optional dynamic manifest}

\item{what}{either 'sigset' or 'beta'}

\item{BPPARAM}{get parallel with MulticoreParam(n)}

\item{...}{parameters to getBetas}
}
\value{
a numeric vector for processed beta values
}
\description{
This function is a simple wrapper of noob + nonlinear dye bias 
correction + pOOBAH masking.
}
\details{
If the input is an IDAT prefix or a \code{SigSet}, the output is
the beta value numerics. If the input is a minfi GenomicRatioSet
or RGChannelSet, the output is the sesamized GenomicRatioSet.
}
\examples{
sset <- sesameDataGet('HM450.1.TCGA.PAAD')$sset
IDATprefixes <- searchIDATprefixes(
    system.file("extdata", "", package = "sesameData"))
betas <- openSesame(IDATprefixes)
}
