% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{openSesame}
\alias{openSesame}
\title{The openSesame pipeline}
\usage{
openSesame(x, platform = "", manifest = NULL, BPPARAM = SerialParam(), ...)
}
\arguments{
\item{x}{SigDF(s), IDAT prefix(es)}

\item{platform}{optional platform string}

\item{manifest}{optional dynamic manifest}

\item{BPPARAM}{get parallel with MulticoreParam(n)}

\item{...}{parameters to getBetas}
}
\value{
a numeric vector for processed beta values
}
\description{
This function is a simple wrapper of noob + nonlinear dye bias 
correction + pOOBAH masking.
}
\details{
If the input is an IDAT prefix or a \code{SigDF}, the output is
the beta value numerics.
}
\examples{
sdf <- sesameDataGet('HM450.1.TCGA.PAAD')$sdf
IDATprefixes <- searchIDATprefixes(
    system.file("extdata", "", package = "sesameData"))
betas <- openSesame(IDATprefixes)
}
