% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO.R
\name{parseGEOsignalMU}
\alias{parseGEOsignalMU}
\title{Convert signal M and U to SigDF}
\usage{
parseGEOsignalMU(sigM, sigU, Probe_IDs, oob = 500)
}
\arguments{
\item{sigM}{methylated signal, a numeric vector}

\item{sigU}{unmethylated signal, a numirc vector}

\item{Probe_IDs}{probe ID vector}

\item{oob}{assumed value for out-of-band signals}
}
\value{
SigDF
}
\description{
This overcomes the issue of missing IDAT files. However,
out-of-band signals will be missing or faked.
}
\examples{
sigM <- c(11436, 6068, 2864)
sigU <- c(1476, 804, 393)
probes <- c("cg07881041", "cg23229610", "cg03513874")
sdf <- parseGEOsignalMU(sigM, sigU, probes)
}
