% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kyCG.R
\name{plotLollipop}
\alias{plotLollipop}
\title{plotLollipop creates a lollipop plot of log(estimate) given data with fields
estimate.}
\usage{
plotLollipop(data, n = 10, title = NA, subtitle = NA)
}
\arguments{
\item{data}{DataFrame where each field is a database name with its estimate.}

\item{n}{Integer representing the number of top enrichments to report.
Optional. (Default: 10)}

\item{title}{String representing the title label. Optional. (Default: NA)}

\item{subtitle}{String representing the subtitle label. Optional. (Default:
NA)}
}
\value{
ggplot lollipop plot
}
\description{
plotLollipop creates a lollipop plot of log(estimate) given data with fields
estimate.
}
\examples{
data = data.frame(estimate=c(runif(10, 0, 10)))
plotLollipop(data)

}
