% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{predictAge}
\alias{predictAge}
\title{Predict age using linear models}
\usage{
predictAge(betas, model, na_fallback = TRUE, min_nonna = 10)
}
\arguments{
\item{betas}{a probeID-named vector of beta values}

\item{model}{a model object from sesameDataGet. should contain
param, intercept, response2age. default to the Horvath353 model.}

\item{na_fallback}{use fall back values if na}

\item{min_nonna}{the minimum number of non-NA values.}
}
\value{
age in the unit specified in the model (usually in year, but
sometimes can be month, like in the mouse clocks).
}
\description{
The function takes a named numeric vector of beta values. The name attribute
contains the probe ID (cg, ch or rs IDs). The function looks for overlapping
probes and estimate age using different models.
}
\details{
You can get the models such as the Horvath aging model (Horvath 2013
Genome Biology) from sesameDataGet. The function outputs a single numeric
of age in years.

Here are some built-in age models:

Horvath353 - Horvath aging model (Horvath 2013 Genome Biol, 353 probes)
SkinBlood - Horvath aging model (Horvath et al. 2018 Aging, 391 probes)
PhenoAge - PhenoAge model (Levine 2018 Aging, 514 cg probes)
Zhou347 - Mouse epigenetic clock (Zhou et al. 2022, 347 probes)
}
\examples{
betas <- sesameDataGet('HM450.1.TCGA.PAAD')$betas
model <- sesameAnno_get("Anno/HM450/Clock_Horvath353.rds")
predictAge(betas, model)

betas <- SummarizedExperiment::assay(
    sesameDataGet('MM285.10.SE.tissue'))[,1]
model <- sesameAnno_get("Anno/MM285/Clock_Zhou347.rds")
predictAge(betas, model)
}
