% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{prefixMask}
\alias{prefixMask}
\title{Mask SigDF by probe ID prefix}
\usage{
prefixMask(sdf, prefixes = NULL, invert = FALSE)
}
\arguments{
\item{sdf}{SigDF}
}
\value{
SigDF
}
\description{
Mask SigDF by probe ID prefix
}
\examples{
sdf <- sesameDataGet("MM285.1.SigDF")
sum(prefixMask(sdf, c("ctl","rs"))$mask)
}
