% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{print.DMLSummary}
\alias{print.DMLSummary}
\title{Print DMLSummary object}
\usage{
\method{print}{DMLSummary}(x, ...)
}
\arguments{
\item{x}{a DMLSummary object}

\item{...}{extra parameter for print}
}
\value{
print DMLSummary result on screen
}
\description{
Print DMLSummary object
}
\examples{
sesameDataCache("HM450") # in case not done yet
data <- sesameDataGet('HM450.76.TCGA.matched')
smry <- DML(data$betas[1:1000,], ~type, meta=data$sampleInfo)
smry

sesameDataClearCache()
}
