% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileSet.R
\name{print.fileSet}
\alias{print.fileSet}
\title{Print a fileSet}
\usage{
\method{print}{fileSet}(x, ...)
}
\arguments{
\item{x}{a sesame::fileSet}

\item{...}{stuff for print}
}
\value{
string representation
}
\description{
Print a fileSet
}
\examples{

fset <- initFileSet('mybetas2', 'HM27', c('s1','s2'))
fset

}
