% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{print.sesameQC}
\alias{print.sesameQC}
\title{Print sesameQC object}
\usage{
\method{print}{sesameQC}(x, ...)
}
\arguments{
\item{x}{a sesameQC object}

\item{...}{extra parameter for print}
}
\value{
print sesameQC result on screen
}
\description{
Print sesameQC object
}
\examples{
sset <- sesameDataGet('EPIC.1.LNCaP')$sset
sesameQC(sset)
}
