% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{qualityMask}
\alias{qualityMask}
\title{Mask beta values by design quality}
\usage{
qualityMask(sdf, verbose = FALSE, ...)
}
\arguments{
\item{sdf}{a \code{SigDF} object}

\item{verbose}{print more messages}

\item{...}{masking details see getMask()}
}
\value{
a filtered \code{SigDF}
}
\description{
Currently quality masking only supports three platforms
see also listAvailableMasks(sdfPlatform(sdf))
}
\examples{
sesameDataCache() # if not done yet
sdf <- sesameDataGet('EPIC.1.SigDF')
sum(sdf$mask)
sum(qualityMask(sdf)$mask)
sum(qualityMask(sdf, mask_names = NULL)$mask)

## list available masks, the dbname column
listAvailableMasks(sdfPlatform(sdf))
listAvailableMasks("EPICv2")

}
