% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{qualityMask}
\alias{qualityMask}
\title{Mask beta values by design quality}
\usage{
qualityMask(sdf, mask_groups = "recommended")
}
\arguments{
\item{sdf}{a \code{SigDF} object}

\item{mask_groups}{mask group name, default to "recommended"}
}
\value{
a filtered \code{SigDF}
}
\description{
Currently quality masking only supports three platforms
see also listAvailableMasks(sdf)
}
\examples{
sesameDataCache() # if not done yet
sdf <- sesameDataGet('EPIC.1.SigDF')
sum(sdf$mask)
sum(qualityMask(sdf)$mask)

## list available masks, the mask_name column
listAvailableMasks(sdf)

}
