% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{qualityMask}
\alias{qualityMask}
\title{Mask beta values by design quality}
\usage{
qualityMask(
  sdf,
  mask.use.manifest = TRUE,
  manifest = NULL,
  mask.use.tcga = FALSE
)
}
\arguments{
\item{sdf}{a \code{SigDF} object}

\item{mask.use.manifest}{use manifest to mask probes}

\item{mask.use.tcga}{whether to use TCGA masking, only applies to HM450}
}
\value{
a filtered \code{SigDF}
}
\description{
Currently quality masking only supports three platforms
}
\examples{
sesameDataCache("EPIC") # if not done yet
sdf <- sesameDataGet('EPIC.1.SigDF')
sum(sdf$mask)
sum(qualityMask(sdf)$mask)
}
