% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{qualityMask}
\alias{qualityMask}
\title{Mask beta values by design quality}
\usage{
qualityMask(sset, mask.use.tcga = FALSE)
}
\arguments{
\item{mask.use.tcga}{whether to use TCGA masking, only applies to HM450}

\item{betas}{numeric vector of beta values}

\item{platform}{HM27, HM450 or EPIC}
}
\value{
a filtered \code{SigSet}
}
\description{
Currently quality masking only supports three platforms
}
\examples{
sset <- sesameDataGet('EPIC.1.LNCaP')$sset
sset.masked <- qualityMask(sset)
}
