% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{qualityRank}
\alias{qualityRank}
\title{This function looks at public data of similar nature
e.g., tissue, FFPE vs non-FFPE, etc to evaluate the quality
of the target data quality}
\usage{
qualityRank(sset, tissue = NULL, samplePrep = NULL, raw = FALSE)
}
\arguments{
\item{sset}{a raw (unprocessed) \code{SigSet}}

\item{tissue}{A string (blood,buccal and saliva)}

\item{samplePrep}{FFPE, fresh, frozen}

\item{raw}{to return the raw comparison table}
}
\value{
three numbers:
1. The number of public samples compared.
2. The fraction of public samples with more nondetection, and
3. The fraction of public samples with lower mean intensity
4. The higher the fraction, the better the sample.
}
\description{
This function looks at public data of similar nature
e.g., tissue, FFPE vs non-FFPE, etc to evaluate the quality
of the target data quality
}
\examples{

# to be added

}
