% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{readIDATpair}
\alias{readIDATpair}
\title{Import a pair of IDATs from one sample}
\usage{
readIDATpair(prefix.path, verbose = FALSE)
}
\arguments{
\item{prefix.path}{sample prefix without _Grn.idat and _Red.idat}

\item{verbose}{be verbose?  (FALSE)}
}
\value{
a \code{SigSet}
}
\description{
The function takes a prefix string that are shared with _Grn.idat
and _Red.idat. The function returns a \code{SigSet}.
}
\examples{
sset <- readIDATpair(sub('_Grn.idat','',system.file(
    "extdata", "4207113116_A_Grn.idat", package = "sesameData")))
}
