% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{readIDATs}
\alias{readIDATs}
\title{Import IDATs from a list of samples}
\usage{
readIDATs(sample.names, base.dir = NULL, raw = FALSE, num.processes = 1)
}
\arguments{
\item{sample.names}{a sample list}

\item{base.dir}{base directory}

\item{raw}{to return raw data without mapping to signal}

\item{num.processes}{number of parallele processes, serial if 1}
}
\value{
a list of \code{SignalSet}s or a list of matrices if `raw=TRUE`
}
\description{
If `raw = FALSE` (default), a list of \code{SignalSet}s are returned. Each
\code{SignalSet} contains all the signal measurement on the platform. If
`raw = TRUE`, a list is returned with each element of the returned list
containing a matrix having signal intensity indexed by chip address.
}
\details{
Sample.names is a vector of common prefixes between the *_Grn.idat and
*_Red.idat. `num.processes` controls the number of parallel workers. It
is default to 1 which means serial.
}
\examples{
ssets <- readIDATs(sub('_Grn.idat','',system.file(
    "extdata", "4207113116_A_Grn.idat", package = "sesameData")))
}
