% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{readIDATs}
\alias{readIDATs}
\title{Import IDATs from a list of samples}
\usage{
readIDATs(sample.names, base.dir = NULL, raw = FALSE, mc = FALSE,
  mc.cores = NULL)
}
\arguments{
\item{sample.names}{a sample list}

\item{base.dir}{base directory}

\item{raw}{to return raw data without mapping to signal}

\item{mc}{use multiple cores}

\item{mc.cores}{number of cores to use, default to getOption('mc.cores'),
then to 4.}
}
\value{
a list of \code{SignalSet}s or a list of matrices if `raw=TRUE`
}
\description{
If `raw = FALSE` (default), a list of \code{SignalSet}s are returned. Each
\code{SignalSet} contains all the signal measurement on the platform. If
`raw = TRUE`, a list is returned with each element of the returned list
containing a matrix having signal intensity indexed by chip address.
}
\details{
Sample.names is a vector of common prefixes between the *_Grn.idat and
*_Red.idat. `mc = TRUE/FALSE` controls whether to process samples in
parallel.`mc.cores` specifies the number of parallel workers
}
\examples{
ssets <- readIDATs(sub('_Grn.idat','',system.file(
    "extdata", "4207113116_A_Grn.idat", package = "sesameData")))
}
