% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{resetMask}
\alias{resetMask}
\title{Reset Masking}
\usage{
resetMask(sset)
}
\arguments{
\item{sset}{a \code{SigSet}}
}
\value{
a new \code{SigSet} with mask reset to empty
}
\description{
Reset Masking
}
\examples{
sesameDataCache("EPIC") # if not done yet
sset <- sesameDataGet('EPIC.1.LNCaP')$sset
sset.no.mask <- resetMask(sset)
}
