% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{resetMask}
\alias{resetMask}
\title{Reset Masking}
\usage{
resetMask(sdf, verbose = FALSE)
}
\arguments{
\item{sdf}{a \code{SigDF}}

\item{verbose}{print more messages}
}
\value{
a new \code{SigDF} with mask reset to all FALSE
}
\description{
Reset Masking
}
\examples{
sesameDataCache() # if not done yet
sdf <- sesameDataGet('EPIC.1.SigDF')
sum(sdf$mask)
sdf <- addMask(sdf, c("cg14057072", "cg22344912"))
sum(sdf$mask)
sum(resetMask(sdf)$mask)
}
