% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesamize.R
\name{RGChannelSetToSigSets}
\alias{RGChannelSetToSigSets}
\title{Convert RGChannelSet (minfi) to a list of SigSet (SeSAMe)}
\usage{
RGChannelSetToSigSets(rgSet, manifest = NULL, BPPARAM = SerialParam())
}
\arguments{
\item{rgSet}{a minfi::RGChannelSet}

\item{manifest}{manifest file}

\item{BPPARAM}{get parallel with MulticoreParam(n)}
}
\value{
a list of sesame::SigSet
}
\description{
Notice the colData() and rowData() is lost. Most cases, rowData is empty
anyway.
}
\examples{

if (FALSE && require(FlowSorted.Blood.450k)) {
    rgSet <- FlowSorted.Blood.450k[,1:2]
    ssets <- RGChannelSetToSigSets(rgSet)
}
}
