% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{scrub}
\alias{scrub}
\title{SCRUB background correction}
\usage{
scrub(sdf)
}
\arguments{
\item{sdf}{a \code{SigDF}}
}
\value{
a new \code{SigDF} with noob background correction
}
\description{
This function takes a \code{SigDF} and returns a modified \code{SigDF}
with background subtracted. scrub subtracts residual background using
background median
}
\details{
This function is meant to be used after noob.
}
\examples{
sdf <- sesameDataGet('EPIC.1.SigDF')
sdf.nb <- noob(sdf)
sdf.nb.scrub <- scrub(sdf.nb)
}
