% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigDFMethods.R
\name{sdf_write_table}
\alias{sdf_write_table}
\title{write SigDF to table file}
\usage{
sdf_write_table(sdf, ...)
}
\arguments{
\item{sdf}{the \code{SigDF} to output}

\item{...}{additional argument to write.table}
}
\value{
write SigDF to table file
}
\description{
write SigDF to table file
}
\examples{
sesameDataCache() # if not done yet
sdf <- sesameDataGet('EPIC.1.SigDF')
sdf_write_table(sdf, file=sprintf("\%s/sigdf.txt", tempdir()))
}
