% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{segmentDMR}
\alias{segmentDMR}
\title{segment DMR}
\usage{
segmentDMR(betas, cf, dist.cutoff = NULL, dist.cutoff.quantile = 0.5,
  platform = "EPIC", refversion = "hg38")
}
\arguments{
\item{betas}{beta values for distance calculation}

\item{cf}{coefficient table from diffMeth}

\item{dist.cutoff}{distance cutoff (default to use dist.cutoff.quantile)}

\item{dist.cutoff.quantile}{quantile to use in selecting cutoff distance}

\item{platform}{EPIC or hm450}

\item{refversion}{hg38 or hg19}
}
\value{
coefficient table with segment ID and segment P-value
}
\description{
This subroutine uses Euclidean distance to group CpGs and
then combine p-values for each segment.
}

