% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\docType{package}
\name{sesame-package}
\alias{_PACKAGE}
\alias{sesame}
\alias{sesame-package}
\title{Analyze DNA methylation data}
\description{
SEnsible and step-wise analysis of DNA methylation data
}
\details{
This package complements array functionalities that allow
processing >10,000 samples in parallel on clusters.
}
\examples{
\dontrun{
library(sesame)

## read IDATs
ssets <- ReadIDATsFromDir(sample.dir, mc=TRUE)

## normalization
ssets <- lapply(ssets, noob)
ssets <- lapply(ssets, dyeBiasCorr)

## convert signal to beta values
betas <- sapply(ssets, function(sset) sset$toBeta())

## mask repeat and SNPs
betas <- sapply(ssets, function(sset) sset$mask())

}
}
\author{
Wanding Zhou \email{Wanding.Zhou@vai.org},
Hui Shen \email{Hui.Shen@vai.org}
}
\keyword{DNAMethylation}
\keyword{Microarray}
\keyword{QualityControl}

