% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\docType{package}
\name{sesame-package}
\alias{_PACKAGE}
\alias{sesame}
\alias{sesame-package}
\title{Analyze DNA methylation data}
\description{
SEnsible and step-wise analysis of DNA methylation data
}
\details{
This package complements array functionalities that allow
processing >10,000 samples in parallel on clusters.
}
\examples{
\dontrun{
library(sesame)

## read IDATs
ssets <- readIDATsFromDir(sample.dir, mc=T)

## normalization
ssets <- mclapply(ssets, noob)
ssets <- mclapply(ssets, dyeBiasCorr)

## convert signal to beta values
betas <- signalToBeta(ssets, mc=T)

## be more parallel
options(mc.cores=28)
dms <- readIDATsFromSheet(sample.sheet, raw=T, mc=T)
simplify2array(mclapply(names(dms), function(nm) {
  dm <- dms[[nm]]
  sset <- chipAddressToSignal(dm)
  sset <- noob(sset)
  sset <- dyeBiasCorr(sset)
  sset$toBeta()
}))
}
}
\author{
Wanding Zhou \email{Wanding.Zhou@vai.org},
Hui Shen \email{Hui.Shen@vai.org}
}
\keyword{DNAMethylation}
\keyword{Microarray}
\keyword{QualityControl}

