% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\docType{package}
\name{sesame-package}
\alias{sesame-package}
\alias{_PACKAGE}
\alias{sesame}
\title{Analyze DNA methylation data}
\description{
SEnsible and step-wise analysis of DNA methylation data
}
\details{
This package complements array functionalities that allow
processing >10,000 samples in parallel on clusters.
}
\examples{
\dontrun{
library(sesame)

## read IDATs
ssets <- readIDATsFromDir(sample.dir, mc=T, mc.cores=4)

## normalization
ssets <- bplapply(ssets, noob, mc.preschedule=F)
ssets <- bplapply(ssets, dyeBiasCorrTypeINorm)

## convert signal to beta values
betas <- do.call(cbind, bplapply(ssets, getBetas))
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zwdzwd/sesame}
  \item Report bugs at \url{https://github.com/zwdzwd/sesame/issues}
}

}
\author{
Wanding Zhou \email{Wanding.Zhou@vai.org},
Hui Shen \email{Hui.Shen@vai.org}
}
\keyword{DNAMethylation}
\keyword{Microarray}
\keyword{QualityControl}
