% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesameAnno.R
\name{sesameAnno_download}
\alias{sesameAnno_download}
\title{Download additional annotation files}
\usage{
sesameAnno_download(title, dest_dir, version = anno_base_default_version)
}
\arguments{
\item{title}{title of the annotation file}

\item{dest_dir}{download to this directory}

\item{version}{version number}
}
\value{
annotation file
}
\description{
From the Infinium annotation website associated github repo
e.g., 
https://github.com/zhou-lab/InfiniumAnnotationV1
}
\details{
The default version number should always work. One need to
refer to the actual repo to see which one of the other versions
also work.

See also
http://zwdzwd.github.io/InfiniumAnnotation
}
\examples{

## avoid testing as this function uses external host
if (FALSE) {
sesameAnno_download("Test/3999492009_R01C01_Grn.idat", tempdir())
}

}
