% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesameAnno.R
\name{sesameAnno_get}
\alias{sesameAnno_get}
\title{Retrieve additional annotation Rds data}
\usage{
sesameAnno_get(title, version = anno_base_default_version, dest_dir = NULL)
}
\arguments{
\item{title}{title of the annotation file}

\item{version}{version number}

\item{dest_dir}{if not NULL, download to this directory}
}
\value{
annotation file
}
\description{
From the Infinium annotation website associated github repo
e.g., 
https://github.com/zhou-lab/InfiniumAnnotationV1
}
\details{
The default version number should always work. One need to
refer to the actual repo to see which one of the other versions
also work.

See also
http://zwdzwd.github.io/InfiniumAnnotation
}
\examples{

## avoided testing as this function uses external host
if (FALSE) {
annoI <- anno_get("Anno/EPIC/EPIC.hg19.typeI_overlap_b151.rds")
}

}
