% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sesameGetExample}
\alias{sesameGetExample}
\title{Retrieve SeSAMe examples}
\usage{
sesameGetExample(nm, platform = "")
}
\arguments{
\item{nm}{name}

\item{platform}{optional, HM450, EPIC or HM27}
}
\value{
example object
}
\description{
Retrieve SeSAMe examples
}
\examples{
betas <- sesameGetExample('TCGA-2L-AAQA-01A-21D-A38H-05')
}
