% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{sesamePlotIntensVsBetas}
\alias{sesamePlotIntensVsBetas}
\title{Plot Total Signal Intensities vs Beta Values
This plot is helpful in revealing the extent of signal background
and dye bias.}
\usage{
sesamePlotIntensVsBetas(sdf, mask = TRUE, intens.range = c(5, 15), ...)
}
\arguments{
\item{sdf}{a \code{SigDF}}

\item{mask}{whether to remove probes that are masked}

\item{intens.range}{plot range of signal intensity}

\item{...}{additional arguments to smoothScatter}
}
\value{
create a total signal intensity vs beta value plot
}
\description{
Plot Total Signal Intensities vs Beta Values
This plot is helpful in revealing the extent of signal background
and dye bias.
}
\examples{
sesameDataCache("EPIC")
sdf <- # if not done yet
sdf <- sesameDataGet('EPIC.1.SigDF')
sesamePlotIntensVsBetas(sdf)
}
