% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{sesameQC}
\alias{sesameQC}
\title{Generate summary numbers that indicative of experiment quality
Please provide a raw sigset (before any preprocessing). Usually
directly from readIDATpair}
\usage{
sesameQC(sset)
}
\arguments{
\item{sset}{a \code{SigSet} object}
}
\value{
a sesameQC class object
}
\description{
Generate summary numbers that indicative of experiment quality
Please provide a raw sigset (before any preprocessing). Usually
directly from readIDATpair
}
\examples{
sesameDataCache("EPIC") # if not done yet
sset <- sesameDataGet('EPIC.1.LNCaP')$sset
sesameQC(sset)
}
