% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{sesameQC}
\alias{sesameQC}
\title{Generate summary numbers that indicative of experiment quality}
\usage{
sesameQC(sset, betas = NULL)
}
\arguments{
\item{sset}{a \code{SigSet} object}

\item{betas}{processed beta values}
}
\value{
a sesameQC class object
}
\description{
Generate summary numbers that indicative of experiment quality
}
\examples{
sset <- sesameDataGet('EPIC.1.LNCaP')$sset
sesameQC(sset)
}
