% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{sesameQC_calcStats}
\alias{sesameQC_calcStats}
\title{A convenience function to call one or multiple
sesameQC_calcStats functions}
\usage{
sesameQC_calcStats(sdf, funs = NULL, use_all = FALSE)
}
\arguments{
\item{sdf}{a SigDF object}

\item{funs}{a sesameQC_calcStats_* function or a list of them
default to sesameQC_calcStats_detection}
}
\description{
A convenience function to call one or multiple
sesameQC_calcStats functions
}
\examples{
sesameDataCache("EPIC") # if not done yet
sdf <- sesameDataGet('EPIC.1.SigDF')
sesameQC_calcStats(sdf)
}
