% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{sesameQC_plotBetaByDesign}
\alias{sesameQC_plotBetaByDesign}
\title{Plot betas distinguishing different Infinium chemistries}
\usage{
sesameQC_plotBetaByDesign(
  sdf,
  prep = NULL,
  legend_pos = "top",
  mar = c(3, 3, 1, 1),
  main = "",
  ...
)
}
\arguments{
\item{sdf}{SigDF}

\item{prep}{prep codes to step through}

\item{legend_pos}{legend position (default: top)}

\item{mar}{margin of layout when showing steps of prep}

\item{main}{main title in plots}

\item{...}{additional options to plot}
}
\value{
create a density plot
}
\description{
Plot betas distinguishing different Infinium chemistries
}
\examples{
sdf <- sesameDataGet("EPIC.1.SigDF")
sesameQC_plotBetaByDesign(sdf, prep="DB")
}
