% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{sesameQC_plotHeatSNPs}
\alias{sesameQC_plotHeatSNPs}
\title{Plot SNP heatmap}
\usage{
sesameQC_plotHeatSNPs(sdfs, cluster = TRUE, filter.nonvariant = TRUE)
}
\arguments{
\item{sdfs}{beta value matrix, row: probes; column: samples}

\item{cluster}{show clustered heatmap}

\item{filter.nonvariant}{whether to filter nonvariant (range < 0.3)}
}
\value{
a grid graphics object
}
\description{
Plot SNP heatmap
}
\examples{

sdfs <- sesameDataGet("EPIC.5.SigDF.normal")[1:2]
plt <- sesameQC_plotHeatSNPs(sdfs, filter.nonvariant = FALSE)
}
