% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{sesameQC_plotRedGrnQQ}
\alias{sesameQC_plotRedGrnQQ}
\title{Plot red-green QQ-Plot using Infinium-I Probes}
\usage{
sesameQC_plotRedGrnQQ(sdf, main = "R-G QQ Plot", ...)
}
\arguments{
\item{sdf}{a \code{SigDF}}

\item{main}{plot title}

\item{...}{additional options to qqplot}
}
\value{
create a qqplot
}
\description{
Plot red-green QQ-Plot using Infinium-I Probes
}
\examples{
sesameDataCache() # if not done yet
sdf <- sesameDataGet('EPIC.1.SigDF')
sesameQC_plotRedGrnQQ(sdf)
}
