% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesamize.R
\name{sesamize}
\alias{sesamize}
\title{"fix" an RGset (for which IDATs may be unavailable) with Sesame
The input is an RGSet and the output is a sesamized RGSet}
\usage{
sesamize(rgSet, naFrac = 1, BPPARAM = SerialParam())
}
\arguments{
\item{rgSet}{an RGChannelSet, perhaps with colData of various flavors}

\item{naFrac}{maximum NA fraction for a probe before it gets dropped (1)}

\item{BPPARAM}{get parallel with MulticoreParam(n)}
}
\value{
a sesamized GenomicRatioSet
}
\description{
"fix" an RGset (for which IDATs may be unavailable) with Sesame
The input is an RGSet and the output is a sesamized RGSet
}
\examples{
# Takes about two minutes to process 48 samples on my 48-core desktop
if (require(FlowSorted.CordBloodNorway.450k)) {
    sesamized <- sesamize(
        FlowSorted.CordBloodNorway.450k[,1:2], BPPARAM=MulticoreParam(2))
}
}
