% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sesamize}
\alias{sesamize}
\title{sesamize function is deprecated.
Please check https://github.com/zwdzwd/sesamize for previous scripts}
\usage{
sesamize(...)
}
\arguments{
\item{...}{arguments for sesamize}
}
\value{
a message text for deprecated function
}
\description{
sesamize function is deprecated.
Please check https://github.com/zwdzwd/sesamize for previous scripts
}
\examples{
cat("Deprecated. see https://github.com/zwdzwd/sesamize")
}
