% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\docType{methods}
\name{show,SigSet-method}
\alias{show,SigSet-method}
\title{The display method for SigSet}
\usage{
\S4method{show}{SigSet}(object)
}
\arguments{
\item{object}{displayed object}
}
\value{
message of number of probes in each category.
}
\description{
The function outputs the number of probes in each category and the first
few signal measurements.
}
\examples{
sset <- readRDS(system.file(
    'extdata','EPIC.sset.LNCaP.Rep1.rds',package='sesameData'))
print(sset)
}
