% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigDFMethods.R
\name{SigDF}
\alias{SigDF}
\title{SigDF validation from a plain data frame}
\usage{
SigDF(df, platform = "EPIC", ctl = NULL)
}
\arguments{
\item{df}{a \code{data.frame} with Probe_ID, MG, MR, UG, UR, col and mask}

\item{platform}{a string to specify the array platform}

\item{ctl}{optional control probe data frame}
}
\value{
a \code{SigDF} object
}
\description{
SigDF validation from a plain data frame
}
\examples{
sesameDataCache() # if not done yet
sdf <- sesameDataGet('EPIC.1.SigDF')
}
