% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigDFMethods.R
\name{signalMU}
\alias{signalMU}
\title{report M and U for regular probes}
\usage{
signalMU(sdf, mask = TRUE, MU = FALSE)
}
\arguments{
\item{sdf}{a \code{SigDF}}

\item{mask}{whether to apply mask}

\item{MU}{add a column for M+U}
}
\value{
a data frame of M and U columns
}
\description{
report M and U for regular probes
}
\examples{
sesameDataCache() # if not done yet
sdf <- sesameDataGet('EPIC.1.SigDF')
head(signalMU(sdf))
}
