% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\docType{class}
\name{SigSet-class}
\alias{SigSet-class}
\alias{SigSet}
\alias{initialize,SigSet-method}
\title{SigSet class}
\usage{
\S4method{initialize}{SigSet}(.Object, platform, ...)

SigSet(...)
}
\arguments{
\item{.Object}{target object}

\item{platform}{"EPIC", "HM450", "HM27" or other strings for custom arrays}

\item{...}{additional arguments}
}
\value{
a \code{SigSet} object

a \code{SigSet} object
}
\description{
This is the main data class for SeSAMe. The class holds different
classes of signal intensities.

The function takes a string describing the platform of the data. It can be
one of "HM27", "HM450" or "EPIC".

The function takes a string describing the platform of the data. It can be
one of "HM27", "HM450" or "EPIC".
}
\details{
The NBeads* slots are normally left empty but can be optionally
turned on.
}
\section{Slots}{

\describe{
\item{\code{IG}}{intensity table for type I probes in green channel}

\item{\code{IR}}{intensity table for type I probes in red channel}

\item{\code{IGG}}{Type-I green that is inferred to be green}

\item{\code{IRR}}{Type-I red that is inferred to be red}

\item{\code{II}}{intensity table for type II probes}

\item{\code{oobG}}{out-of-band probes in green channel}

\item{\code{oobR}}{out-of-band probes in red channel}

\item{\code{NBeadsIG}}{Number of Beads for Infinium I green channel}

\item{\code{NBeadsIR}}{Number of Beads for Infinium I red channel}

\item{\code{NBeadsII}}{Number of Beads for Infinium II}

\item{\code{ctl}}{all the control probe intensities}

\item{\code{pval}}{named numeric vector of p-values}

\item{\code{platform}}{"EPIC", "HM450" or "HM27"}
}}

\examples{
## Create an empty EPIC object.
SigSet("EPIC")
SigSet('EPIC')
}
