% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\docType{class}
\name{SigSet-class}
\alias{SigSet-class}
\alias{SigSet}
\alias{initialize,SigSet-method}
\alias{SigSet}
\title{SigSet class}
\usage{
\S4method{initialize}{SigSet}(.Object, platform = "EPIC", ...)

SigSet(...)
}
\arguments{
\item{.Object}{target object}

\item{platform}{"EPIC", "HM450" or "HM27"}

\item{...}{additional arguments}
}
\value{
a \code{SigSet} object

a \code{SigSet} object
}
\description{
SigSet class

Constructor method of SigSet Class.

Wrapper function
}
\section{Slots}{

\describe{
\item{\code{IG}}{intensity table for type I probes in green channel}

\item{\code{IR}}{intensity table for type I probes in red channel}

\item{\code{II}}{intensity table for type II probes}

\item{\code{oobG}}{out-of-band probes in green channel}

\item{\code{oobR}}{out-of-band probes in red channel}

\item{\code{ctl}}{all the control probe intensities}

\item{\code{pval}}{named numeric vector of p-values}

\item{\code{platform}}{"EPIC", "HM450" or "HM27"}
}}

\examples{
## Create an empty EPIC object.
SigSet("EPIC")
}
