% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigSetList.R
\name{SigSetList-methods}
\alias{SigSetList-methods}
\alias{show,SigSetList-method}
\title{SigSetList methods (centralized). 
Currently scarce...
 
`show`         print a summary of the SigSetList.}
\usage{
\S4method{show}{SigSetList}(object)
}
\arguments{
\item{object}{a SigSetList}
}
\value{
Description of SigSetList
}
\description{
SigSetList methods (centralized). 
Currently scarce...
 
`show`         print a summary of the SigSetList.
}
\examples{
SigSetListFromPath(system.file("extdata", "", package = "sesameData"))
}
