% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigSetList.R
\name{SigSetList}
\alias{SigSetList}
\title{constructor}
\usage{
SigSetList(..., pCutoff = NULL)
}
\arguments{
\item{...}{the SigSet objects that will be the List elements}

\item{pCutoff}{(optional) if not NULL, mask probes with poobah >= pCutoff}
}
\value{
a SigSetList
}
\description{
constructor
}
