% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigSetList.R
\name{SigSetListFromPath}
\alias{SigSetListFromPath}
\title{read an entire directory's worth of IDATs into a SigSetList}
\usage{
SigSetListFromPath(path = ".", parallel = FALSE)
}
\arguments{
\item{path}{the path from which to read IDATs (default ".")}

\item{parallel}{run in parallel? (default FALSE)}
}
\value{
a SigSetList
}
\description{
read an entire directory's worth of IDATs into a SigSetList
}
