% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesamize.R
\name{SigSetToRGChannelSet}
\alias{SigSetToRGChannelSet}
\title{Convert SeSAMe::SigSet to minfi::RGChannelSet}
\usage{
SigSetToRGChannelSet(ssets)
}
\arguments{
\item{ssets}{a list of SigSet}
}
\value{
a minfi::RGChannelSet
}
\description{
Convert SeSAMe::SigSet to minfi::RGChannelSet
}
\examples{

sset <- sesameDataGet('EPIC.1.LNCaP')$sset
rgSet <- SigSetToRGChannelSet(sset)

}
