% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgroup.R
\name{[.WGroup}
\alias{[.WGroup}
\title{subset WGroup}
\usage{
\method{[}{WGroup}(x, i)
}
\arguments{
\item{x}{a WGroup object}

\item{i}{integer indexing element}
}
\value{
a subset of WGroup or NULL
}
\description{
subset WGroup
}
