% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{subsetSignal}
\alias{subsetSignal}
\title{Select a subset of probes}
\usage{
subsetSignal(sset, probes)
}
\arguments{
\item{sset}{a \code{SigSet}}

\item{probes}{target probes}
}
\value{
another sset with probes specified
}
\description{
The function takes a \code{SigSet} as input and output another
\code{SigSet} with probes from the given probe selection.
}
\examples{
sset <- readRDS(system.file(
    'extdata','EPIC.sset.LNCaP.Rep1.rds',package='sesameData'))
subsetSignal(sset, rownames(slot(sset, 'IR')))
}
