% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KYCG.R
\name{testEnrichment1}
\alias{testEnrichment1}
\title{testEnrichment1 tests for the enrichment of set of probes (query set)
in a single given feature (database set)}
\usage{
testEnrichment1(query, database, universe)
}
\arguments{
\item{query}{Vector of probes of interest (e.g., significant probes)}

\item{database}{Vector corresponding to the database sets of
interest with associated meta data as an attribute to each element.}

\item{universe}{Vector of probes in the universe set containing all of
the probes to be considered in the test.}
}
\value{
One list containing features corresponding the test estimate,
p-value, and type of test.
}
\description{
testEnrichment1 tests for the enrichment of set of probes (query set)
in a single given feature (database set)
}
