% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KYCG.R
\name{testEnrichmentGene}
\alias{testEnrichmentGene}
\title{Convenient function for testing enrichment of gene linkage}
\usage{
testEnrichmentGene(query, platform = NULL, silent = FALSE, ...)
}
\arguments{
\item{query}{probe set of interest}

\item{platform}{string corresponding to the type of platform to use. Either
MM285, EPIC, HM450, or HM27. If it is not provided, it will be inferred
from the query set probe IDs.}

\item{silent}{whether to output message}

\item{...}{addition argument provided to testEnrichment}
}
\value{
A data frame containing features corresponding to the test estimate,
p-value, and type of test etc.
}
\description{
Convenient function for testing enrichment of gene linkage
}
\examples{
query <- c("cg04707299", "cg13380562", "cg00480749")
testEnrichment(query, platform = "EPIC")
}
