% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KYCG.R
\name{testEnrichmentGSEA}
\alias{testEnrichmentGSEA}
\title{testEnrichmentGSEA uses the GSEA test to estimate the association of a
categorical variable against a continuous variable.}
\usage{
testEnrichmentGSEA(
  query,
  databases = NULL,
  platform = NULL,
  silent = FALSE,
  precise = FALSE
)
}
\arguments{
\item{query}{query, if numerical, expect categorical database, if
categorical expect numerical database}

\item{platform}{EPIC, MM285, ..., infer if not given}

\item{silent}{suppress message (default: FALSE)}

\item{precise}{whether to compute precise p-value (up to numerical limit)
of interest.}

\item{database}{database, numerical or categorical, but needs to be
different from query}
}
\value{
A DataFrame with the estimate/statistic, p-value, and name of test
for the given results.
}
\description{
estimate represent enrichment score and negative estimate indicate a
test for depletion
}
\examples{
query <- KYCG_getDBs("KYCG.MM285.designGroup")[["TSS"]]
res <- testEnrichmentGSEA(query, "MM285.seqContextN")
}
