% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_methylation.R
\name{topLoci}
\alias{topLoci}
\title{Top loci in differential methylation}
\usage{
topLoci(cf1)
}
\arguments{
\item{cf1}{coefficient table of one factor from diffMeth}
}
\value{
coefficient table ordered by p-value of each locus
}
\description{
This is a convenience function to show top differential methylated segments.
The function takes a coefficient table as input and output the same
table ordered by the sigificance of the locus.
}
\examples{
betas <- readRDS(system.file(
    'extdata','HM450.betas.76matchedTCGAchr20.rds',
    package='sesameData'))
sample.info <- readRDS(system.file(
    'extdata','HM450.sampleinfo.76matchedTCGAchr20.rds',
    package='sesameData'))
cf <- DMR(betas, sample.info, ~type)
topLoci(cf[[1]])
}
