% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_methylation.R
\name{topSegments}
\alias{topSegments}
\title{Top segments in differential methylation}
\usage{
topSegments(cf1)
}
\arguments{
\item{cf1}{coefficient table of one factor from DMR}
}
\value{
coefficient table ordered by adjusted p-value of segments
}
\description{
This is a utility function to show top differential methylated segments.
The function takes a coefficient table as input and output the same
table ordered by the sigificance of the segments.
}
\examples{
data <- sesameDataGet('HM450.76.TCGA.matched')
cf <- DMR(data$betas, data$sampleInfo, ~type)
topSegments(cf[[1]])
}
