% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{totalIntensities}
\alias{totalIntensities}
\title{M+U Intensities for All Probes}
\usage{
totalIntensities(sset)
}
\arguments{
\item{sset}{a \code{SigSet}}
}
\value{
a vector of M+U signal for each probe
}
\description{
The function takes one single \code{SigSet} and computes total
intensity of all the in-band measurements by summing methylated and
unmethylated alleles. This function outputs a single numeric for the mean.
}
\examples{
sset <- makeExampleSeSAMeDataSet()
totalIntensities(sset)
}
