% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesame.R
\name{totalIntensities}
\alias{totalIntensities}
\title{M+U Intensities Array}
\usage{
totalIntensities(sdf, mask = FALSE)
}
\arguments{
\item{sdf}{a \code{SigDF}}

\item{mask}{whether to mask probes using mask column}
}
\value{
a vector of M+U signal for each probe
}
\description{
The function takes one single \code{SigDF} and computes total
intensity of all the in-band measurements by summing methylated and
unmethylated alleles. This function outputs a single numeric for the mean.
}
\examples{
sesameDataCache() # if not done yet
sdf <- sesameDataGet('EPIC.1.SigDF')
intensities <- totalIntensities(sdf)
}
