% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualizeGene}
\alias{visualizeGene}
\title{visualize gene}
\usage{
visualizeGene(geneName, betas, platform = "EPIC", upstream = 2000,
  dwstream = 2000, refversion = "hg38", ...)
}
\arguments{
\item{geneName}{gene name}

\item{betas}{beta value matrix (row: probes, column: samples)}

\item{platform}{HM450 (default) or EPIC}

\item{upstream}{distance to extend upstream}

\item{dwstream}{distance to extend downstream}

\item{refversion}{hg19 or hg38}

\item{...}{additional options, see visualizeRegion}
}
\value{
None
}
\description{
visualize gene
}
\examples{
betas <- readRDS(system.file(
    'extdata','HM450.betas.76matchedTCGAchr20.rds',
    package='sesameData'))
visualizeGene('ADA', betas, 'HM450')
}
