% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualizeProbes}
\alias{visualizeProbes}
\title{Visualize region that contains the specified probes}
\usage{
visualizeProbes(probeNames, betas, platform = "EPIC", refversion = "hg38",
  upstream = 1000, dwstream = 1000, ...)
}
\arguments{
\item{probeNames}{probe names}

\item{betas}{beta value matrix (row: probes, column: samples)}

\item{platform}{HM450 (default) or EPIC}

\item{refversion}{hg19 or hg38}

\item{upstream}{distance to extend upstream}

\item{dwstream}{distance to extend downstream}

\item{...}{additional options, see visualizeRegion}
}
\value{
None
}
\description{
Visualize region that contains the specified probes
}
\examples{
betas <- readRDS(system.file(
    'extdata','HM450.betas.76matchedTCGAchr20.rds',
    package='sesameData'))
visualizeProbes(c('cg22316575', 'cg16084772', 'cg20622019'), betas, 'HM450')
}
