% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualizeProbes}
\alias{visualizeProbes}
\title{Visualize Region that Contains the Specified Probes}
\usage{
visualizeProbes(probeNames, betas, platform = c("EPIC", "HM450"),
  refversion = c("hg38", "hg19"), upstream = 1000, dwstream = 1000, ...)
}
\arguments{
\item{probeNames}{probe names}

\item{betas}{beta value matrix (row: probes, column: samples)}

\item{platform}{HM450 or EPIC (default)}

\item{refversion}{hg19 or hg38 (default)}

\item{upstream}{distance to extend upstream}

\item{dwstream}{distance to extend downstream}

\item{...}{additional options, see visualizeRegion}
}
\value{
None
}
\description{
Visualize the beta value in heatmaps for the genomic region containing
specified probes. The function works only if specified probes can be
spanned by a single genomic region. The region can cover more probes
than specified. Hence the plotting heatmap may encompass more probes.
The function takes as input a string vector of probe IDs (cg/ch/rs-numbers).
if draw is FALSE, the function returns the subset beta value matrix
otherwise it returns the grid graphics object.
}
\examples{
betas <- readRDS(system.file(
    'extdata','HM450.betas.76matchedTCGAchr20.rds',
    package='sesameData'))
visualizeProbes(c('cg22316575', 'cg16084772', 'cg20622019'), betas, 'HM450')
}
