% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualizeRegion}
\alias{visualizeRegion}
\title{visualize region}
\usage{
visualizeRegion(betas, chrm, plt.beg, plt.end, platform = "EPIC",
  refversion = "hg19", draw = TRUE, show.sampleNames = TRUE,
  show.probeNames = TRUE)
}
\arguments{
\item{betas}{beta value matrix (row: probes, column: samples)}

\item{chrm}{chromosome}

\item{plt.beg}{begin of the region}

\item{plt.end}{end of the region}

\item{platform}{hm450 (default) or EPIC}

\item{refversion}{hg19 or hg38}

\item{draw}{draw figure or return betas}

\item{show.sampleNames}{whether to show sample names}

\item{show.probeNames}{whether to show probe names}
}
\description{
visualize region
}

