% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualizeRegion}
\alias{visualizeRegion}
\title{visualize region}
\usage{
visualizeRegion(chrm, plt.beg, plt.end, betas, platform = "EPIC",
  refversion = "hg38", heat.height = NULL, draw = TRUE,
  show.sampleNames = TRUE, show.probeNames = TRUE, na.rm = FALSE)
}
\arguments{
\item{chrm}{chromosome}

\item{plt.beg}{begin of the region}

\item{plt.end}{end of the region}

\item{betas}{beta value matrix (row: probes, column: samples)}

\item{platform}{hm450 (default) or EPIC}

\item{refversion}{hg19 or hg38}

\item{heat.height}{heatmap height (auto inferred based on rows)}

\item{draw}{draw figure or return betas}

\item{show.sampleNames}{whether to show sample names}

\item{show.probeNames}{whether to show probe names}

\item{na.rm}{remove probes with all NA.}
}
\description{
visualize region
}

