% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualizeRegion}
\alias{visualizeRegion}
\title{visualize region}
\usage{
visualizeRegion(chrm, plt.beg, plt.end, betas, platform = "EPIC",
  refversion = "hg38", sample.name.fontsize = 10, heat.height = NULL,
  draw = TRUE, show.sampleNames = TRUE, show.samples.n = NULL,
  show.probeNames = TRUE, cluster.samples = FALSE, na.rm = FALSE,
  dmin = 0, dmax = 1)
}
\arguments{
\item{chrm}{chromosome}

\item{plt.beg}{begin of the region}

\item{plt.end}{end of the region}

\item{betas}{beta value matrix (row: probes, column: samples)}

\item{platform}{EPIC or HM450}

\item{refversion}{hg38 or hg19}

\item{sample.name.fontsize}{sample name font size}

\item{heat.height}{heatmap height (auto inferred based on rows)}

\item{draw}{draw figure or return betas}

\item{show.sampleNames}{whether to show sample names}

\item{show.samples.n}{number of samples to show (default: all)}

\item{show.probeNames}{whether to show probe names}

\item{cluster.samples}{whether to cluster samples}

\item{na.rm}{remove probes with all NA.}

\item{dmin}{data min}

\item{dmax}{data max}
}
\value{
graphics or a matrix containing the captured beta values
}
\description{
visualize region
}
\examples{
betas <- readRDS(system.file(
    'extdata','HM450.betas.76matchedTCGAchr20.rds',
    package='sesameData'))
visualizeRegion('chr20', 44648623, 44652152, betas, 'HM450')
}
