% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.R
\name{visualizeSegments}
\alias{visualizeSegments}
\title{Visualize segments}
\usage{
visualizeSegments(seg, to.plot = NULL)
}
\arguments{
\item{seg}{a \code{CNSegment} object}

\item{to.plot}{chromosome to plot (by default plot all chromosomes)}
}
\value{
plot graphics
}
\description{
require ggplot2, scales
}
\examples{
seg <- readRDS(system.file(
    'extdata','EPIC.seg.LNCaP.Rep1.rds',package='sesameData'))
visualizeSegments(seg)

}
