% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.R
\name{visualizeSegments}
\alias{visualizeSegments}
\title{Visualize segments}
\usage{
visualizeSegments(seg, to.plot = NULL)
}
\arguments{
\item{seg}{a \code{CNSegment} object}

\item{to.plot}{chromosome to plot (by default plot all chromosomes)}
}
\value{
plot graphics
}
\description{
The function takes a \code{CNSegment} object obtained from cnSegmentation
and plot the bin signals and segments (as horizontal lines).
}
\details{
require ggplot2, scales
}
\examples{

sesameDataCache("EPIC") # in case not done yet
sdf <- sesameDataGet('EPIC.1.SigDF')
sdfs.normal <- sesameDataGet('EPIC.5.SigDF.normal')
seg <- cnSegmentation(sdf, sdfs.normal)

visualizeSegments(seg)

sesameDataClearCache()

}
